Cut a Dataset for Analysis at a Specified Date

Cut a Dataset for Analysis at a Specified Date

cutData(x, cutDate)

Arguments

x
a time-to-event dataset, e.g., generated by simPWSurv

cutDate
date relative to start of randomization (cte from input dataset) at which dataset is to be cut off for analysis

Value

A dataset ready for survival analysis

Examples

# Use default enrollment and event rates and cut at calendar time 5 after start
# of randomization
library(dplyr)
simPWSurv(n=20) %>% cutData(5)

#> # A tibble: 20 x 4
#> # Groups:   Stratum, Treatment [2]
#>      tte event Stratum Treatment   
#>    <dbl> <dbl> <chr>   <chr>       
<<<<<<< HEAD
#>  1  4.70     0 All     Experimental
#>  2  4.51     0 All     Control     
#>  3  2.84     1 All     Experimental
#>  4  4.44     0 All     Control     
#>  5  1.08     1 All     Control     
#>  6  4.18     0 All     Experimental
#>  7  1.13     1 All     Experimental
#>  8  3.92     0 All     Control     
#>  9  3.92     0 All     Experimental
#> 10  3.80     0 All     Experimental
#> 11  3.72     0 All     Control     
#> 12  3.59     0 All     Control     
#> 13  3.58     0 All     Control     
#> 14  3.54     0 All     Experimental
#> 15  3.49     0 All     Experimental
#> 16  3.46     0 All     Control     
#> 17  3.20     0 All     Control     
#> 18  3.18     0 All     Control     
#> 19  3.17     0 All     Experimental
#> 20  3.09     0 All     Experimental

======= #> 1 4.52 0 All Experimental #> 2 0.194 1 All Experimental #> 3 4.34 0 All Control #> 4 2.29 1 All Control #> 5 4.26 0 All Experimental #> 6 4.16 0 All Experimental #> 7 4.08 0 All Control #> 8 3.98 0 All Control #> 9 3.95 0 All Experimental #> 10 3.54 0 All Control #> 11 3.16 0 All Experimental #> 12 3.12 0 All Control #> 13 0.294 1 All Experimental #> 14 2.55 0 All Experimental #> 15 2.14 0 All Control #> 16 2.00 0 All Control #> 17 1.82 0 All Control #> 18 1.77 0 All Experimental #> 19 1.70 0 All Control #> 20 1.56 0 All Experimental
>>>>>>> 8da6ac0d6f0042026e5db4aae1934a972cd63c44