Cut a Dataset for Analysis at a Specified Event Count

`cutDataAtCount` takes a time-to-event data set and cuts the data at which an event count is reached.

cutDataAtCount(x, count)

Arguments

x
a time-to-event dataset, e.g., generated by simPWSurv

count
event count at which data cutoff is to be made

Value

a tibble ready for survival analysis, including culumns time to event (`tte`), `event`, the `stratum` and the `treatment.`

Examples

library(tidyr)
# Use default enrollment and event rates at cut at 100 events
x <- simPWSurv(n=200) %>% cutDataAtCount(100)
table(x$event,x$Treatment)

#>    
#>     Control Experimental
<<<<<<< HEAD
#>   0      37           63
#>   1      63           37

======= #> 0 41 59 #> 1 59 41

>>>>>>> 8da6ac0d6f0042026e5db4aae1934a972cd63c44