Permuted fixed block randomization

Fixed block randomization. The `block` input should repeat each treatment code the number of times it is to be included within each block. The final block will be a partial block if `n` is not an exact multiple of the block length.

fixedBlockRand(n = 10, block = c(0, 0, 1, 1))

Arguments

n
sample size to be randomized

block
Vector of treatments to be included in each block

Value

A treatment group sequence (vector) of length `n` with treatments from `block` permuted within each block having block size equal to the length of `block`

Examples

library(dplyr)
# 2:1 randomization with block size 3, treatments "A" and "B"
tibble(x=1:10) %>% mutate(Treatment=fixedBlockRand(block=c("A","B","B")))

#> # A tibble: 10 x 2
#>        x Treatment
#>    <int> <chr>    
#>  1     1 B        
#>  2     2 A        
#>  3     3 B        
#>  4     4 B        
#>  5     5 B        
#>  6     6 A        
#>  7     7 A        
#>  8     8 B        
#>  9     9 B        
#> 10    10 B        

# Stratified randomization
tibble(Stratum=c(rep("A",10),rep("B",10))) %>%
group_by(Stratum) %>%
mutate(Treatment=fixedBlockRand())

#> # A tibble: 20 x 2
#> # Groups:   Stratum [2]
#>    Stratum Treatment
#>    <chr>       <dbl>
#>  1 A               0
#>  2 A               1
#>  3 A               1
#>  4 A               0
#>  5 A               1
#>  6 A               0
#>  7 A               0
#>  8 A               1
#>  9 A               1
#> 10 A               0
#> 11 B               0
#> 12 B               1
#> 13 B               0
#> 14 B               1
#> 15 B               0
#> 16 B               1
#> 17 B               0
#> 18 B               1
#> 19 B               0
#> 20 B               1