MaxCombo p-value

pMaxCombo() computes p-values for the MaxCombo test based on output from simtrial::tenFHcorr(). This is still in an experimental stage and is intended for use with the simtrial::simfix() trial simulation routine. However, it can also be used to analyze clinical trial data such as that provided in the ADaM ADTTE format.

pMaxCombo(Z, dummyvar, algorithm = GenzBretz(maxpts = 50000, abseps = 1e-05))

Arguments

Z
a dataset output from tenFHcorr(); see examples.

dummyvar
a dummy input that allows group_map() to be used to compute p-values for multiple simulations.

algorithm
This is passed directly to the algorithm argument in the mvtnorm::pmvnorm()

Value

A numeric p-value

Examples

library(tidyr)
x <- simfix(nsim=1,timingType=5,rg=tibble::tibble(rho=c(0,0,1),gamma=c(0,1,1)))
head(x)

#>   Events       lnhr rho gamma         Z        V1        V2        V3
#> 1    350 -0.3224632   0     0 -2.999278 1.0000000 0.8552956 0.9328038
#> 2    350 -0.3224632   0     1 -3.345850 0.8552956 1.0000000 0.9419201
#> 3    350 -0.3224632   1     1 -3.239182 0.9328038 0.9419201 1.0000000
#>                             cut Duration Sim
#> 1 Max(min follow-up, event cut)  72.6043   1
#> 2 Max(min follow-up, event cut)  72.6043   1
#> 3 Max(min follow-up, event cut)  72.6043   1

pMaxCombo(x)

#> [1] 0.0008076197

# Only use cuts for events, events + min follow-up
xx <- simfix(nsim=100,timingType=5,rg=tibble::tibble(rho=c(0,0,1),gamma=c(0,1,1)))
head(xx)

#>   Events       lnhr rho gamma         Z        V1        V2        V3
#> 1    373 -0.2758722   0     0 -2.657128 1.0000000 0.8548440 0.9354456
#> 2    373 -0.2758722   0     1 -2.824036 0.8548440 1.0000000 0.9193780
#> 3    373 -0.2758722   1     1 -2.969040 0.9354456 0.9193780 1.0000000
#> 4    370 -0.5198712   0     0 -4.955347 1.0000000 0.8516446 0.9334546
#> 5    370 -0.5198712   0     1 -5.711525 0.8516446 1.0000000 0.9184127
#> 6    370 -0.5198712   1     1 -5.650737 0.9334546 0.9184127 1.0000000
#>                             cut Duration Sim
#> 1 Max(min follow-up, event cut) 74.28377   1
#> 2 Max(min follow-up, event cut) 74.28377   1
#> 3 Max(min follow-up, event cut) 74.28377   1
#> 4 Max(min follow-up, event cut) 71.89120   2
#> 5 Max(min follow-up, event cut) 71.89120   2
#> 6 Max(min follow-up, event cut) 71.89120   2

# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- unlist(xx %>%  dplyr::group_by(Sim) %>% dplyr::group_map(pMaxCombo))
mean(p<.025)

#> [1] 0.97