MaxCombo p-value

pvalue_maxcombo() computes p-values for the MaxCombo test based on output from simtrial::tenFHcorr(). This is still in an experimental stage and is intended for use with the simtrial::sim_fixed_n() trial simulation routine. However, it can also be used to analyze clinical trial data such as that provided in the ADaM ADTTE format.

Usage

pvalue_maxcombo(
  Z,
  dummy_var,
  algorithm = GenzBretz(maxpts = 50000, abseps = 1e-05)
)

Arguments

Z

a dataset output from tenFHcorr(); see examples.

dummy_var

a dummy input that allows group_map() to be used to compute p-values for multiple simulations.

algorithm

This is passed directly to the algorithm argument in the mvtnorm::pmvnorm()

Value

A numeric p-value

Examples

library(tidyr)
library(tibble)
library(dplyr)

# example 1
x <- sim_fixed_n(n_sim = 1,
            timing_type = 5,
            rho_gamma = tibble(rho = c(0, 0, 1),
                        gamma = c(0, 1, 1)))
head(x)
#>   Events       lnhr rho gamma          Z        V1        V2        V3
#> 1    356 -0.1006035   0     0 -0.9465493 1.0000000 0.8538705 0.9336442
#> 2    356 -0.1006035   0     1 -1.1085484 0.8538705 1.0000000 0.9291844
#> 3    356 -0.1006035   1     1 -1.1625115 0.9336442 0.9291844 1.0000000
#>                             cut Duration Sim
#> 1 Max(min follow-up, event cut) 75.82895   1
#> 2 Max(min follow-up, event cut) 75.82895   1
#> 3 Max(min follow-up, event cut) 75.82895   1
pvalue_maxcombo(x)
#> [1] 0.1768348

# example 2
# Only use cuts for events, events + min follow-up
xx <- sim_fixed_n(n_sim = 100,
             timing_type = 5,
             rho_gamma = tibble(rho = c(0, 0, 1),
                         gamma = c(0, 1, 1)))
head(xx)
#>   Events       lnhr rho gamma         Z        V1        V2        V3
#> 1    350 -0.3634486   0     0 -3.381543 1.0000000 0.8560027 0.9329911
#> 2    350 -0.3634486   0     1 -3.927115 0.8560027 1.0000000 0.9434671
#> 3    350 -0.3634486   1     1 -3.725732 0.9329911 0.9434671 1.0000000
#> 4    352 -0.3836014   0     0 -3.588231 1.0000000 0.8470113 0.9313598
#> 5    352 -0.3836014   0     1 -3.359856 0.8470113 1.0000000 0.9110861
#> 6    352 -0.3836014   1     1 -3.270217 0.9313598 0.9110861 1.0000000
#>                             cut Duration Sim
#> 1 Max(min follow-up, event cut) 69.74127   1
#> 2 Max(min follow-up, event cut) 69.74127   1
#> 3 Max(min follow-up, event cut) 69.74127   1
#> 4 Max(min follow-up, event cut) 69.43770   2
#> 5 Max(min follow-up, event cut) 69.43770   2
#> 6 Max(min follow-up, event cut) 69.43770   2
# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx %>% group_by(Sim) %>% group_map(pvalue_maxcombo) %>% unlist()
mean(p < .025)
#> [1] 0.99