Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All        0.00441 Control         36.2         286. 36.2      1
#>  2 All        0.153   Control         18.9         121. 19.1      1
#>  3 All        0.274   Experimental    10.5         445. 10.7      1
#>  4 All        0.369   Experimental    20.0         375. 20.3      1
#>  5 All        0.552   Experimental    10.1        2999. 10.6      1
#>  6 All        0.807   Experimental    25.4         777. 26.2      1
#>  7 All        0.937   Control          4.50        135.  5.43     1
#>  8 All        1.01    Control         10.9         238. 11.9      1
#>  9 All        1.03    Experimental    62.7         854. 63.7      1
#> 10 All        1.09    Control         16.3        1111. 17.4      1
#> 11 All        1.31    Experimental    15.6        1400. 16.9      1
#> 12 All        1.35    Control          7.49       3498.  8.84     1
#> 13 All        1.38    Control          6.45       1401.  7.82     1
#> 14 All        1.43    Experimental    30.9         196. 32.3      1
#> 15 All        1.57    Experimental    10.1         527. 11.7      1
#> 16 All        1.80    Control         25.4         947. 27.2      1
#> 17 All        1.98    Experimental    34.7         274. 36.7      1
#> 18 All        2.18    Experimental    22.2        1849. 24.4      1
#> 19 All        2.28    Control          5.88        928.  8.15     1
#> 20 All        2.30    Control          1.29        680.  3.58     1
 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All        0.00180 Control         0.490      494.    0.491     1
#>  2 All        0.0412  Experimental   29.6          2.30  2.34      0
#>  3 All        0.159   Experimental   48.5        857.   48.7       1
#>  4 All        0.266   Experimental   16.6        908.   16.9       1
#>  5 All        0.306   Experimental   13.0        516.   13.3       1
#>  6 All        0.338   Control        11.5        284.   11.8       1
#>  7 All        0.374   Experimental   54.1        391.   54.5       1
#>  8 All        0.494   Experimental    2.81       539.    3.31      1
#>  9 All        0.603   Experimental    0.416     1107.    1.02      1
#> 10 All        0.748   Control         3.52       513.    4.27      1
#> 11 All        0.780   Experimental   88.1       1045.   88.9       1
#> 12 All        0.817   Experimental    0.168      248.    0.985     1
#> 13 All        0.931   Experimental   24.3        281.   25.2       1
#> 14 All        1.08    Experimental   33.1       1405.   34.1       1
#> 15 All        1.12    Control        26.1       1494.   27.2       1
#> 16 All        1.14    Experimental   46.9       1605.   48.0       1
#> 17 All        1.15    Experimental    2.64       522.    3.79      1
#> 18 All        1.26    Experimental   10.9       1000.   12.2       1
#> 19 All        1.36    Experimental    8.33      2828.    9.70      1
#> 20 All        1.69    Control        25.5       1254.   27.2       1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 High         0.100 Experimental    0.369       385.    0.469     1
#>  2 High         0.239 Control        13.8        1089.   14.0       1
#>  3 High         0.290 Experimental    0.617       208.    0.907     1
#>  4 Low          0.881 Control        11.3          88.4  12.1       1
#>  5 High         1.01  Control        16.9          44.7  17.9       1
#>  6 Low          1.11  Control         5.30       1566.    6.41      1
#>  7 High         1.22  Experimental   17.8         688.   19.1       1
#>  8 High         1.28  Experimental   17.5        1340.   18.8       1
#>  9 High         1.45  Control         4.18        222.    5.63      1
#> 10 High         1.53  Control        17.2         823.   18.7       1
#> 11 High         1.54  Experimental   11.3         119.   12.8       1
#> 12 Low          1.60  Experimental    2.52        344.    4.13      1
#> 13 High         1.66  Experimental  101.          785.  103.        1
#> 14 High         1.82  Control         4.07        326.    5.89      1
#> 15 High         2.17  Control         3.86       2903.    6.03      1
#> 16 High         2.41  Control         5.62         11.8   8.03      1
#> 17 High         2.52  Control         7.78        617.   10.3       1
#> 18 High         2.52  Experimental   32.5         289.   35.0       1
#> 19 Low          2.91  Experimental   31.0          61.5  33.9       1
#> 20 High         3.02  Experimental    1.14       1094.    4.16      1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High        0.0131 Experimental    22.0       2573.   22.0      1
#>  2 High        0.0408 Control         32.3        195.   32.3      1
#>  3 High        0.506  Control          8.37       518.    8.88     1
#>  4 High        0.755  Experimental    17.6       1261.   18.4      1
#>  5 High        1.09   Experimental   132.         696.  133.       1
#>  6 High        1.25   Control         64.8        449.   66.1      1
#>  7 High        1.30   Control         25.5        803.   26.8      1
#>  8 Low         1.34   Control         50.9        424.   52.2      1
#>  9 High        1.35   Experimental     1.47      1282.    2.82     1
#> 10 Low         1.66   Experimental    75.1         28.8  30.5      0
#> 11 High        1.69   Control          3.75       398.    5.44     1
#> 12 High        1.75   Experimental     5.53      1108.    7.27     1