Simulate a stratified time-to-event outcome randomized trial

simPWSurv_ enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv_(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment
    = c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simtrial:::simPWSurv_(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.112 Control        19.1          118. 19.2      1
#>  2 All          0.250 Control         1.25         112.  1.50     1
#>  3 All          0.303 Experimental   18.0          560. 18.3      1
#>  4 All          0.455 Experimental   17.0         1197. 17.4      1
#>  5 All          0.616 Experimental   46.5         1271. 47.1      1
#>  6 All          1.09  Experimental   36.4          447. 37.5      1
#>  7 All          1.15  Control        12.6          110. 13.7      1
#>  8 All          1.30  Control         2.50         132.  3.79     1
#>  9 All          1.36  Control         7.36         115.  8.72     1
#> 10 All          1.43  Experimental   15.7         2156. 17.1      1
#> 11 All          1.55  Experimental   72.6          888. 74.2      1
#> 12 All          1.59  Control         1.51        1889.  3.10     1
#> 13 All          1.67  Experimental   11.3          288. 13.0      1
#> 14 All          1.88  Control        14.4          326. 16.3      1
#> 15 All          1.88  Control        13.6         2107. 15.5      1
#> 16 All          1.91  Experimental   48.2         1206. 50.1      1
#> 17 All          2.26  Control         8.12         644. 10.4      1
#> 18 All          2.60  Experimental   35.0         1663. 37.6      1
#> 19 All          2.71  Experimental    0.635       1458.  3.35     1
#> 20 All          2.74  Control         1.56         835.  4.30     1
 # 3:1 randomization
 simtrial:::simPWSurv_(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0329 Control        18.1         441.  18.1      1
#>  2 All         0.206  Experimental    7.40       1569.   7.61     1
#>  3 All         0.208  Experimental    5.89        901.   6.10     1
#>  4 All         0.449  Experimental   25.0         531.  25.4      1
#>  5 All         0.558  Experimental   23.2        2009.  23.8      1
#>  6 All         0.650  Experimental   58.5         300.  59.2      1
#>  7 All         1.08   Experimental    0.491       313.   1.57     1
#>  8 All         1.17   Control         2.82        360.   4.00     1
#>  9 All         1.30   Experimental    1.06        383.   2.36     1
#> 10 All         1.31   Control        13.6        2790.  14.9      1
#> 11 All         1.42   Experimental    4.05        118.   5.47     1
#> 12 All         1.45   Experimental    8.29         80.8  9.74     1
#> 13 All         1.50   Experimental   30.1         141.  31.6      1
#> 14 All         1.74   Experimental    5.87        345.   7.61     1
#> 15 All         1.75   Experimental    1.38        273.   3.13     1
#> 16 All         1.75   Control        10.1          70.2 11.9      1
#> 17 All         1.77   Experimental   34.0         178.  35.8      1
#> 18 All         1.92   Experimental    2.49        406.   4.40     1
#> 19 All         1.93   Control        15.7         133.  17.6      1
#> 20 All         1.98   Experimental   11.6        6042.  13.6      1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simtrial:::simPWSurv_(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low          0.101 Control         58.4        738.  58.5      1
#>  2 High         0.109 Control         28.4        266.  28.5      1
#>  3 Low          0.143 Control         71.8       3132.  71.9      1
#>  4 Low          0.448 Experimental     6.63       242.   7.08     1
#>  5 High         0.604 Experimental    21.3       4661.  21.9      1
#>  6 High         0.641 Experimental    55.9        853.  56.5      1
#>  7 High         0.811 Control          5.12       110.   5.93     1
#>  8 High         0.811 Experimental    10.7        446.  11.5      1
#>  9 Low          1.00  Experimental     8.17      1521.   9.18     1
#> 10 Low          1.04  Experimental    11.1       1200.  12.1      1
#> 11 High         1.13  Control         17.5        252.  18.6      1
#> 12 Low          1.14  Experimental    78.7       1743.  79.8      1
#> 13 High         1.20  Control         23.2         14.4 15.6      0
#> 14 High         1.36  Experimental    15.6        764.  17.0      1
#> 15 Low          1.39  Control         22.8        165.  24.2      1
#> 16 Low          1.52  Control         36.0        263.  37.5      1
#> 17 High         1.56  Experimental    29.1        612.  30.7      1
#> 18 Low          1.61  Experimental     3.37      1917.   4.98     1
#> 19 High         1.62  Control          3.92       654.   5.53     1
#> 20 Low          1.72  Control         54.8        604.  56.5      1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simtrial:::simPWSurv_(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High         0.493 Experimental  190.          669.  190.       1
#>  2 Low          0.512 Control        48.6         991.   49.1      1
#>  3 Low          0.606 Experimental   23.1        2260.   23.7      1
#>  4 High         0.634 Control        29.2         111.   29.9      1
#>  5 High         0.824 Control        37.6          30.0  30.9      0
#>  6 High         0.839 Experimental    0.289       453.    1.13     1
#>  7 High         0.969 Experimental    3.51        204.    4.48     1
#>  8 High         1.07  Control         0.765       980.    1.84     1
#>  9 High         1.15  Control        10.3         155.   11.5      1
#> 10 High         1.29  Experimental   33.0         369.   34.3      1
#> 11 High         1.41  Control        23.8         411.   25.2      1
#> 12 High         1.52  Control         8.50        937.   10.0      1