Simulation of fixed sample size design for time-to-event endpoint

`simfix()` provide simulations of a single endpoint two-arm trial where the enrollment, hazard ratio, and failure and dropout rates change over time.

simfix(
  nsim = 1000,
  sampleSize = 500,
  targetEvents = 350,
  strata = tibble::tibble(Stratum = "All", p = 1),
  enrollRates = tibble::tibble(duration = c(2, 2, 10), rate = c(3, 6, 9)),
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2)),
  totalDuration = 30,
  block = rep(c("Experimental", "Control"), 2),
  timingType = 1:5,
  rg = tibble::tibble(rho = 0, gamma = 0)
)

Arguments

nsim
Number of simulations to perform.

sampleSize
Total sample size per simulation.

targetEvents
Targeted event count for analysis.

strata
A tibble with strata specified in `Stratum`, probability (incidence) of each stratum in `p`.

enrollRates
Piecewise constant enrollment rates by time period. Note that these are overall population enrollment rates and the `strata` argument controls the random distribution between strata.

failRates
Piecewise constant control group failure rates, hazard ratio for experimental vs control, and dropout rates by stratum and time period.

totalDuration
Total follow-up from start of enrollment to data cutoff.

block
As in `simtrial::simPWSurv()`. Vector of treatments to be included in each block.

timingType
A numeric vector determining data cutoffs used; see details. Default is to include all available cutoff methods.

rg
As in `simtrial::tenFHCorr()`. A tibble with variables rho and gamma, both greater than equal to zero, to specify one Fleming-Harrington weighted logrank test per row.

Value

A tibble including columns Events (event count), lnhr (log-hazard ratio), Z (normal test statistic; < 0 favors experimental) cut (text describing cutoff used), Duration (duration of trial at cutoff for analysis) and sim (sequential simulation id). One row per simulated dataset per cutoff specified in timingType, per test statistic specified. If multiple Fleming-Harrington tests are specified in rg, then columns rho,gamma are also included.

Details

timingType has up to 5 elements indicating different options for data cutoff. 1 uses the planned study duration, 2 the time the targeted event count is achieved, 3 the planned minimum follow-up after enrollment is complete, 4 the maximum of planned study duration and targeted event count cuts (1 and 2), 5 the maximum of targeted event count and minimum follow-up cuts (2 and 3).

Examples

library(tidyr)
library(dplyr)
# Show output structure
simfix(nsim=3)

#>    Events       lnhr          Z                              cut Duration Sim
#> 1      99 -0.5017231 -2.4867983                 Planned duration 30.00000   1
#> 2     350 -0.5127519 -4.7534852                  Targeted events 70.95562   1
#> 3     369 -0.5132016 -4.8868930                Minimum follow-up 75.49585   1
#> 4     350 -0.5127519 -4.7534852 Max(planned duration, event cut) 70.95562   1
#> 5     369 -0.5132016 -4.8868930    Max(min follow-up, event cut) 75.49585   1
#> 6      95 -0.5423255 -2.6275921                 Planned duration 30.00000   2
#> 7     350 -0.4260287 -3.9697757                  Targeted events 71.19390   2
#> 8     359 -0.4194617 -3.9589607                Minimum follow-up 73.97906   2
#> 9     350 -0.4260287 -3.9697757 Max(planned duration, event cut) 71.19390   2
#> 10    359 -0.4194617 -3.9589607    Max(min follow-up, event cut) 73.97906   2
#> 11     99 -0.1881764 -0.9347157                 Planned duration 30.00000   3
#> 12    350 -0.3076297 -2.8683511                  Targeted events 67.35884   3
#> 13    355 -0.3086422 -2.8945053                Minimum follow-up 69.71603   3
#> 14    350 -0.3076297 -2.8683511 Max(planned duration, event cut) 67.35884   3
#> 15    355 -0.3086422 -2.8945053    Max(min follow-up, event cut) 69.71603   3

# Example with 2 tests: logrank and FH(0,1)
simfix(nsim=1,rg=tibble::tibble(rho=0,gamma=c(0,1)))

#>    Events       lnhr rho gamma         Z        V1        V2
#> 1     106 -0.2454674   0     0 -1.258183 1.0000000 0.8365670
#> 2     106 -0.2454674   0     1 -1.714048 0.8365670 1.0000000
#> 3     350 -0.2750308   0     0 -2.567805 1.0000000 0.8465478
#> 4     350 -0.2750308   0     1 -2.604712 0.8465478 1.0000000
#> 5     372 -0.3069135   0     0 -2.953437 1.0000000 0.8518861
#> 6     372 -0.3069135   0     1 -3.102061 0.8518861 1.0000000
#> 7     350 -0.2750308   0     0 -2.567805 1.0000000 0.8465478
#> 8     350 -0.2750308   0     1 -2.604712 0.8465478 1.0000000
#> 9     372 -0.3069135   0     0 -2.953437 1.0000000 0.8518861
#> 10    372 -0.3069135   0     1 -3.102061 0.8518861 1.0000000
#>                                 cut Duration Sim
#> 1                  Planned duration 30.00000   1
#> 2                  Planned duration 30.00000   1
#> 3                   Targeted events 70.90062   1
#> 4                   Targeted events 70.90062   1
#> 5                 Minimum follow-up 75.67859   1
#> 6                 Minimum follow-up 75.67859   1
#> 7  Max(planned duration, event cut) 70.90062   1
#> 8  Max(planned duration, event cut) 70.90062   1
#> 9     Max(min follow-up, event cut) 75.67859   1
#> 10    Max(min follow-up, event cut) 75.67859   1

# Power by test
# Only use cuts for events, events + min follow-up
xx <- simfix(nsim=100,timingType=c(2,5),rg=tibble::tibble(rho=0,gamma=c(0,1)))
# Get power approximation for FH, data cutoff combination
xx %>% group_by(cut,rho,gamma) %>% summarise(mean(Z<=qnorm(.025)))

#> `summarise()` has grouped output by 'cut', 'rho'. You can override using the `.groups` argument.

#> # A tibble: 4 x 4
#> # Groups:   cut, rho [2]
#>   cut                             rho gamma `mean(Z <= qnorm(0.025))`
#>   <chr>                         <dbl> <dbl>                     <dbl>
#> 1 Max(min follow-up, event cut)     0     0                      0.98
#> 2 Max(min follow-up, event cut)     0     1                      0.98
#> 3 Targeted events                   0     0                      0.97
#> 4 Targeted events                   0     1                      0.99

# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx %>%  filter(cut != "Targeted events") %>% group_by(Sim) %>% group_map(pMaxCombo)

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

p <- unlist(p)
mean(p<.025)

#> [1] 0.99

# MaxCombo estimate for targeted events cutoff
p <- unlist(xx %>%  filter(cut == "Targeted events") %>% group_by(Sim) %>% group_map(pMaxCombo))

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

#> Warning: Unknown or uninitialised column: `c`.

mean(p<.025)

#> [1] 0.99

>>>>>>> 8da6ac0d6f0042026e5db4aae1934a972cd63c44