Simulation of fixed sample size design for time-to-event endpoint

simfix_() provide simulations of a single endpoint two-arm trial where the enrollment, hazard ratio, and failure and dropout rates change over time.

Usage

simfix_(
  nsim = 1000,
  sampleSize = 500,
  targetEvents = 350,
  strata = tibble::tibble(Stratum = "All", p = 1),
  enrollRates = tibble::tibble(duration = c(2, 2, 10), rate = c(3, 6, 9)),
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2)),
  totalDuration = 30,
  block = rep(c("Experimental", "Control"), 2),
  timingType = 1:5,
  rg = tibble::tibble(rho = 0, gamma = 0)
)

Arguments

nsim

Number of simulations to perform.

sampleSize

Total sample size per simulation.

targetEvents

Targeted event count for analysis.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p.

enrollRates

Piecewise constant enrollment rates by time period. Note that these are overall population enrollment rates and the strata argument controls the random distribution between strata.

failRates

Piecewise constant control group failure rates, hazard ratio for experimental vs control, and dropout rates by stratum and time period.

totalDuration

Total follow-up from start of enrollment to data cutoff.

block

As in simtrial:::simPWSurv_(). Vector of treatments to be included in each block.

timingType

A numeric vector determining data cutoffs used; see details. Default is to include all available cutoff methods.

rg

As in simtrial::tenFHCorr(). A tibble with variables rho and gamma, both greater than equal to zero, to specify one Fleming-Harrington weighted logrank test per row.

Value

A tibble including columns Events (event count), lnhr (log-hazard ratio), Z (normal test statistic; < 0 favors experimental) cut (text describing cutoff used), Duration (duration of trial at cutoff for analysis) and sim (sequential simulation id). One row per simulated dataset per cutoff specified in timingType, per test statistic specified. If multiple Fleming-Harrington tests are specified in rg, then columns rho,gamma

are also included.

Details

timingType has up to 5 elements indicating different options for data cutoff. 1 uses the planned study duration, 2 the time the targeted event count is achieved, 3 the planned minimum follow-up after enrollment is complete, 4 the maximum of planned study duration and targeted event count cuts (1 and 2), 5 the maximum of targeted event count and minimum follow-up cuts (2 and 3).

Examples

library(tidyr)
library(dplyr)
# Show output structure
simtrial:::simfix_(nsim=3)
#>    Events       lnhr         Z                              cut Duration Sim
#> 1     103 -0.5280793 -2.640435                 Planned duration 30.00000   1
#> 2     350 -0.5034721 -4.651405                  Targeted events 77.59402   1
#> 3     340 -0.5022811 -4.572105                Minimum follow-up 74.70954   1
#> 4     350 -0.5034721 -4.651405 Max(planned duration, event cut) 77.59402   1
#> 5     350 -0.5034721 -4.651405    Max(min follow-up, event cut) 77.59402   1
#> 6      90 -0.5224711 -2.449599                 Planned duration 30.00000   2
#> 7     350 -0.3116702 -2.908206                  Targeted events 72.04140   2
#> 8     370 -0.3165746 -3.037855                Minimum follow-up 76.83469   2
#> 9     350 -0.3116702 -2.908206 Max(planned duration, event cut) 72.04140   2
#> 10    370 -0.3165746 -3.037855    Max(min follow-up, event cut) 76.83469   2
#> 11     92 -0.4185950 -1.988818                 Planned duration 30.00000   3
#> 12    350 -0.5436922 -5.041697                  Targeted events 73.28002   3
#> 13    348 -0.5324415 -4.923599                Minimum follow-up 72.95646   3
#> 14    350 -0.5436922 -5.041697 Max(planned duration, event cut) 73.28002   3
#> 15    350 -0.5436922 -5.041697    Max(min follow-up, event cut) 73.28002   3
# Example with 2 tests: logrank and FH(0,1)
simtrial:::simfix_(nsim=1,rg=tibble::tibble(rho=0,gamma=c(0,1)))
#>    Events         lnhr rho gamma           Z        V1        V2
#> 1     103  0.009106792   0     0  0.04614088 1.0000000 0.8265539
#> 2     103  0.009106792   0     1  0.41256497 0.8265539 1.0000000
#> 3     350 -0.462403621   0     0 -4.30200199 1.0000000 0.8475055
#> 4     350 -0.462403621   0     1 -4.67850095 0.8475055 1.0000000
#> 5     366 -0.452064433   0     0 -4.30024567 1.0000000 0.8534493
#> 6     366 -0.452064433   0     1 -4.69260307 0.8534493 1.0000000
#> 7     350 -0.462403621   0     0 -4.30200199 1.0000000 0.8475055
#> 8     350 -0.462403621   0     1 -4.67850095 0.8475055 1.0000000
#> 9     366 -0.452064433   0     0 -4.30024567 1.0000000 0.8534493
#> 10    366 -0.452064433   0     1 -4.69260307 0.8534493 1.0000000
#>                                 cut Duration Sim
#> 1                  Planned duration 30.00000   1
#> 2                  Planned duration 30.00000   1
#> 3                   Targeted events 69.82225   1
#> 4                   Targeted events 69.82225   1
#> 5                 Minimum follow-up 74.93619   1
#> 6                 Minimum follow-up 74.93619   1
#> 7  Max(planned duration, event cut) 69.82225   1
#> 8  Max(planned duration, event cut) 69.82225   1
#> 9     Max(min follow-up, event cut) 74.93619   1
#> 10    Max(min follow-up, event cut) 74.93619   1
# Power by test
# Only use cuts for events, events + min follow-up
xx <- simtrial:::simfix_(nsim=100,timingType=c(2,5),rg=tibble::tibble(rho=0,gamma=c(0,1)))
# Get power approximation for FH, data cutoff combination
xx %>% group_by(cut,rho,gamma) %>% summarise(mean(Z<=qnorm(.025)))
#> `summarise()` has grouped output by 'cut', 'rho'. You can override using the
#> `.groups` argument.
#> # A tibble: 4 × 4
#> # Groups:   cut, rho [2]
#>   cut                             rho gamma `mean(Z <= qnorm(0.025))`
#>   <chr>                         <dbl> <dbl>                     <dbl>
#> 1 Max(min follow-up, event cut)     0     0                      0.96
#> 2 Max(min follow-up, event cut)     0     1                      0.99
#> 3 Targeted events                   0     0                      0.94
#> 4 Targeted events                   0     1                      0.97
# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx %>%  filter(cut != "Targeted events") %>% group_by(Sim) %>% group_map(pMaxCombo)
p <- unlist(p)
mean(p<.025)
#> [1] 0.98
# MaxCombo estimate for targeted events cutoff
p <- unlist(xx %>%  filter(cut == "Targeted events") %>% group_by(Sim) %>% group_map(pMaxCombo))
mean(p<.025)
#> [1] 0.97