Fleming-Harrington Weighted Logrank Tests

With output from the function tensurv

tenFH(
  x = simPWSurv(n = 200) %>% cutDataAtCount(150) %>% tensurv(txval =
    "Experimental"),
  rg = tibble(rho = c(0, 0, 1, 1), gamma = c(0, 1, 0, 1)),
  returnVariance = FALSE
)

Arguments

x
a tensurv-class tibble with a counting process dataset

rg
a tibble with variables rho and gamma, both greater than equal to zero, to specify one Fleming-Harrington weighted logrank test per row; Default: tibble(rho = c(0, 0, 1, 1), gamma = c(0, 1, 0, 1))

returnVariance
a logical flag that, if true, adds columns estimated variance for weighted sum of observed minus expected; see details; Default: FALSE

Value

a `tibble` with rg as input and the FH test statistic for the data in x (Z, a directional square root of the usual weighted logrank test); if variance calculations are specified (e.g., to be used for covariances in a combination test), the this will be returned in the column Var

Details

The input value x produced by tensurv() produces a counting process dataset grouped by strata and sorted within strata by increasing times where events occur.

\(Z\) - standardized normal Fleming-Harrington weighted logrank test

\(i\) - stratum index

\(d_i\) - number of distinct times at which events occurred in stratum \(i\)

\(t_{ij}\) - ordered times at which events in stratum \(i\), \(j=1,2,\ldots d_i\) were observed; for each observation, \(t_{ij}\) represents the time post study entry

O_ij. - total number of events in stratum \(i\) that occurred at time \(t_{ij}\)

O_ije - total number of events in stratum \(i\) in the experimental treatment group that occurred at time \(t_{ij}\)

N_ij. - total number of study subjects in stratum \(i\) who were followed for at least duration

E_ije - expected observations in experimental treatment group given random selection of \(O_{ij.}\) from those in stratum \(i\) at risk at time \(t_{ij}\)

V_ije - hypergeometric variance for \(E_{ije}\) as produced in Var from the tensurv() routine

N_ije - total number of study subjects in stratum \(i\) in the experimental treatment group who were followed for at least duration \(t_{ij}\)

E_ije - expected observations in experimental group in stratum \(i\) at time \(t_{ij}\) conditioning on the overall number of events and at risk populations at that time and sampling at risk observations without replacement: $$E_{ije} = O_{ij.} N_{ije}/N_{ij.}$$

\(S_{ij}\) - Kaplan-Meier estimate of survival in combined treatment groups immediately prior to time \(t_{ij}\)

\(\rho, \gamma\) - real parameters for Fleming-Harrington test

\(X_i\) - Numerator for signed logrank test in stratum \(i\) $$X_i = \sum_{j=1}^d_i S_{ij}^\rho(1-S_{ij}^\gamma)(O_{ije}-E_{ije})$$

V_ij - variance used in denominator for Fleming-Harrington weighted logrank tests $$V_i = \sum_{j=1}^d_i (S_{ij}^\rho(1-S_{ij}^\gamma))^2V_{ij})$$

The stratified Fleming-Harrington weighted logrank test is then computed as: $$Z = \sum_i X_i/\sqrt{\sum_i V_i}$$

Examples

library(tidyr)
# Use default enrollment and event rates at cut at 100 events
x <- simPWSurv(n=200) %>% cutDataAtCount(100) %>% tensurv(txval="Experimental")
# compute logrank (FH(0,0)) and FH(0,1)
tenFH(x,rg=tibble(rho=c(0,0),gamma=c(0,1)))

#> # A tibble: 2 x 3
#>     rho gamma     Z
#>   <dbl> <dbl> <dbl>
#> 1     0     0 -2.37
#> 2     0     1 -2.54

>>>>>>> 8da6ac0d6f0042026e5db4aae1934a972cd63c44