Fleming-Harrington Weighted Logrank Tests plus Correlations

Correlations can be used with mvtnorm::pmvnorm to compute p-value for MaxCombo, the maximum of the specifed Fleming-Harrington tests

tenFHcorr(
  x = simPWSurv(n = 200) %>% cutDataAtCount(100) %>% tensurv(txval =
    "Experimental"),
  rg = tibble(rho = c(0, 0, 1, 1), gamma = c(0, 1, 0, 1)),
  corr = TRUE
)

Arguments

x
a tensurv-class tibble with a counting process dataset

rg
a tibble with variables rho and gamma, both greater than equal to zero, to specify one Fleming-Harrington weighted logrank test per row

corr
a logical; if TRUE (default), return correlation matrix; otherwise, return covariance matrix

Value

a `tibble` with rg as input, the FH test statistics specified for the data in Z, and the correlation or covariance matrix for these tests in variables starting with V

Examples

library(tidyr)
library(dplyr)
# Use default enrollment and event rates at cut of 100 events
x <- simPWSurv(n=200) %>% cutDataAtCount(100) %>% tensurv(txval="Experimental")
# compute logrank (FH(0,0)) and FH(0,1)
x <- tenFHcorr(rg=tibble(rho=c(0,0),gamma=c(0,1)),x=x)
# compute p-value for MaxCombo
library(mvtnorm)
1-pmvnorm(lower=rep(min(x$Z),nrow(x)),corr=data.matrix(select(x,-c(rho,gamma,Z))),
algorithm=GenzBretz(maxpts=50000,abseps=0.00001))[1]

#> [1] 0.001174039

# check that covariance is as expected
x <- simPWSurv(n=200) %>%
         cutDataAtCount(100) %>%
         tensurv(txval="Experimental")
x %>% tenFHcorr(rg=tibble(rho=c(0,0),gamma=c(0,1)),corr=FALSE)

#>   rho gamma         Z        V1       V2
#> 1   0     0 -2.202743 24.438767 7.205294
#> 2   0     1 -2.862075  7.205294 3.038234

# Off-diagonal element should be variance in following
x %>% tenFHcorr(rg=tibble(rho=0,gamma=.5),corr=FALSE)

#>   rho gamma         Z       V1
#> 1   0   0.5 -2.749318 7.205294

# compare off diagonal result with tenFH()
x %>% tenFH(rg=tibble(rho=0,gamma=.5))

#> # A tibble: 1 x 3
#>     rho gamma     Z
#>   <dbl> <dbl> <dbl>
#> 1     0   0.5 -2.75