Cut a Dataset for Analysis at a Specified Date

Cut a Dataset for Analysis at a Specified Date

Usage

cutData(x, cutDate)

Arguments

x

a time-to-event dataset, e.g., generated by simPWSurv

cutDate

date relative to start of randomization (cte from input dataset) at which dataset is to be cut off for analysis

Value

A dataset ready for survival analysis

Examples

# Use default enrollment and event rates and
# cut at calendar time 5 after start of randomization
library(dplyr)
simPWSurv(n = 20) %>% cutData(5)
#> # A tibble: 20 × 4
#> # Groups:   Stratum, Treatment [2]
#>      tte event Stratum Treatment   
#>    <dbl> <dbl> <chr>   <chr>       
#>  1  2.92     1 All     Control     
#>  2  4.79     0 All     Control     
#>  3  1.46     1 All     Experimental
#>  4  4.58     0 All     Experimental
#>  5  3.02     1 All     Control     
#>  6  4.38     1 All     Experimental
#>  7  4.37     0 All     Control     
#>  8  4.33     0 All     Experimental
#>  9  4.07     1 All     Control     
#> 10  4.30     0 All     Experimental
#> 11  4.10     0 All     Control     
#> 12  4.09     0 All     Experimental
#> 13  3.98     0 All     Control     
#> 14  3.98     0 All     Experimental
#> 15  3.90     0 All     Experimental
#> 16  3.77     0 All     Control     
#> 17  3.74     0 All     Control     
#> 18  3.57     0 All     Control     
#> 19  3.55     0 All     Experimental
#> 20  3.52     0 All     Experimental