Cut a dataset for analysis at a specified date

Cut a dataset for analysis at a specified date

Usage

cut_data_by_date(x, cut_date)

Arguments

x

A time-to-event dataset, for example, generated by sim_pw_surv().

cut_date

Date relative to start of randomization (cte from input dataset) at which dataset is to be cut off for analysis.

Value

A dataset ready for survival analysis.

Examples

# Use default enrollment and event rates and
# cut at calendar time 5 after start of randomization
sim_pw_surv(n = 20) |> cut_data_by_date(5)
#>          tte event stratum    treatment
#> 1  4.7974994     0     All experimental
#> 2  4.7750421     0     All experimental
#> 3  4.7325350     0     All      control
#> 4  4.5329518     0     All      control
#> 5  4.4633469     0     All      control
#> 6  0.5081653     1     All      control
#> 7  0.2365449     1     All experimental
#> 8  4.0788610     0     All experimental
#> 9  1.2684367     1     All experimental
#> 10 3.9451478     0     All      control
#> 11 0.6922059     1     All experimental
#> 12 3.8044547     0     All      control
#> 13 3.7759752     0     All      control
#> 14 1.7540322     1     All      control
#> 15 1.5856859     1     All experimental
#> 16 3.6729583     0     All experimental
#> 17 3.4931218     0     All experimental
#> 18 3.4444251     0     All experimental
#> 19 3.3635008     0     All      control
#> 20 3.2398176     0     All      control