MaxCombo p-value

pMaxCombo() computes p-values for the MaxCombo test based on output from simtrial::tenFHcorr(). This is still in an experimental stage and is intended for use with the simtrial::simfix() trial simulation routine. However, it can also be used to analyze clinical trial data such as that provided in the ADaM ADTTE format.

Usage

pMaxCombo(Z, dummy_var, algorithm = GenzBretz(maxpts = 50000, abseps = 1e-05))

Arguments

Z

a dataset output from tenFHcorr(); see examples.

dummy_var

a dummy input that allows group_map() to be used to compute p-values for multiple simulations.

algorithm

This is passed directly to the algorithm argument in the mvtnorm::pmvnorm()

Value

A numeric p-value

Examples

library(tidyr)
library(tibble)
library(dplyr)

# example 1
x <- simfix(nsim = 1,
            timingType = 5,
            rg = tibble(rho = c(0, 0, 1),
                        gamma = c(0, 1, 1)))
head(x)
#>   Events       lnhr rho gamma         Z        V1        V2        V3
#> 1    354 -0.4667545   0     0 -4.373377 1.0000000 0.8545091 0.9328463
#> 2    354 -0.4667545   0     1 -3.733605 0.8545091 1.0000000 0.9411226
#> 3    354 -0.4667545   1     1 -4.480740 0.9328463 0.9411226 1.0000000
#>                             cut Duration Sim
#> 1 Max(min follow-up, event cut) 74.19343   1
#> 2 Max(min follow-up, event cut) 74.19343   1
#> 3 Max(min follow-up, event cut) 74.19343   1
pMaxCombo(x)
#> [1] 3.727767e-06

# example 2
# Only use cuts for events, events + min follow-up
xx <- simfix(nsim = 100,
             timingType = 5,
             rg = tibble(rho = c(0, 0, 1),
                         gamma = c(0, 1, 1)))
head(xx)
#>   Events       lnhr rho gamma         Z        V1        V2        V3
#> 1    355 -0.2955135   0     0 -2.776648 1.0000000 0.8561885 0.9340511
#> 2    355 -0.2955135   0     1 -2.825529 0.8561885 1.0000000 0.9376215
#> 3    355 -0.2955135   1     1 -3.154769 0.9340511 0.9376215 1.0000000
#> 4    367 -0.2861542   0     0 -2.725323 1.0000000 0.8535956 0.9341805
#> 5    367 -0.2861542   0     1 -3.814155 0.8535956 1.0000000 0.9199942
#> 6    367 -0.2861542   1     1 -3.822690 0.9341805 0.9199942 1.0000000
#>                             cut Duration Sim
#> 1 Max(min follow-up, event cut) 75.82148   1
#> 2 Max(min follow-up, event cut) 75.82148   1
#> 3 Max(min follow-up, event cut) 75.82148   1
#> 4 Max(min follow-up, event cut) 75.13340   2
#> 5 Max(min follow-up, event cut) 75.13340   2
#> 6 Max(min follow-up, event cut) 75.13340   2
# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx %>% group_by(Sim) %>% group_map(pMaxCombo) %>% unlist()
mean(p < .025)
#> [1] 0.98