MaxCombo p-value

pMaxCombo() computes p-values for the MaxCombo test based on output from simtrial::tenFHcorr(). This is still in an experimental stage and is intended for use with the simtrial::simfix() trial simulation routine. However, it can also be used to analyze clinical trial data such as that provided in the ADaM ADTTE format.

Usage

pMaxCombo(Z, dummyvar, algorithm = GenzBretz(maxpts = 50000, abseps = 1e-05))

Arguments

Z

a dataset output from tenFHcorr(); see examples.

dummyvar

a dummy input that allows group_map() to be used to compute p-values for multiple simulations.

algorithm

This is passed directly to the algorithm argument in the mvtnorm::pmvnorm()

Value

A numeric p-value

Examples

library(tidyr)
x <- simfix(nsim=1,timingType=5,rg=tibble::tibble(rho=c(0,0,1),gamma=c(0,1,1)))
head(x)
#>   Events       lnhr rho gamma         Z        V1        V2        V3
#> 1    361 -0.3322218   0     0 -3.146791 1.0000000 0.8571103 0.9347641
#> 2    361 -0.3322218   0     1 -3.318562 0.8571103 1.0000000 0.9379794
#> 3    361 -0.3322218   1     1 -3.798803 0.9347641 0.9379794 1.0000000
#>                             cut Duration Sim
#> 1 Max(min follow-up, event cut) 72.16034   1
#> 2 Max(min follow-up, event cut) 72.16034   1
#> 3 Max(min follow-up, event cut) 72.16034   1
pMaxCombo(x)
#> [1] 7.351073e-05
# Only use cuts for events, events + min follow-up
xx <- simfix(nsim=100,timingType=5,rg=tibble::tibble(rho=c(0,0,1),gamma=c(0,1,1)))
head(xx)
#>   Events       lnhr rho gamma         Z        V1        V2        V3
#> 1    370 -0.3088979   0     0 -2.961613 1.0000000 0.8560167 0.9356853
#> 2    370 -0.3088979   0     1 -3.221966 0.8560167 1.0000000 0.9283988
#> 3    370 -0.3088979   1     1 -3.514724 0.9356853 0.9283988 1.0000000
#> 4    350 -0.4692477   0     0 -4.356558 1.0000000 0.8549328 0.9321061
#> 5    350 -0.4692477   0     1 -4.874285 0.8549328 1.0000000 0.9425538
#> 6    350 -0.4692477   1     1 -4.712256 0.9321061 0.9425538 1.0000000
#>                             cut Duration Sim
#> 1 Max(min follow-up, event cut) 75.02873   1
#> 2 Max(min follow-up, event cut) 75.02873   1
#> 3 Max(min follow-up, event cut) 75.02873   1
#> 4 Max(min follow-up, event cut) 71.81227   2
#> 5 Max(min follow-up, event cut) 71.81227   2
#> 6 Max(min follow-up, event cut) 71.81227   2
# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- unlist(xx %>%  dplyr::group_by(Sim) %>% dplyr::group_map(pMaxCombo))
mean(p<.025)
#> [1] 0.98