MaxCombo p-value

Computes p-values for the MaxCombo test based on output from fh_weight(). This is still in an experimental stage and is intended for use with the sim_fixed_n() trial simulation routine. However, it can also be used to analyze clinical trial data such as that provided in the ADaM ADTTE format.

Usage

pvalue_maxcombo(
  z,
  algorithm = mvtnorm::GenzBretz(maxpts = 50000, abseps = 1e-05)
)

Arguments

z

A dataset output from fh_weight(); see examples.

algorithm

This is passed directly to the algorithm argument in mvtnorm::pmvnorm().

Value

A numeric p-value.

Examples

library(dplyr)

# Example 1
x <- sim_fixed_n(
  n_sim = 1,
  timing_type = 5,
  rho_gamma = data.frame(
    rho = c(0, 0, 1),
    gamma = c(0, 1, 1)
  )
)
#> Backend uses sequential processing.
#> Loading required package: foreach
#> Loading required package: future
#> 
#> Attaching package: ‘future’
#> The following object is masked from ‘package:survival’:
#> 
#>     cluster
head(x)
#>   event      ln_hr rho gamma         z        v1        v2        v3
#> 1   359 -0.3345059   0     0 -3.151779 1.0000000 0.8545668 0.9336124
#> 2   359 -0.3345059   0     1 -3.163773 0.8545668 1.0000000 0.9337118
#> 3   359 -0.3345059   1     1 -3.376780 0.9336124 0.9337118 1.0000000
#>                             cut duration sim
#> 1 Max(min follow-up, event cut) 76.78496   1
#> 2 Max(min follow-up, event cut) 76.78496   1
#> 3 Max(min follow-up, event cut) 76.78496   1
pvalue_maxcombo(x)
#> [1] 0.0006909124

# Example 2
# Only use cuts for events, events + min follow-up
xx <- sim_fixed_n(
  n_sim = 100,
  timing_type = 5,
  rho_gamma = data.frame(
    rho = c(0, 0, 1),
    gamma = c(0, 1, 1)
  )
)
#> Backend uses sequential processing.
head(xx)
#>   event      ln_hr rho gamma         z        v1        v2        v3
#> 1   356 -0.3128620   0     0 -2.945655 1.0000000 0.8514028 0.9333614
#> 2   356 -0.3128620   0     1 -2.817452 0.8514028 1.0000000 0.9224452
#> 3   356 -0.3128620   1     1 -3.166569 0.9333614 0.9224452 1.0000000
#> 4   350 -0.4061278   0     0 -3.783454 1.0000000 0.8554978 0.9327910
#> 5   350 -0.4061278   0     1 -3.655435 0.8554978 1.0000000 0.9432697
#> 6   350 -0.4061278   1     1 -3.741710 0.9327910 0.9432697 1.0000000
#>                             cut duration sim
#> 1 Max(min follow-up, event cut) 75.29919   1
#> 2 Max(min follow-up, event cut) 75.29919   1
#> 3 Max(min follow-up, event cut) 75.29919   1
#> 4 Max(min follow-up, event cut) 74.41309   2
#> 5 Max(min follow-up, event cut) 74.41309   2
#> 6 Max(min follow-up, event cut) 74.41309   2

# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx |>
  group_by(sim) |>
  group_map(~ pvalue_maxcombo(.x)) |>
  unlist()
mean(p < .025)
#> [1] 0.98