MaxCombo p-value

Computes p-values for the MaxCombo test based on output from fh_weight(). This is still in an experimental stage and is intended for use with the sim_fixed_n() trial simulation routine. However, it can also be used to analyze clinical trial data such as that provided in the ADaM ADTTE format.

Usage

pvalue_maxcombo(
  z,
  algorithm = mvtnorm::GenzBretz(maxpts = 50000, abseps = 1e-05)
)

Arguments

z

A dataset output from fh_weight(); see examples.

algorithm

This is passed directly to the algorithm argument in mvtnorm::pmvnorm().

Value

A numeric p-value.

Examples

library(dplyr)

# Example 1
x <- sim_fixed_n(
  n_sim = 1,
  timing_type = 5,
  rho_gamma = data.frame(
    rho = c(0, 0, 1),
    gamma = c(0, 1, 1)
  )
)
#> Backend uses sequential processing.
#> Loading required package: foreach
#> Loading required package: future
#> 
#> Attaching package: ‘future’
#> The following object is masked from ‘package:survival’:
#> 
#>     cluster
head(x)
#>   event      ln_hr rho gamma         z        v1        v2        v3
#> 1   350 -0.3913266   0     0 -3.637665 1.0000000 0.8523636 0.9324825
#> 2   350 -0.3913266   0     1 -4.420310 0.8523636 1.0000000 0.9321649
#> 3   350 -0.3913266   1     1 -4.540739 0.9324825 0.9321649 1.0000000
#>                             cut duration sim
#> 1 Max(min follow-up, event cut) 73.05896   1
#> 2 Max(min follow-up, event cut) 73.05896   1
#> 3 Max(min follow-up, event cut) 73.05896   1
pvalue_maxcombo(x)
#> [1] 2.806222e-06

# Example 2
# Only use cuts for events, events + min follow-up
xx <- sim_fixed_n(
  n_sim = 100,
  timing_type = 5,
  rho_gamma = data.frame(
    rho = c(0, 0, 1),
    gamma = c(0, 1, 1)
  )
)
#> Backend uses sequential processing.
head(xx)
#>   event      ln_hr rho gamma         z        v1        v2        v3
#> 1   367 -0.3603407   0     0 -3.433180 1.0000000 0.8508939 0.9329539
#> 2   367 -0.3603407   0     1 -3.227307 0.8508939 1.0000000 0.9168869
#> 3   367 -0.3603407   1     1 -3.707276 0.9329539 0.9168869 1.0000000
#> 4   370 -0.3551767   0     0 -3.406090 1.0000000 0.8523183 0.9339574
#> 5   370 -0.3551767   0     1 -3.207531 0.8523183 1.0000000 0.9165166
#> 6   370 -0.3551767   1     1 -3.310361 0.9339574 0.9165166 1.0000000
#>                             cut duration sim
#> 1 Max(min follow-up, event cut) 77.50178   1
#> 2 Max(min follow-up, event cut) 77.50178   1
#> 3 Max(min follow-up, event cut) 77.50178   1
#> 4 Max(min follow-up, event cut) 71.48793   2
#> 5 Max(min follow-up, event cut) 71.48793   2
#> 6 Max(min follow-up, event cut) 71.48793   2

# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx |>
  group_by(sim) |>
  group_map(~ pvalue_maxcombo(.x)) |>
  unlist()
mean(p < .025)
#> [1] 0.98