Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0805 Experimental  38.1          402.  38.2      1
#>  2 All         0.131  Control        2.12         223.   2.26     1
#>  3 All         0.411  Control        0.691        845.   1.10     1
#>  4 All         0.414  Experimental  24.9          610.  25.3      1
#>  5 All         0.533  Experimental   1.43        1240.   1.96     1
#>  6 All         0.667  Experimental  56.9           95.3 57.6      1
#>  7 All         0.737  Control       15.1          258.  15.9      1
#>  8 All         1.12   Control        0.565        457.   1.69     1
#>  9 All         1.20   Control        1.87         261.   3.07     1
#> 10 All         1.26   Experimental  31.0          468.  32.3      1
#> 11 All         1.43   Experimental  23.3         1704.  24.7      1
#> 12 All         1.68   Control        6.61         522.   8.29     1
#> 13 All         1.87   Experimental   1.23        1003.   3.10     1
#> 14 All         1.97   Experimental   1.12         709.   3.09     1
#> 15 All         1.97   Control       13.7           86.9 15.7      1
#> 16 All         1.98   Control       23.6          957.  25.6      1
#> 17 All         2.14   Control        0.541       3336.   2.68     1
#> 18 All         2.34   Experimental   0.0946       651.   2.43     1
#> 19 All         2.36   Experimental  17.8           95.2 20.2      1
#> 20 All         2.55   Control       22.1          766.  24.7      1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All         0.0222 Control        0.493        259.   0.516     1
#>  2 All         0.0845 Experimental  15.6          198.  15.6       1
#>  3 All         0.189  Experimental   0.105        490.   0.294     1
#>  4 All         0.211  Experimental  14.3          217.  14.5       1
#>  5 All         0.286  Experimental  10.1         1051.  10.4       1
#>  6 All         0.362  Experimental  36.7          369.  37.1       1
#>  7 All         0.385  Experimental  24.6         1747.  25.0       1
#>  8 All         0.789  Control        0.0109       441.   0.800     1
#>  9 All         0.873  Experimental   0.982        534.   1.86      1
#> 10 All         0.891  Experimental  49.8          177.  50.7       1
#> 11 All         0.906  Experimental   1.72        2425.   2.62      1
#> 12 All         1.38   Control       14.5          724.  15.9       1
#> 13 All         1.42   Experimental   5.28        1093.   6.70      1
#> 14 All         1.42   Experimental  77.9          933.  79.3       1
#> 15 All         1.47   Control        8.68         298.  10.2       1
#> 16 All         1.52   Experimental   9.67         547.  11.2       1
#> 17 All         1.58   Experimental   1.16          98.2  2.74      1
#> 18 All         1.60   Experimental  29.7          374.  31.3       1
#> 19 All         1.97   Experimental   1.92         506.   3.88      1
#> 20 All         2.28   Control        8.44         355.  10.7       1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low          0.164 Control        11.3         812.  11.5      1
#>  2 High         0.221 Experimental   38.1        1568.  38.3      1
#>  3 High         0.227 Control         9.71        344.   9.94     1
#>  4 High         0.626 Experimental    3.47        422.   4.10     1
#>  5 High         0.799 Control         0.764       211.   1.56     1
#>  6 High         0.987 Experimental   31.4         687.  32.4      1
#>  7 High         0.998 Control        71.2          24.1 25.1      0
#>  8 High         1.08  Experimental   18.6        2046.  19.7      1
#>  9 High         1.17  Control        38.6        1404.  39.8      1
#> 10 High         1.23  Experimental   17.8          45.7 19.0      1
#> 11 High         1.47  Experimental   58.1        2221.  59.6      1
#> 12 Low          1.54  Control        26.5         119.  28.0      1
#> 13 High         1.57  Control        10.3         251.  11.9      1
#> 14 High         1.60  Control         6.97        470.   8.56     1
#> 15 High         1.60  Control        18.2        1451.  19.8      1
#> 16 High         1.67  Control        51.6         860.  53.3      1
#> 17 High         1.75  Experimental   46.7         106.  48.5      1
#> 18 High         1.85  Experimental    6.39       2183.   8.24     1
#> 19 Low          2.22  Experimental   49.6         581.  51.8      1
#> 20 High         2.30  Control        17.5         223.  19.8      1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low         0.0921 Experimental     3.13      634.    3.22     1
#>  2 High        0.225  Control          6.17      814.    6.39     1
#>  3 High        0.269  Experimental    18.0         6.84  7.11     0
#>  4 High        0.301  Control          1.71      168.    2.01     1
#>  5 High        0.312  Experimental    11.7       817.   12.0      1
#>  6 High        0.492  Control          2.54     2040.    3.04     1
#>  7 High        0.547  Experimental    30.5      1615.   31.0      1
#>  8 High        0.548  Experimental    11.9        22.9  12.4      1
#>  9 High        0.585  Control          6.05      608.    6.64     1
#> 10 High        0.689  Control         13.2      1377.   13.9      1
#> 11 High        0.734  Experimental     1.93      118.    2.66     1
#> 12 Low         0.735  Control          3.97      156.    4.71     1