Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0124 Experimental  18.5         1403.  18.6      1
#>  2 All         0.0476 Experimental  32.8          476.  32.9      1
#>  3 All         0.182  Control       24.7         1696.  24.9      1
#>  4 All         0.675  Control        2.21         961.   2.88     1
#>  5 All         0.738  Experimental  62.1          886.  62.9      1
#>  6 All         0.803  Control       16.2          193.  17.0      1
#>  7 All         0.846  Experimental   7.20        2515.   8.05     1
#>  8 All         0.883  Control        0.774        308.   1.66     1
#>  9 All         1.04   Experimental   0.253        208.   1.29     1
#> 10 All         1.09   Control       17.6          602.  18.7      1
#> 11 All         1.18   Control        1.41         206.   2.59     1
#> 12 All         1.31   Experimental  43.2         4162.  44.5      1
#> 13 All         1.34   Experimental  13.1          927.  14.4      1
#> 14 All         1.73   Control       16.7          131.  18.4      1
#> 15 All         1.74   Control        0.0196      1155.   1.76     1
#> 16 All         1.83   Experimental  20.2          633.  22.0      1
#> 17 All         1.89   Control       32.2         1344.  34.1      1
#> 18 All         1.94   Control        5.41        1016.   7.35     1
#> 19 All         1.94   Experimental  52.9           64.5 54.9      1
#> 20 All         2.00   Experimental  37.1          956.  39.1      1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0362 Experimental    11.9        239.  11.9      1
#>  2 All         0.0684 Control         45.4        747.  45.4      1
#>  3 All         0.302  Experimental    23.5       3606.  23.8      1
#>  4 All         0.423  Experimental     6.59      1036.   7.02     1
#>  5 All         0.496  Control          2.19      1292.   2.69     1
#>  6 All         0.726  Experimental     9.25       502.   9.98     1
#>  7 All         0.762  Experimental    28.0       2174.  28.7      1
#>  8 All         1.02   Experimental    12.9         37.9 13.9      1
#>  9 All         1.14   Control         11.3        581.  12.5      1
#> 10 All         1.17   Experimental    21.2        525.  22.4      1
#> 11 All         1.25   Experimental     5.29       270.   6.54     1
#> 12 All         1.45   Experimental     5.06       722.   6.51     1
#> 13 All         1.46   Experimental     6.03       940.   7.49     1
#> 14 All         1.70   Control          4.78       459.   6.49     1
#> 15 All         1.70   Experimental     2.33      1679.   4.03     1
#> 16 All         1.89   Experimental    49.0        207.  50.9      1
#> 17 All         2.21   Control          8.63       292.  10.8      1
#> 18 All         2.32   Experimental     5.94       203.   8.26     1
#> 19 All         2.43   Experimental    63.9       2150.  66.4      1
#> 20 All         2.81   Experimental    43.3        707.  46.1      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low          0.169 Experimental   55.2        2597.  55.4      1
#>  2 Low          0.294 Control        10.9         168.  11.2      1
#>  3 Low          0.484 Experimental    9.12        716.   9.60     1
#>  4 High         0.534 Control        11.5          29.0 12.1      1
#>  5 High         0.564 Experimental   16.2         322.  16.8      1
#>  6 High         0.800 Experimental    0.993      1527.   1.79     1
#>  7 Low          0.832 Control        13.7        1584.  14.5      1
#>  8 Low          0.991 Experimental   80.0        1339.  81.0      1
#>  9 High         1.01  Control        13.9          23.6 15.0      1
#> 10 Low          1.11  Experimental   11.0         795.  12.1      1
#> 11 High         1.19  Experimental    0.491      1912.   1.68     1
#> 12 Low          1.20  Control         0.818      1065.   2.02     1
#> 13 High         1.27  Control         8.43        525.   9.71     1
#> 14 Low          1.28  Control        20.1         223.  21.3      1
#> 15 High         1.30  Experimental   35.9         527.  37.2      1
#> 16 High         1.32  Control        78.2         757.  79.5      1
#> 17 High         1.35  Control         9.92        104.  11.3      1
#> 18 Low          1.44  Control        16.1        1259.  17.5      1
#> 19 High         1.54  Experimental   25.4         324.  27.0      1
#> 20 Low          1.64  Experimental    3.20        899.   4.84     1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [3]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High        0.0655 Experimental    29.4        214.   29.5      1
#>  2 Low         0.253  Control          3.52       388.    3.77     1
#>  3 High        0.328  Experimental     9.46       767.    9.79     1
#>  4 High        0.364  Control         48.8       1074.   49.2      1
#>  5 High        0.517  Control          3.54      2717.    4.06     1
#>  6 High        0.638  Control          8.51        48.5   9.15     1
#>  7 High        0.726  Experimental    67.9        469.   68.6      1
#>  8 High        0.739  Control         21.4        552.   22.1      1
#>  9 High        0.950  Experimental     5.14      1233.    6.09     1
#> 10 High        1.13   Experimental    15.1       2961.   16.3      1
#> 11 High        1.35   Experimental   142.         218.  143.       1
#> 12 High        1.36   Control         60.7       1159.   62.0      1