Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0489 Control         47.6      2068.   47.7      1
#>  2 All         0.197  Experimental    32.7       482.   32.9      1
#>  3 All         0.402  Control          7.91      557.    8.31     1
#>  4 All         0.846  Experimental    23.0       385.   23.8      1
#>  5 All         1.18   Experimental     7.12       94.5   8.30     1
#>  6 All         1.18   Experimental     1.33      136.    2.52     1
#>  7 All         1.19   Control         50.5       335.   51.6      1
#>  8 All         1.35   Control         13.2      2671.   14.6      1
#>  9 All         1.40   Experimental    35.3        71.2  36.7      1
#> 10 All         1.63   Control         17.6       557.   19.2      1
#> 11 All         1.73   Experimental    38.7      1818.   40.4      1
#> 12 All         1.74   Control          3.31      488.    5.05     1
#> 13 All         1.77   Experimental    50.4       579.   52.1      1
#> 14 All         1.91   Experimental    19.7      1883.   21.6      1
#> 15 All         2.03   Control         13.8         9.19 11.2      0
#> 16 All         2.14   Control          6.08      225.    8.21     1
#> 17 All         2.19   Control         10.2      1312.   12.4      1
#> 18 All         2.30   Experimental    10.9      1174.   13.2      1
#> 19 All         2.35   Control          7.01      162.    9.35     1
#> 20 All         2.38   Experimental    58.1       470.   60.4      1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All         0.0725 Control       12.7           80.0 12.8       1
#>  2 All         0.119  Experimental  36.4         2443.  36.5       1
#>  3 All         0.216  Experimental  73.9          402.  74.1       1
#>  4 All         0.251  Experimental   6.23         158.   6.48      1
#>  5 All         0.302  Experimental  13.2           83.6 13.5       1
#>  6 All         0.405  Control       10.1         1921.  10.5       1
#>  7 All         0.410  Experimental   0.555       1194.   0.965     1
#>  8 All         0.609  Experimental  15.9          322.  16.5       1
#>  9 All         0.790  Experimental  26.1          471.  26.9       1
#> 10 All         0.834  Control        0.0114       867.   0.845     1
#> 11 All         0.968  Experimental  21.7         1509.  22.7       1
#> 12 All         1.02   Experimental   0.615       1367.   1.64      1
#> 13 All         1.09   Experimental  51.2         1872.  52.3       1
#> 14 All         1.10   Experimental  23.2         1099.  24.3       1
#> 15 All         1.10   Control        8.79        1908.   9.90      1
#> 16 All         1.12   Experimental  11.4          690.  12.5       1
#> 17 All         1.12   Experimental  13.3         1101.  14.4       1
#> 18 All         1.17   Control       17.6          911.  18.8       1
#> 19 All         1.46   Experimental   3.84         752.   5.30      1
#> 20 All         2.01   Experimental  27.1          218.  29.1       1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.173 Control         6.46       2296.   6.63     1
#>  2 High         0.237 Control        18.4         791.  18.6      1
#>  3 Low          0.370 Control         3.31        263.   3.68     1
#>  4 High         0.691 Experimental   15.8          80.2 16.5      1
#>  5 High         0.773 Experimental   31.2         335.  32.0      1
#>  6 High         0.797 Experimental   39.9        1551.  40.7      1
#>  7 High         0.802 Control        34.7         102.  35.5      1
#>  8 Low          0.803 Control         6.91       1803.   7.71     1
#>  9 High         0.830 Control         8.65       1285.   9.48     1
#> 10 High         0.834 Experimental   55.7         356.  56.5      1
#> 11 High         0.988 Experimental    0.888      1580.   1.88     1
#> 12 High         0.997 Control         6.98        276.   7.98     1
#> 13 High         1.07  Control        22.8        1779.  23.9      1
#> 14 High         1.25  Experimental    2.96        135.   4.21     1
#> 15 High         1.26  Experimental    2.57        416.   3.83     1
#> 16 Low          1.35  Experimental    4.82        144.   6.17     1
#> 17 High         1.41  Control        15.4        3250.  16.8      1
#> 18 High         1.46  Control         3.28       1060.   4.75     1
#> 19 High         1.57  Experimental   35.2         752.  36.8      1
#> 20 Low          1.70  Experimental   14.1         740.  15.8      1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [3]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High       0.00569 Control         36.8       3928.   36.8      1
#>  2 High       0.168   Experimental   141.        1611.  141.       1
#>  3 High       0.280   Control         37.6        450.   37.9      1
#>  4 High       0.352   Experimental    44.5       1401.   44.9      1
#>  5 High       0.476   Experimental    36.9         43.6  37.4      1
#>  6 High       0.503   Control         55.5       2423.   56.0      1
#>  7 High       0.526   Control          3.78       743.    4.31     1
#>  8 Low        0.607   Control         42.2       1087.   42.8      1
#>  9 High       0.626   Experimental     2.54       557.    3.16     1
#> 10 High       1.09    Experimental    76.5       1029.   77.6      1
#> 11 High       1.27    Control         26.8        665.   28.1      1
#> 12 High       1.27    Control          1.54       810.    2.81     1