Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.302 Experimental    1.42        1591.  1.73     1
#>  2 All          0.519 Experimental   57.3         2629. 57.8      1
#>  3 All          0.713 Control        41.2          843. 41.9      1
#>  4 All          0.781 Control        33.2         1868. 34.0      1
#>  5 All          0.814 Control        65.5          860. 66.3      1
#>  6 All          0.857 Control        17.4          309. 18.2      1
#>  7 All          1.00  Experimental   29.9         1069. 30.9      1
#>  8 All          1.14  Experimental    2.16         624.  3.31     1
#>  9 All          1.51  Experimental   52.4         1265. 54.0      1
#> 10 All          1.64  Control         3.86         693.  5.50     1
#> 11 All          1.78  Control         6.00        3801.  7.79     1
#> 12 All          1.84  Experimental   12.4          338. 14.3      1
#> 13 All          1.90  Experimental   14.0          346. 15.9      1
#> 14 All          1.93  Control         5.31         990.  7.25     1
#> 15 All          2.39  Experimental    2.38         202.  4.76     1
#> 16 All          2.54  Control         7.85        4197. 10.4      1
#> 17 All          2.54  Control         0.654       1924.  3.19     1
#> 18 All          2.65  Experimental    6.49         553.  9.15     1
#> 19 All          2.65  Control        18.4          178. 21.0      1
#> 20 All          2.71  Experimental   16.1          473. 18.8      1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 All          0.145 Control        0.284         847.   0.429     1
#>  2 All          0.284 Experimental  32.1           213.  32.4       1
#>  3 All          0.297 Experimental   0.0924       1167.   0.390     1
#>  4 All          0.355 Experimental  36.4          4154.  36.8       1
#>  5 All          0.757 Experimental  23.0          1066.  23.8       1
#>  6 All          0.802 Experimental   0.402        3927.   1.20      1
#>  7 All          0.955 Control        3.63          595.   4.59      1
#>  8 All          1.04  Experimental   8.16          855.   9.20      1
#>  9 All          1.16  Experimental   6.74         1470.   7.90      1
#> 10 All          1.20  Experimental   0.0257        677.   1.22      1
#> 11 All          1.71  Control       27.1           123.  28.9       1
#> 12 All          1.80  Experimental   9.31         2997.  11.1       1
#> 13 All          1.81  Experimental   6.71          135.   8.52      1
#> 14 All          1.85  Experimental  99.8           404. 102.        1
#> 15 All          1.92  Experimental   5.00         1669.   6.92      1
#> 16 All          1.92  Control        7.05         3296.   8.97      1
#> 17 All          1.96  Control        7.03          761.   8.99      1
#> 18 All          2.15  Experimental   2.12         1142.   4.27      1
#> 19 All          2.32  Experimental  48.3           294.  50.6       1
#> 20 All          2.46  Experimental   5.85         1914.   8.31      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [3]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High        0.0928 Experimental    4.88       2234.    4.98     1
#>  2 High        0.0980 Control        38.6         760.   38.7      1
#>  3 High        0.139  Control         7.46       5455.    7.59     1
#>  4 High        0.236  Experimental  124.          343.  125.       1
#>  5 High        0.239  Experimental    2.59         88.0   2.83     1
#>  6 High        0.254  Control         9.69        539.    9.95     1
#>  7 Low         0.445  Control         2.12       1466.    2.57     1
#>  8 High        0.717  Control         0.859       922.    1.58     1
#>  9 High        0.752  Experimental   17.5        1506.   18.3      1
#> 10 High        0.863  Control        17.2        1094.   18.1      1
#> 11 High        1.47   Control        10.7          63.1  12.2      1
#> 12 Low         1.54   Control        27.6         698.   29.1      1
#> 13 High        1.61   Experimental   21.6        1624.   23.2      1
#> 14 High        1.64   Experimental   19.4         665.   21.1      1
#> 15 High        1.70   Experimental   12.2        2023.   13.9      1
#> 16 High        1.91   Experimental    4.64       1241.    6.55     1
#> 17 High        2.16   Control        39.7         868.   41.8      1
#> 18 High        2.52   Control        11.1          27.6  13.7      1
#> 19 High        2.63   Experimental    2.19       1241.    4.83     1
#> 20 High        2.79   Control         8.39        434.   11.2      1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [3]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 High        0.0223 Experimental   55.5         474.   55.6       1
#>  2 High        0.0261 Control        21.2         683.   21.3       1
#>  3 High        0.0911 Control        42.9         253.   43.0       1
#>  4 Low         0.131  Control        22.5        4048.   22.7       1
#>  5 High        0.246  Experimental  129.          103.  103.        0
#>  6 High        0.255  Control         0.648       375.    0.903     1
#>  7 High        0.550  Experimental    5.59        599.    6.14      1
#>  8 High        0.589  Control         4.03       1296.    4.62      1
#>  9 High        0.720  Experimental   10.4         755.   11.1       1
#> 10 High        0.881  Experimental   25.3          17.9  18.8       0
#> 11 High        1.02   Control         7.14         18.3   8.16      1
#> 12 High        1.06   Control         1.60        193.    2.66      1