Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0337 Experimental  15.0        1234.   15.0      1
#>  2 All          0.0863 Control        7.38        535.    7.46     1
#>  3 All          0.221  Control        1.97       1951.    2.20     1
#>  4 All          0.302  Experimental  13.7         487.   14.0      1
#>  5 All          0.312  Control        0.951       782.    1.26     1
#>  6 All          0.322  Experimental  37.5           4.40  4.72     0
#>  7 All          0.331  Control       10.1         491.   10.4      1
#>  8 All          0.387  Experimental   7.54        182.    7.93     1
#>  9 All          0.760  Experimental   7.86        236.    8.62     1
#> 10 All          0.804  Control       30.7         131.   31.5      1
#> 11 All          1.01   Experimental  64.9        1688.   65.9      1
#> 12 All          1.02   Control       19.2         419.   20.2      1
#> 13 All          1.14   Control        5.87       1046.    7.01     1
#> 14 All          1.34   Experimental  10.4         306.   11.7      1
#> 15 All          1.38   Control       24.0        2563.   25.4      1
#> 16 All          1.60   Experimental  27.6        1549.   29.2      1
#> 17 All          1.61   Control       12.2         694.   13.8      1
#> 18 All          1.65   Experimental   0.0649      353.    1.71     1
#> 19 All          1.69   Experimental   1.45        851.    3.14     1
#> 20 All          1.75   Control        5.85        182.    7.60     1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0181 Control         1.87       1369.   1.89     1
#>  2 All          0.106  Experimental   13.6         857.  13.7      1
#>  3 All          0.108  Experimental    6.32         58.1  6.43     1
#>  4 All          0.177  Experimental   39.1        1234.  39.3      1
#>  5 All          0.243  Experimental   34.9        3344.  35.1      1
#>  6 All          0.264  Experimental    2.55        716.   2.82     1
#>  7 All          0.452  Experimental   84.8         308.  85.2      1
#>  8 All          0.540  Control         0.551       227.   1.09     1
#>  9 All          0.548  Control        11.7        1694.  12.2      1
#> 10 All          0.558  Experimental   64.2         515.  64.8      1
#> 11 All          0.589  Experimental   49.9          87.7 50.5      1
#> 12 All          0.617  Experimental    9.20        982.   9.82     1
#> 13 All          0.727  Experimental   43.6        2112.  44.3      1
#> 14 All          0.767  Control        19.0         381.  19.8      1
#> 15 All          0.898  Experimental   58.7         705.  59.6      1
#> 16 All          0.968  Experimental   16.0         248.  17.0      1
#> 17 All          1.09   Control        17.4         125.  18.5      1
#> 18 All          1.46   Experimental   23.8        1606.  25.3      1
#> 19 All          1.54   Experimental    2.89        157.   4.43     1
#> 20 All          1.54   Experimental   16.9        1079.  18.4      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enroll_rate
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.0403 Experimental   5.85         420.   5.89     1
#>  2 High         0.251  Control       41.4          249.  41.6      1
#>  3 High         0.400  Control       11.5          447.  11.9      1
#>  4 High         0.409  Experimental  52.6          159.  53.0      1
#>  5 High         0.500  Control        3.56        1100.   4.06     1
#>  6 High         0.788  Experimental  86.9          170.  87.7      1
#>  7 High         0.842  Experimental   0.246        285.   1.09     1
#>  8 High         0.924  Control        2.92         236.   3.85     1
#>  9 High         1.00   Experimental  70.6          691.  71.6      1
#> 10 Low          1.09   Experimental   0.527       3380.   1.61     1
#> 11 High         1.13   Experimental   7.93         650.   9.06     1
#> 12 High         1.22   Control        1.55         254.   2.77     1
#> 13 Low          1.24   Control        0.0635       773.   1.31     1
#> 14 High         1.31   Control        9.44         422.  10.7      1
#> 15 High         1.36   Control        3.03        1657.   4.39     1
#> 16 Low          1.45   Control       14.3          607.  15.7      1
#> 17 Low          1.47   Experimental  16.1          172.  17.6      1
#> 18 High         1.48   Control        7.83          46.0  9.31     1
#> 19 High         1.59   Experimental   8.09         949.   9.67     1
#> 20 Low          1.75   Control       14.3          908.  16.0      1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    failTime dropoutTime     cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 High         0.0911 Experimental    0.291       353.    0.382     1
#>  2 High         0.272  Control         5.27       2228.    5.55      1
#>  3 Low          0.566  Experimental   40.0        1585.   40.6       1
#>  4 High         0.733  Experimental  179.           86.0  86.7       0
#>  5 Low          0.742  Control        28.1         633.   28.8       1
#>  6 High         0.743  Control        34.1        2162.   34.8       1
#>  7 Low          1.03   Experimental   49.9        1309.   50.9       1
#>  8 High         1.08   Control        15.4         864.   16.5       1
#>  9 High         1.16   Control         3.98       1129.    5.14      1
#> 10 High         1.30   Experimental    0.513       331.    1.81      1
#> 11 Low          1.38   Control       129.          116.  117.        0
#> 12 High         1.42   Experimental   22.2        2052.   23.6       1