Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.102 Experimental   20.2         759.   20.3      1
#>  2 All          0.114 Control         6.21        197.    6.33     1
#>  3 All          0.200 Control        21.1        1376.   21.3      1
#>  4 All          0.347 Experimental   23.4        1953.   23.8      1
#>  5 All          0.431 Control         5.17        697.    5.60     1
#>  6 All          0.463 Control         4.82        419.    5.28     1
#>  7 All          0.562 Experimental    1.36       1854.    1.92     1
#>  8 All          0.584 Experimental   47.3         127.   47.9      1
#>  9 All          0.838 Control         5.25        915.    6.09     1
#> 10 All          1.01  Experimental  124.         2421.  125.       1
#> 11 All          1.01  Experimental    0.577       453.    1.58     1
#> 12 All          1.11  Control        16.9        1120.   18.0      1
#> 13 All          1.22  Control         6.49        198.    7.71     1
#> 14 All          1.42  Control         5.88       1064.    7.30     1
#> 15 All          1.58  Experimental    8.58        642.   10.2      1
#> 16 All          1.77  Experimental    6.86        835.    8.63     1
#> 17 All          1.83  Experimental    6.67        478.    8.49     1
#> 18 All          1.83  Experimental   21.7          71.2  23.5      1
#> 19 All          1.84  Control         4.13       1218.    5.96     1
#> 20 All          2.03  Control         1.54       1083.    3.56     1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All         0.0245 Experimental    5.76        360.    5.79     1
#>  2 All         0.0844 Experimental   19.1        1306.   19.2      1
#>  3 All         0.0997 Experimental   28.1        1566.   28.2      1
#>  4 All         0.268  Control         3.66        934.    3.93     1
#>  5 All         0.285  Experimental   36.4         668.   36.7      1
#>  6 All         0.676  Control        23.6         372.   24.3      1
#>  7 All         0.698  Experimental   38.1         371.   38.8      1
#>  8 All         0.702  Experimental   62.9        1426.   63.6      1
#>  9 All         0.753  Experimental    0.405       152.    1.16     1
#> 10 All         0.883  Experimental    1.98       3493.    2.86     1
#> 11 All         1.03   Control         8.89        170.    9.92     1
#> 12 All         1.08   Experimental    6.07        129.    7.15     1
#> 13 All         1.11   Experimental    1.52       1531.    2.63     1
#> 14 All         1.37   Experimental    1.81       2633.    3.18     1
#> 15 All         1.53   Experimental   93.7        1649.   95.3      1
#> 16 All         1.53   Control         0.776       615.    2.31     1
#> 17 All         1.53   Experimental    2.57        216.    4.10     1
#> 18 All         1.63   Control         8.85        336.   10.5      1
#> 19 All         1.73   Experimental  135.          591.  137.       1
#> 20 All         2.08   Experimental   26.0          32.6  28.1      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High        0.0317 Control         22.5         41.7 22.6      1
#>  2 High        0.254  Experimental    21.2       3270.  21.5      1
#>  3 Low         0.403  Experimental    15.9         94.7 16.3      1
#>  4 Low         0.433  Experimental    10.7         95.9 11.1      1
#>  5 High        0.610  Experimental    61.5       1000.  62.1      1
#>  6 Low         0.628  Control         31.1        411.  31.7      1
#>  7 High        0.944  Control          1.33      1528.   2.27     1
#>  8 High        0.990  Experimental     1.13      4686.   2.12     1
#>  9 High        1.03   Control         11.6        461.  12.6      1
#> 10 High        1.07   Control          2.95       804.   4.02     1
#> 11 High        1.11   Experimental    27.7        458.  28.8      1
#> 12 High        1.14   Experimental    41.1        299.  42.3      1
#> 13 Low         1.16   Control         32.0        155.  33.1      1
#> 14 High        1.17   Control         10.3        854.  11.4      1
#> 15 High        1.26   Experimental    29.0        258.  30.3      1
#> 16 Low         1.44   Experimental    93.4       2319.  94.8      1
#> 17 High        1.68   Control         28.2        193.  29.9      1
#> 18 Low         1.80   Control         15.6        309.  17.4      1
#> 19 Low         1.82   Experimental    38.2       1006.  40.0      1
#> 20 Low         1.90   Control         22.1       1885.  24.0      1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low         0.0853 Experimental    24.0       1354.  24.1      1
#>  2 High        0.222  Experimental    18.7       2407.  18.9      1
#>  3 Low         0.232  Control         36.6        246.  36.8      1
#>  4 High        0.347  Experimental     9.85       979.  10.2      1
#>  5 High        0.465  Control         66.2        435.  66.7      1
#>  6 High        0.755  Control         11.7        237.  12.5      1
#>  7 High        0.813  Experimental     3.59       483.   4.40     1
#>  8 High        1.24   Control         35.5        969.  36.7      1
#>  9 Low         1.28   Experimental    36.6       2768.  37.9      1
#> 10 Low         1.39   Control          9.70       460.  11.1      1
#> 11 High        1.58   Control         26.1         75.8 27.7      1
#> 12 Low         1.82   Control         66.1        242.  68.0      1