Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment
    = c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.299 Experimental   46.2        2078.  46.5      1
#>  2 All          0.364 Control        13.1        1681.  13.5      1
#>  3 All          0.442 Control        12.7        1585.  13.2      1
#>  4 All          0.555 Experimental    2.99        510.   3.55     1
#>  5 All          0.631 Experimental   10.1         146.  10.8      1
#>  6 All          0.673 Control        27.7         171.  28.3      1
#>  7 All          0.746 Control        19.3         165.  20.1      1
#>  8 All          0.853 Experimental   13.7         281.  14.6      1
#>  9 All          1.01  Control        11.6          97.6 12.6      1
#> 10 All          1.12  Experimental   32.9        1173.  34.0      1
#> 11 All          1.12  Experimental    0.356      1494.   1.47     1
#> 12 All          1.44  Control         6.12        508.   7.56     1
#> 13 All          1.49  Experimental   11.2         194.  12.7      1
#> 14 All          1.63  Control        19.4          33.5 21.0      1
#> 15 All          1.66  Experimental    0.851       573.   2.51     1
#> 16 All          1.67  Control         6.51       1378.   8.18     1
#> 17 All          1.70  Control         0.627       515.   2.33     1
#> 18 All          1.88  Control        15.7         632.  17.5      1
#> 19 All          1.92  Experimental    1.92       1599.   3.84     1
#> 20 All          1.96  Experimental   14.7         278.  16.7      1
 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.327 Experimental    0.935       577.   1.26     1
#>  2 All          0.355 Control        12.1         466.  12.4      1
#>  3 All          0.388 Experimental   23.9         134.  24.2      1
#>  4 All          0.538 Experimental   21.6         258.  22.2      1
#>  5 All          0.937 Experimental    8.67        100.   9.61     1
#>  6 All          0.963 Experimental    0.615      2426.   1.58     1
#>  7 All          1.12  Control         2.90        235.   4.02     1
#>  8 All          1.24  Experimental   11.1         301.  12.3      1
#>  9 All          1.29  Experimental   43.2         625.  44.5      1
#> 10 All          1.55  Experimental   57.8          26.4 27.9      0
#> 11 All          1.58  Control         9.53       1189.  11.1      1
#> 12 All          1.65  Experimental   34.7         303.  36.4      1
#> 13 All          1.97  Experimental   19.1         301.  21.1      1
#> 14 All          2.10  Control         2.03        807.   4.13     1
#> 15 All          2.12  Experimental   93.5        2818.  95.6      1
#> 16 All          2.41  Experimental   87.1         738.  89.5      1
#> 17 All          2.45  Control         8.95         99.1 11.4      1
#> 18 All          2.48  Experimental    9.77        147.  12.3      1
#> 19 All          2.49  Experimental   39.0         199.  41.5      1
#> 20 All          2.61  Experimental   41.2        1517.  43.8      1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 High        0.0152 Control         4.68        526.    4.70      1
#>  2 High        0.0564 Experimental    0.722       297.    0.778     1
#>  3 High        0.108  Experimental    0.752       655.    0.860     1
#>  4 High        0.165  Control        17.6         530.   17.8       1
#>  5 Low         0.212  Experimental   38.0         394.   38.2       1
#>  6 Low         0.329  Control        11.7         806.   12.0       1
#>  7 High        0.650  Experimental   31.2        2200.   31.8       1
#>  8 Low         0.739  Experimental    1.30        497.    2.04      1
#>  9 High        0.888  Control         4.99        130.    5.87      1
#> 10 High        1.04   Control        37.3          89.2  38.3       1
#> 11 High        1.05   Experimental    0.242       621.    1.29      1
#> 12 High        1.32   Control         5.19        373.    6.51      1
#> 13 High        1.62   Experimental   62.3         619.   63.9       1
#> 14 High        1.71   Control        20.2         528.   21.9       1
#> 15 Low         1.93   Control        26.6         360.   28.6       1
#> 16 High        1.98   Experimental    2.07        325.    4.05      1
#> 17 High        2.01   Experimental    2.46       2965.    4.47      1
#> 18 Low         2.01   Experimental   41.2        1290.   43.2       1
#> 19 High        2.07   Control        14.5        4562.   16.5       1
#> 20 Low         2.15   Experimental  100.          549.  102.        1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 Low        0.00332 Experimental  206.          119.  119.        0
#>  2 High       0.0408  Control        14.3        1172.   14.4       1
#>  3 High       0.0445  Experimental   22.9          68.1  22.9       1
#>  4 High       0.145   Control        18.2         316.   18.3       1
#>  5 High       0.507   Experimental    0.292       489.    0.798     1
#>  6 High       0.527   Control        28.0        1516.   28.6       1
#>  7 High       0.565   Experimental    7.69        693.    8.25      1
#>  8 Low        0.937   Experimental    2.56       1618.    3.49      1
#>  9 Low        1.05    Control         5.91        625.    6.97      1
#> 10 High       1.44    Experimental   93.4         105.   94.9       1
#> 11 High       1.59    Control        35.4         668.   37.0       1
#> 12 High       1.65    Control        11.7        1996.   13.3       1