Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.141 Experimental  20.2          227.  20.4      1
#>  2 All          0.326 Experimental  18.9          274.  19.2      1
#>  3 All          0.376 Control       33.7         1138.  34.1      1
#>  4 All          0.561 Control        4.55         423.   5.11     1
#>  5 All          0.976 Experimental   1.96         217.   2.94     1
#>  6 All          0.996 Experimental  12.7          907.  13.7      1
#>  7 All          1.23  Control        2.68         483.   3.92     1
#>  8 All          1.27  Control        6.25         904.   7.52     1
#>  9 All          1.61  Control        2.14         535.   3.75     1
#> 10 All          1.62  Experimental   0.563       1042.   2.19     1
#> 11 All          1.83  Control        1.05         259.   2.88     1
#> 12 All          1.88  Experimental  50.0         1085.  51.9      1
#> 13 All          1.88  Experimental  64.6          679.  66.5      1
#> 14 All          1.97  Experimental  26.8          231.  28.8      1
#> 15 All          2.00  Control        0.367         71.4  2.37     1
#> 16 All          2.01  Control       10.4          331.  12.4      1
#> 17 All          2.05  Control        9.08         600.  11.1      1
#> 18 All          2.07  Experimental   0.0342      1828.   2.10     1
#> 19 All          2.27  Experimental  13.2          527.  15.5      1
#> 20 All          2.38  Control        7.33        1238.   9.72     1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All        0.00381 Experimental   17.4         212.   17.4      1
#>  2 All        0.159   Experimental    3.41       1364.    3.57     1
#>  3 All        0.350   Control         6.60        916.    6.95     1
#>  4 All        0.696   Experimental   96.6         424.   97.3      1
#>  5 All        1.00    Experimental   29.1        2283.   30.1      1
#>  6 All        1.01    Experimental    5.26       1396.    6.27     1
#>  7 All        1.03    Experimental   14.0         900.   15.1      1
#>  8 All        1.08    Control        14.0         274.   15.1      1
#>  9 All        1.10    Experimental    0.643       422.    1.74     1
#> 10 All        1.20    Control         3.49       1475.    4.69     1
#> 11 All        1.35    Experimental    4.85       2865.    6.19     1
#> 12 All        1.63    Experimental   42.8         720.   44.4      1
#> 13 All        1.76    Control        29.6         430.   31.3      1
#> 14 All        2.03    Experimental    8.89         41.3  10.9      1
#> 15 All        2.08    Experimental   14.5         182.   16.6      1
#> 16 All        2.26    Experimental    1.49        606.    3.75     1
#> 17 All        2.26    Experimental  149.         2009.  152.       1
#> 18 All        2.36    Experimental   10.8         456.   13.2      1
#> 19 All        2.38    Control         5.38        341.    7.76     1
#> 20 All        2.41    Experimental    9.20         10.7  11.6      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High        0.0571 Experimental    43.7        891.  43.7      1
#>  2 Low         0.101  Experimental    34.1        467.  34.2      1
#>  3 High        0.369  Experimental    21.8       1031.  22.2      1
#>  4 High        0.390  Control          7.80      1675.   8.19     1
#>  5 High        0.668  Control         45.9        214.  46.6      1
#>  6 High        0.670  Experimental     1.66       583.   2.33     1
#>  7 High        0.724  Control          2.01      6490.   2.73     1
#>  8 High        0.730  Control          2.01      1280.   2.74     1
#>  9 High        0.787  Experimental    50.5        798.  51.3      1
#> 10 High        1.06   Experimental    16.7        393.  17.8      1
#> 11 High        1.45   Experimental     3.78        90.1  5.24     1
#> 12 High        2.04   Control          3.95      2911.   5.99     1
#> 13 Low         2.06   Control          5.36       112.   7.42     1
#> 14 High        2.16   Control          8.60       402.  10.8      1
#> 15 High        2.23   Control         15.8        991.  18.0      1
#> 16 High        2.26   Experimental    12.2        695.  14.4      1
#> 17 High        2.30   Experimental     5.20       442.   7.51     1
#> 18 Low         2.31   Control         42.5        829.  44.8      1
#> 19 High        2.42   Control         13.0       1451.  15.4      1
#> 20 Low         2.50   Experimental    12.6       3217.  15.1      1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 Low         0.0188 Experimental   109.         4910. 109.       1
#>  2 Low         0.150  Control          5.72        183.   5.87     1
#>  3 High        0.433  Control         20.0        4100.  20.4      1
#>  4 Low         0.811  Control         32.9        4644.  33.7      1
#>  5 High        0.818  Experimental    20.0        1708.  20.8      1
#>  6 High        0.897  Control         14.2         522.  15.1      1
#>  7 Low         1.09   Experimental    22.2         202.  23.3      1
#>  8 High        1.19   Experimental    52.6        1348.  53.8      1
#>  9 Low         1.20   Experimental    36.7        1443.  37.9      1
#> 10 High        1.29   Control         20.0         451.  21.3      1
#> 11 Low         1.56   Control         39.8        1970.  41.4      1
#> 12 High        1.77   Experimental    20.8        3560.  22.6      1