Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.100 Control          8.20      1156.    8.30     1
#>  2 All          0.168 Experimental   103.         170.  103.       1
#>  3 All          0.398 Control         18.1       1230.   18.5      1
#>  4 All          0.412 Experimental    20.3         58.0  20.7      1
#>  5 All          0.481 Control         18.1        291.   18.5      1
#>  6 All          0.519 Experimental     2.27      4772.    2.79     1
#>  7 All          0.629 Experimental    26.9        493.   27.5      1
#>  8 All          0.689 Control          5.16        78.3   5.85     1
#>  9 All          0.746 Control         54.1       4068.   54.9      1
#> 10 All          0.897 Experimental    24.5         30.5  25.4      1
#> 11 All          1.07  Experimental     2.67       569.    3.74     1
#> 12 All          1.15  Control          9.43       171.   10.6      1
#> 13 All          1.43  Control          4.67       809.    6.10     1
#> 14 All          1.69  Experimental    10.5        909.   12.2      1
#> 15 All          1.70  Control         19.7        239.   21.4      1
#> 16 All          1.74  Experimental     9.90      2481.   11.6      1
#> 17 All          1.77  Experimental     4.52      1247.    6.29     1
#> 18 All          1.85  Control         10.3        227.   12.2      1
#> 19 All          2.06  Control          1.62      1140.    3.68     1
#> 20 All          2.29  Experimental    11.8        165.   14.1      1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.110 Experimental  19.4         1697.  19.5       1
#>  2 All          0.190 Experimental   0.0225       820.   0.213     1
#>  3 All          0.355 Control        3.52         326.   3.88      1
#>  4 All          0.431 Experimental  28.2          162.  28.6       1
#>  5 All          0.548 Experimental   9.32         948.   9.87      1
#>  6 All          0.660 Experimental   4.48         973.   5.14      1
#>  7 All          0.695 Experimental   5.19         763.   5.89      1
#>  8 All          0.740 Control       10.9         1794.  11.6       1
#>  9 All          0.903 Control        8.22        1363.   9.13      1
#> 10 All          1.17  Experimental   0.899       1654.   2.07      1
#> 11 All          1.19  Experimental   0.323        226.   1.51      1
#> 12 All          1.19  Experimental  22.7          900.  23.9       1
#> 13 All          1.28  Experimental  48.2         1893.  49.5       1
#> 14 All          1.41  Experimental   8.05          63.8  9.46      1
#> 15 All          1.57  Experimental  42.7          124.  44.3       1
#> 16 All          1.82  Control       24.4         1881.  26.2       1
#> 17 All          1.94  Experimental   2.63          11.4  4.57      1
#> 18 All          2.18  Control       15.6         1805.  17.8       1
#> 19 All          2.24  Experimental  20.3         1008.  22.6       1
#> 20 All          2.56  Experimental   5.80        1673.   8.36      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High        0.0364 Experimental   52.4        1032.  52.5       1
#>  2 High        0.0667 Control        11.2          95.7 11.2       1
#>  3 High        0.255  Experimental   79.0         642.  79.3       1
#>  4 High        0.367  Control         0.552        90.6  0.918     1
#>  5 Low         0.630  Control        25.1        1424.  25.7       1
#>  6 High        0.751  Experimental   20.9         448.  21.7       1
#>  7 Low         0.791  Control        10.5         478.  11.3       1
#>  8 Low         0.849  Experimental   10.3        1101.  11.1       1
#>  9 High        0.867  Experimental   23.3        1227.  24.2       1
#> 10 High        0.905  Control        14.5        2168.  15.4       1
#> 11 High        1.02   Control         0.120       347.   1.14      1
#> 12 Low         1.05   Experimental   11.2         683.  12.3       1
#> 13 High        1.09   Experimental   28.8         781.  29.9       1
#> 14 High        1.10   Control        13.4          85.0 14.5       1
#> 15 Low         1.17   Experimental   45.1          47.8 46.3       1
#> 16 High        1.18   Experimental   53.4         877.  54.6       1
#> 17 High        1.19   Control        26.6          95.3 27.8       1
#> 18 Low         1.20   Experimental    7.49        108.   8.70      1
#> 19 High        1.22   Experimental    1.12       1218.   2.34      1
#> 20 Low         1.36   Control        95.1        2425.  96.4       1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.139 Experimental    50.1        681.  50.2      1
#>  2 Low          0.515 Control         58.6         55.0 55.5      0
#>  3 High         0.535 Experimental     3.34       115.   3.88     1
#>  4 Low          0.537 Experimental    29.2       1749.  29.8      1
#>  5 High         0.587 Control          3.78       326.   4.37     1
#>  6 High         0.669 Control          4.18       549.   4.85     1
#>  7 High         0.820 Experimental    33.2       1143.  34.0      1
#>  8 Low          1.48  Experimental    15.5        289.  17.0      1
#>  9 High         1.53  Control         32.9       1525.  34.4      1
#> 10 Low          1.58  Control         83.1        775.  84.7      1
#> 11 High         1.73  Control         24.2        603.  26.0      1
#> 12 Low          1.80  Control          2.26       172.   4.07     1