Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0512 Control         13.0       224.   13.1      1
#>  2 All         0.0551 Experimental    18.0        49.4  18.1      1
#>  3 All         0.123  Control         17.8       411.   17.9      1
#>  4 All         0.209  Experimental    63.8      1377.   64.0      1
#>  5 All         0.214  Experimental     6.51      962.    6.73     1
#>  6 All         0.607  Control          2.90      687.    3.51     1
#>  7 All         0.744  Control         18.9      1276.   19.7      1
#>  8 All         0.872  Experimental     3.66      678.    4.53     1
#>  9 All         0.928  Experimental    28.4      1843.   29.3      1
#> 10 All         0.948  Control         14.3       158.   15.2      1
#> 11 All         0.977  Experimental    13.2       434.   14.2      1
#> 12 All         1.17   Control          4.57     1318.    5.74     1
#> 13 All         1.23   Experimental   125.          3.23  4.46     0
#> 14 All         1.24   Control         25.2      3232.   26.5      1
#> 15 All         1.27   Control          3.47      360.    4.74     1
#> 16 All         1.46   Experimental    29.5      4831.   30.9      1
#> 17 All         1.53   Experimental    19.9      1031.   21.4      1
#> 18 All         1.53   Control         30.3       251.   31.8      1
#> 19 All         1.64   Experimental    13.9       357.   15.5      1
#> 20 All         1.95   Control          9.39       88.9  11.3      1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All         0.0352 Experimental   25.3         734.   25.3      1
#>  2 All         0.129  Experimental    8.28        494.    8.41     1
#>  3 All         0.336  Control         7.28        267.    7.61     1
#>  4 All         0.423  Experimental   26.2        3947.   26.6      1
#>  5 All         0.431  Experimental   13.6         307.   14.1      1
#>  6 All         0.562  Experimental   70.2        1308.   70.8      1
#>  7 All         0.593  Experimental   12.7         211.   13.3      1
#>  8 All         0.598  Control        39.9          44.4  40.5      1
#>  9 All         0.611  Control        29.5        1047.   30.2      1
#> 10 All         0.635  Experimental   31.1         819.   31.8      1
#> 11 All         0.851  Experimental   22.7        1231.   23.5      1
#> 12 All         0.940  Experimental    0.794       144.    1.73     1
#> 13 All         0.982  Experimental   14.7        1012.   15.7      1
#> 14 All         1.12   Experimental    1.83       1795.    2.95     1
#> 15 All         1.31   Experimental   46.4        1234.   47.7      1
#> 16 All         1.33   Control         7.87       1061.    9.20     1
#> 17 All         1.58   Experimental   21.9         282.   23.4      1
#> 18 All         1.62   Experimental  114.          854.  115.       1
#> 19 All         1.80   Control        28.1        1708.   29.9      1
#> 20 All         1.84   Experimental   15.0         545.   16.8      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.304 Experimental   21.2        765.   21.5      1
#>  2 High         0.424 Experimental   40.0        674.   40.5      1
#>  3 High         0.449 Control        15.5       2079.   16.0      1
#>  4 High         0.475 Control         4.59       854.    5.06     1
#>  5 High         0.700 Control        30.0       1652.   30.7      1
#>  6 Low          0.827 Control        45.9        394.   46.7      1
#>  7 High         0.841 Experimental   23.7       1961.   24.5      1
#>  8 Low          0.873 Experimental    0.630     5337.    1.50     1
#>  9 High         0.885 Experimental   11.9        216.   12.8      1
#> 10 High         1.02  Control         3.11       313.    4.13     1
#> 11 High         1.06  Control         5.42      1616.    6.48     1
#> 12 High         1.07  Experimental   36.7       1485.   37.7      1
#> 13 High         1.09  Experimental   59.5         69.3  60.5      1
#> 14 Low          1.13  Experimental   14.7       2617.   15.8      1
#> 15 High         1.23  Control         2.20       315.    3.43     1
#> 16 High         1.31  Experimental    6.35        34.5   7.66     1
#> 17 High         1.66  Control        10.2        545.   11.9      1
#> 18 High         1.79  Control         8.89         6.54  8.33     0
#> 19 Low          1.97  Control        23.7       2363.   25.6      1
#> 20 Low          2.30  Control         5.35       723.    7.65     1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High         0.107 Experimental    28.7         47.4  28.8      1
#>  2 High         0.386 Experimental     2.58      1922.    2.97     1
#>  3 Low          0.482 Experimental    65.0        268.   65.5      1
#>  4 Low          0.509 Control         86.7        239.   87.2      1
#>  5 Low          0.521 Experimental    16.7       3256.   17.2      1
#>  6 High         0.619 Control          7.45       830.    8.07     1
#>  7 Low          0.881 Control          6.09      3053.    6.97     1
#>  8 Low          0.940 Control         75.3        316.   76.3      1
#>  9 High         1.13  Control          1.87       402.    3.00     1
#> 10 High         1.18  Control         10.4       1058.   11.6      1
#> 11 High         1.29  Experimental     2.81       113.    4.10     1
#> 12 Low          1.36  Experimental   110.         989.  112.       1