Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.135 Experimental    6.28     1929.     6.42      1
#>  2 All          0.139 Control         5.32      135.     5.46      1
#>  3 All          0.148 Control         3.02      103.     3.17      1
#>  4 All          0.159 Experimental    6.13      519.     6.29      1
#>  5 All          0.372 Experimental   10.5         0.358  0.730     0
#>  6 All          0.433 Control        28.0        52.4   28.5       1
#>  7 All          0.528 Control        21.4       390.    21.9       1
#>  8 All          0.744 Experimental   17.9      2494.    18.6       1
#>  9 All          0.838 Control        11.9      5003.    12.8       1
#> 10 All          0.885 Experimental   33.2       619.    34.1       1
#> 11 All          0.919 Control         5.24     2670.     6.16      1
#> 12 All          0.950 Experimental    1.84      745.     2.79      1
#> 13 All          1.25  Experimental   13.5        32.0   14.8       1
#> 14 All          1.30  Control        35.8       450.    37.1       1
#> 15 All          1.45  Control         3.48     1633.     4.93      1
#> 16 All          1.48  Experimental    0.364    1051.     1.84      1
#> 17 All          1.51  Experimental    3.57      539.     5.08      1
#> 18 All          1.56  Control        12.6      1079.    14.1       1
#> 19 All          1.80  Experimental    5.69      239.     7.49      1
#> 20 All          1.99  Control        10.9      1881.    12.9       1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0443 Experimental   24.9         567.  24.9      1
#>  2 All         0.0739 Control        14.0         126.  14.1      1
#>  3 All         0.166  Experimental   41.1         593.  41.3      1
#>  4 All         0.244  Experimental    1.23         18.9  1.47     1
#>  5 All         0.408  Experimental    4.62       1594.   5.03     1
#>  6 All         0.420  Experimental   15.7         757.  16.1      1
#>  7 All         0.552  Control         3.77       1322.   4.32     1
#>  8 All         0.682  Experimental    7.60       1602.   8.28     1
#>  9 All         0.784  Experimental    4.67       2437.   5.46     1
#> 10 All         0.842  Experimental   10.8         644.  11.6      1
#> 11 All         0.882  Control        27.9         687.  28.8      1
#> 12 All         1.30   Experimental   15.9        2440.  17.2      1
#> 13 All         1.41   Experimental   11.4         534.  12.8      1
#> 14 All         1.48   Experimental    2.86         97.9  4.34     1
#> 15 All         1.89   Experimental    0.902        95.1  2.79     1
#> 16 All         2.06   Control        10.6        1902.  12.7      1
#> 17 All         2.27   Experimental   45.1        1114.  47.4      1
#> 18 All         2.45   Experimental   21.7         762.  24.2      1
#> 19 All         2.71   Control         2.11       4455.   4.82     1
#> 20 All         3.03   Experimental   43.2        1684.  46.2      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.187 Experimental   19.8        1345.  20.0      1
#>  2 Low          0.188 Experimental   12.4         443.  12.5      1
#>  3 High         0.296 Control        37.2          77.6 37.5      1
#>  4 Low          0.339 Control        43.8         270.  44.2      1
#>  5 High         0.576 Experimental    9.83        606.  10.4      1
#>  6 High         0.618 Control         4.57        887.   5.18     1
#>  7 High         0.666 Control         2.17       1325.   2.83     1
#>  8 Low          0.876 Control        15.2         320.  16.1      1
#>  9 Low          0.947 Experimental   95.6         993.  96.6      1
#> 10 High         1.06  Experimental   54.3         250.  55.3      1
#> 11 High         1.09  Control         6.74        992.   7.83     1
#> 12 High         1.10  Experimental    0.821        51.0  1.92     1
#> 13 High         1.26  Experimental   35.9          59.5 37.1      1
#> 14 High         1.28  Control        14.7        2179.  15.9      1
#> 15 Low          1.31  Experimental    4.37        329.   5.68     1
#> 16 High         1.34  Control         7.00       1965.   8.33     1
#> 17 High         1.37  Experimental   30.0        2400.  31.4      1
#> 18 High         1.40  Experimental    0.489        30.8  1.89     1
#> 19 Low          1.41  Experimental   56.5        1416.  57.9      1
#> 20 High         1.53  Experimental    3.04        146.   4.57     1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High        0.0265 Experimental     1.26        162.  1.29     1
#>  2 High        0.0707 Experimental     1.86       2488.  1.93     1
#>  3 High        0.247  Control         49.6        1642. 49.8      1
#>  4 Low         0.440  Experimental    18.5        2001. 18.9      1
#>  5 High        0.637  Control          9.78        767. 10.4      1
#>  6 High        0.696  Experimental    22.0         457. 22.7      1
#>  7 Low         0.734  Control         48.6        1007. 49.3      1
#>  8 High        0.813  Control          2.00        527.  2.81     1
#>  9 High        0.957  Experimental    54.3        1096. 55.3      1
#> 10 Low         0.972  Control         30.3         383. 31.3      1
#> 11 High        0.998  Control         11.3         226. 12.3      1
#> 12 High        1.07   Experimental    41.2         385. 42.3      1