Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment
    = c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0311 Control        24.6         807.  24.6      1
#>  2 All         0.0346 Experimental   69.2         131.  69.2      1
#>  3 All         0.231  Experimental   31.1        1122.  31.3      1
#>  4 All         0.504  Control        10.1        1069.  10.6      1
#>  5 All         0.554  Control        11.4        1182.  11.9      1
#>  6 All         0.746  Experimental    0.921       881.   1.67     1
#>  7 All         0.763  Control        14.3         480.  15.1      1
#>  8 All         1.26   Experimental   47.6         798.  48.8      1
#>  9 All         1.58   Control        15.1        3502.  16.7      1
#> 10 All         1.91   Control         3.28        868.   5.19     1
#> 11 All         1.97   Experimental    6.54        381.   8.52     1
#> 12 All         1.98   Experimental   12.0         471.  14.0      1
#> 13 All         2.01   Experimental   57.4        2373.  59.4      1
#> 14 All         2.07   Control        36.1         659.  38.2      1
#> 15 All         2.46   Control        14.2        4830.  16.6      1
#> 16 All         2.50   Experimental    4.62         98.7  7.11     1
#> 17 All         2.85   Control        27.7         238.  30.5      1
#> 18 All         2.92   Experimental   12.0         607.  14.9      1
#> 19 All         3.01   Experimental   54.5         469.  57.5      1
#> 20 All         3.02   Control         7.68        536.  10.7      1
 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.108 Experimental    0.393       1748.  0.502     1
#>  2 All          0.136 Experimental   44.5         1341. 44.7       1
#>  3 All          0.198 Control         7.70        1091.  7.90      1
#>  4 All          0.389 Experimental    3.02         685.  3.41      1
#>  5 All          0.462 Control         1.67         597.  2.13      1
#>  6 All          0.563 Experimental   49.9          571. 50.5       1
#>  7 All          0.565 Experimental   11.9          478. 12.5       1
#>  8 All          0.879 Experimental   92.3         3118. 93.2       1
#>  9 All          0.917 Experimental   53.5         2198. 54.4       1
#> 10 All          1.03  Experimental    2.50        1356.  3.53      1
#> 11 All          1.08  Control         1.65         432.  2.73      1
#> 12 All          1.11  Experimental    1.59         626.  2.70      1
#> 13 All          1.14  Experimental   17.2          347. 18.3       1
#> 14 All          1.22  Control         1.84        2049.  3.06      1
#> 15 All          1.32  Experimental    7.95        1159.  9.27      1
#> 16 All          1.36  Experimental   16.4          273. 17.8       1
#> 17 All          1.51  Experimental    1.39         794.  2.91      1
#> 18 All          1.51  Control         6.12        2664.  7.64      1
#> 19 All          1.55  Experimental   35.6          601. 37.1       1
#> 20 All          1.61  Experimental   57.6         1504. 59.2       1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High        0.0239 Control         1.86         76.0  1.88     1
#>  2 High        0.0467 Experimental   29.0        1918.  29.1      1
#>  3 Low         0.150  Experimental   40.4         932.  40.5      1
#>  4 High        0.172  Experimental    7.05        660.   7.22     1
#>  5 Low         0.247  Experimental   88.8         307.  89.1      1
#>  6 High        0.289  Control        23.4         621.  23.7      1
#>  7 High        0.507  Control        22.0        3187.  22.5      1
#>  8 High        0.591  Control        25.3         870.  25.9      1
#>  9 High        0.610  Experimental    3.57         73.8  4.18     1
#> 10 Low         0.786  Control        20.5        1165.  21.2      1
#> 11 High        0.840  Experimental    0.395       161.   1.24     1
#> 12 High        0.870  Control        25.4        1415.  26.3      1
#> 13 High        0.928  Control         4.27       1096.   5.20     1
#> 14 High        1.03   Experimental   22.2        1140.  23.2      1
#> 15 High        1.05   Experimental    8.02        316.   9.07     1
#> 16 High        1.30   Control        48.4        2474.  49.7      1
#> 17 High        1.35   Experimental   13.8        1010.  15.2      1
#> 18 High        1.56   Control        18.9         129.  20.5      1
#> 19 Low         1.65   Control        31.6        2593.  33.3      1
#> 20 Low         1.73   Experimental   57.5         451.  59.2      1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 Low          0.184 Control       121.         1798.  121.       1
#>  2 Low          0.566 Experimental    2.17       1886.    2.74     1
#>  3 High         0.604 Control        31.1         239.   31.7      1
#>  4 Low          0.651 Experimental   21.4         459.   22.0      1
#>  5 High         0.715 Control         6.10       1969.    6.82     1
#>  6 Low          0.761 Control        24.7          70.1  25.4      1
#>  7 High         1.10  Experimental   85.9        2729.   87.0      1
#>  8 High         1.15  Experimental   30.2         497.   31.4      1
#>  9 High         1.15  Control        36.3         162.   37.5      1
#> 10 Low          1.18  Control        21.9        1146.   23.1      1
#> 11 Low          1.23  Experimental    1.35        155.    2.58     1
#> 12 Low          1.24  Control         0.354      2291.    1.60     1