Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment
    = c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in `Stratum`, probability (incidence) of each stratum in `p`

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0663 Experimental   40.7         313.  40.8      1
#>  2 All         0.187  Control        25.1         408.  25.3      1
#>  3 All         0.333  Control         9.31        618.   9.64     1
#>  4 All         0.601  Experimental    1.70        914.   2.30     1
#>  5 All         0.618  Control         8.50         61.9  9.12     1
#>  6 All         0.648  Experimental   69.4         133.  70.1      1
#>  7 All         0.658  Experimental    6.53       2379.   7.19     1
#>  8 All         0.711  Control        13.6        1691.  14.3      1
#>  9 All         0.821  Control         2.59        400.   3.42     1
#> 10 All         0.912  Experimental    0.338       469.   1.25     1
#> 11 All         0.928  Control        26.8          85.5 27.8      1
#> 12 All         0.952  Experimental   15.7         787.  16.6      1
#> 13 All         1.00   Experimental   17.3         498.  18.3      1
#> 14 All         1.04   Control         8.92       1677.   9.95     1
#> 15 All         1.07   Experimental   30.4         490.  31.5      1
#> 16 All         1.09   Control        11.5         166.  12.6      1
#> 17 All         1.13   Experimental    3.55        898.   4.68     1
#> 18 All         1.16   Control         7.20        161.   8.36     1
#> 19 All         1.20   Experimental   25.0         774.  26.2      1
#> 20 All         1.24   Control         7.76        995.   9.00     1
 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.200 Experimental     7.19      4296.   7.39     1
#>  2 All          0.254 Experimental    13.2       3531.  13.4      1
#>  3 All          0.299 Control         27.9        515.  28.2      1
#>  4 All          0.424 Experimental    24.5       1145.  24.9      1
#>  5 All          0.539 Experimental    98.3        288.  98.8      1
#>  6 All          0.632 Control          2.83       263.   3.46     1
#>  7 All          0.642 Experimental     8.83       435.   9.47     1
#>  8 All          0.903 Experimental    44.2       1085.  45.1      1
#>  9 All          0.928 Experimental     4.77       956.   5.70     1
#> 10 All          0.929 Experimental     1.73       556.   2.66     1
#> 11 All          1.18  Experimental     4.00      1656.   5.18     1
#> 12 All          1.26  Control         15.5        138.  16.8      1
#> 13 All          1.38  Experimental     1.64       316.   3.02     1
#> 14 All          1.48  Experimental     2.46        89.7  3.93     1
#> 15 All          1.52  Experimental     6.66      2608.   8.18     1
#> 16 All          1.58  Control         73.2       1177.  74.8      1
#> 17 All          1.72  Experimental    27.0       1130.  28.8      1
#> 18 All          1.72  Experimental    19.8       2174.  21.6      1
#> 19 All          1.99  Control         19.1        303.  21.1      1
#> 20 All          2.10  Experimental    14.2        112.  16.3      1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.236 Control        1.84        164.    2.07     1
#>  2 Low          0.939 Experimental  36.7         954.   37.7      1
#>  3 Low          1.28  Control        8.03       1372.    9.31     1
#>  4 High         1.41  Control       12.8          86.5  14.3      1
#>  5 Low          1.44  Control        8.77       1486.   10.2      1
#>  6 High         1.56  Experimental   0.153       645.    1.71     1
#>  7 High         1.87  Experimental  33.7           5.52  7.39     0
#>  8 High         1.99  Control        8.66        521.   10.7      1
#>  9 High         2.03  Control       14.3         339.   16.3      1
#> 10 Low          2.08  Experimental  13.8         488.   15.9      1
#> 11 Low          2.11  Control       10.7         103.   12.8      1
#> 12 Low          2.15  Experimental   0.460       994.    2.61     1
#> 13 High         2.21  Experimental  11.5        1730.   13.7      1
#> 14 High         2.26  Experimental   0.544       334.    2.80     1
#> 15 High         2.26  Experimental   5.62        596.    7.88     1
#> 16 Low          2.42  Control       32.8        1221.   35.2      1
#> 17 High         2.42  Experimental   8.66        507.   11.1      1
#> 18 High         2.78  Control       43.7         459.   46.4      1
#> 19 High         3.04  Control        0.0724      577.    3.11     1
#> 20 High         3.10  Control        7.85       1005.   11.0      1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.116 Control         1.98         364.  2.10     1
#>  2 Low          0.135 Control         7.24         304.  7.38     1
#>  3 High         0.136 Control        14.5         2475. 14.6      1
#>  4 High         0.155 Experimental    9.74        2957.  9.90     1
#>  5 High         0.162 Experimental    6.58        1142.  6.74     1
#>  6 High         0.312 Experimental    1.60         736.  1.91     1
#>  7 High         0.426 Experimental   11.4          237. 11.8      1
#>  8 Low          0.624 Experimental    0.851       1486.  1.47     1
#>  9 High         0.744 Control        18.6         5482. 19.3      1
#> 10 Low          0.837 Control         4.38         722.  5.22     1
#> 11 Low          0.871 Experimental   50.1          707. 51.0      1
#> 12 High         0.937 Control        13.4         1812. 14.3      1