Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment
    = c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 All         0.0399 Experimental    0.382      1750.    0.422     1
#>  2 All         0.287  Experimental    0.964       200.    1.25      1
#>  3 All         0.370  Control         9.32       1708.    9.69      1
#>  4 All         0.473  Control         1.69         60.8   2.17      1
#>  5 All         0.787  Control         0.127       488.    0.915     1
#>  6 All         0.793  Experimental    2.31        591.    3.10      1
#>  7 All         0.805  Experimental  130.          401.  131.        1
#>  8 All         0.862  Control         9.65       1936.   10.5       1
#>  9 All         0.908  Control         2.74        269.    3.65      1
#> 10 All         0.929  Experimental   17.3          78.9  18.2       1
#> 11 All         1.05   Control         3.56        692.    4.61      1
#> 12 All         1.06   Experimental   22.4         108.   23.5       1
#> 13 All         1.17   Control        32.6         633.   33.7       1
#> 14 All         1.26   Experimental    0.538       198.    1.80      1
#> 15 All         1.34   Control         7.54        500.    8.88      1
#> 16 All         1.38   Experimental   32.0         387.   33.4       1
#> 17 All         1.47   Experimental    9.43       2457.   10.9       1
#> 18 All         1.54   Experimental    9.32         44.5  10.9       1
#> 19 All         1.66   Control         7.74        786.    9.40      1
#> 20 All         1.66   Control         8.34        269.   10.0       1
 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 All          0.121 Experimental  126.           451. 126.        1
#>  2 All          0.202 Experimental    6.46        1632.   6.66      1
#>  3 All          0.251 Experimental    0.365       3238.   0.617     1
#>  4 All          0.765 Control        31.4          428.  32.2       1
#>  5 All          0.815 Experimental   26.5         1227.  27.4       1
#>  6 All          0.916 Experimental   28.0          797.  28.9       1
#>  7 All          0.950 Experimental    5.94        1000.   6.89      1
#>  8 All          1.13  Control         9.80         749.  10.9       1
#>  9 All          1.24  Experimental    2.04        3427.   3.28      1
#> 10 All          1.49  Experimental   38.5          543.  40.0       1
#> 11 All          1.65  Control         3.01        1272.   4.66      1
#> 12 All          1.67  Experimental   46.8         1155.  48.4       1
#> 13 All          1.83  Experimental   13.3          526.  15.1       1
#> 14 All          1.94  Experimental    2.39        2083.   4.33      1
#> 15 All          2.02  Control         4.50        1015.   6.51      1
#> 16 All          2.23  Experimental    3.00        1503.   5.23      1
#> 17 All          2.30  Experimental   11.8         1369.  14.1       1
#> 18 All          2.49  Experimental    1.96         981.   4.45      1
#> 19 All          2.56  Control        13.7          714.  16.3       1
#> 20 All          2.62  Experimental   25.1          179.  27.7       1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 High        0.0192 Control       5.32          832.   5.34       1
#>  2 High        0.0655 Experimental  0.00131      1031.   0.0668     1
#>  3 Low         0.370  Experimental  7.55         2312.   7.92       1
#>  4 High        0.474  Control       7.22         1224.   7.69       1
#>  5 Low         0.491  Control      21.7           135.  22.1        1
#>  6 High        0.610  Experimental 10.3           420.  10.9        1
#>  7 Low         0.740  Experimental 32.2            12.9 13.7        0
#>  8 Low         1.02   Control      18.7          1483.  19.8        1
#>  9 Low         1.32   Experimental  7.56         1278.   8.88       1
#> 10 High        1.34   Experimental 25.1          2905.  26.4        1
#> 11 High        1.37   Control       1.77          199.   3.14       1
#> 12 High        1.67   Experimental 14.9           531.  16.5        1
#> 13 High        1.95   Control       2.71           64.9  4.66       1
#> 14 High        2.05   Control       3.56          704.   5.61       1
#> 15 Low         2.17   Experimental 13.9           659.  16.0        1
#> 16 High        2.20   Experimental 27.1          1068.  29.3        1
#> 17 Low         2.38   Control      12.3           163.  14.7        1
#> 18 High        2.47   Experimental 91.5           358.  94.0        1
#> 19 High        2.49   Control       1.26          219.   3.75       1
#> 20 High        2.69   Control       7.03         6876.   9.72       1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 Low        0.00192 Experimental    55.0       1590.   55.0      1
#>  2 High       0.0321  Experimental    37.0         16.9  16.9      0
#>  3 Low        0.204   Experimental    16.8       4340.   17.0      1
#>  4 Low        0.345   Control         12.6        615.   12.9      1
#>  5 High       0.415   Control          2.03      1905.    2.44     1
#>  6 Low        0.419   Control         31.6       1253.   32.0      1
#>  7 High       0.468   Experimental    16.8        485.   17.2      1
#>  8 Low        0.627   Experimental   231.        3402.  232.       1
#>  9 High       0.756   Control         21.4       1794.   22.2      1
#> 10 High       0.839   Experimental    34.4       1038.   35.3      1
#> 11 Low        0.990   Experimental    36.0        577.   37.0      1
#> 12 High       1.05    Control          4.97       154.    6.02     1