Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All         0.0917 Experimental   26.7          292. 26.8       1
#>  2 All         0.113  Experimental   35.6         2659. 35.7       1
#>  3 All         0.144  Control         4.42        3461.  4.56      1
#>  4 All         0.314  Control        29.5         1093. 29.9       1
#>  5 All         0.389  Experimental    0.451        101.  0.840     1
#>  6 All         0.740  Control        16.8         1267. 17.6       1
#>  7 All         0.852  Experimental   66.6         1476. 67.4       1
#>  8 All         1.22   Control         7.54         137.  8.76      1
#>  9 All         1.33   Control         1.87         601.  3.20      1
#> 10 All         1.38   Experimental   20.0          848. 21.3       1
#> 11 All         1.55   Control         5.05        1628.  6.61      1
#> 12 All         1.73   Experimental   31.8          360. 33.5       1
#> 13 All         1.84   Control        11.1         1072. 12.9       1
#> 14 All         2.00   Experimental   42.1          327. 44.1       1
#> 15 All         2.01   Control        17.7          246. 19.7       1
#> 16 All         2.23   Experimental   51.1          413. 53.3       1
#> 17 All         2.29   Experimental   76.7          660. 79.0       1
#> 18 All         2.52   Control         1.09         716.  3.62      1
#> 19 All         2.54   Control        14.3          353. 16.8       1
#> 20 All         2.73   Experimental    1.11         178.  3.85      1
 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.103 Control         5.12        168.    5.22     1
#>  2 All          0.281 Experimental    5.91        419.    6.19     1
#>  3 All          0.326 Experimental    1.41        722.    1.73     1
#>  4 All          0.380 Experimental  113.          247.  114.       1
#>  5 All          0.442 Experimental   24.0         368.   24.4      1
#>  6 All          0.443 Control         1.71        371.    2.15     1
#>  7 All          0.445 Experimental   10.3         442.   10.8      1
#>  8 All          0.491 Experimental    4.80        159.    5.29     1
#>  9 All          0.714 Experimental    6.01       1746.    6.72     1
#> 10 All          0.778 Experimental   35.7         293.   36.5      1
#> 11 All          0.907 Control        29.6          30.0  30.5      1
#> 12 All          0.949 Experimental   45.0        4263.   46.0      1
#> 13 All          1.10  Control        13.6         532.   14.7      1
#> 14 All          1.21  Experimental   17.8         418.   19.0      1
#> 15 All          1.35  Experimental    1.96         51.5   3.31     1
#> 16 All          1.49  Experimental    0.520       179.    2.01     1
#> 17 All          1.54  Experimental   19.8         368.   21.3      1
#> 18 All          1.58  Experimental   48.7        6033.   50.2      1
#> 19 All          1.69  Control         4.37        173.    6.05     1
#> 20 All          1.74  Experimental   23.5        2087.   25.3      1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High        0.0330 Experimental    10.7       2147.  10.7      1
#>  2 High        0.170  Control         34.5       2363.  34.6      1
#>  3 High        0.186  Experimental     1.49       620.   1.68     1
#>  4 High        0.365  Control          5.86       399.   6.22     1
#>  5 High        0.406  Control         22.7       1180.  23.1      1
#>  6 High        0.537  Experimental    29.5       1474.  30.1      1
#>  7 High        0.616  Control         39.2        365.  39.9      1
#>  8 Low         0.618  Experimental    10.3         79.8 10.9      1
#>  9 High        0.670  Experimental     9.10       170.   9.77     1
#> 10 High        0.687  Control          6.30      1173.   6.99     1
#> 11 High        1.06   Experimental    40.1        536.  41.1      1
#> 12 High        1.08   Experimental    34.8        491.  35.9      1
#> 13 Low         1.18   Control          1.07      1181.   2.24     1
#> 14 High        1.22   Control         26.6        252.  27.8      1
#> 15 High        1.30   Control          9.08        80.4 10.4      1
#> 16 High        1.42   Experimental    20.9        230.  22.3      1
#> 17 High        1.55   Control          2.17       911.   3.72     1
#> 18 High        1.59   Experimental     5.97       709.   7.56     1
#> 19 High        1.62   Control         22.6        540.  24.2      1
#> 20 High        1.79   Experimental     9.09      2104.  10.9      1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High        0.0361 Control         25.0       2590.  25.1      1
#>  2 Low         0.0647 Experimental     6.20       325.   6.27     1
#>  3 Low         0.161  Control         13.7         91.5 13.8      1
#>  4 High        0.199  Experimental    53.5        142.  53.7      1
#>  5 High        0.224  Experimental    76.6        887.  76.8      1
#>  6 High        0.237  Control         25.7         19.0 19.3      0
#>  7 High        0.278  Experimental     2.80       170.   3.08     1
#>  8 Low         0.298  Control          9.81       644.  10.1      1
#>  9 High        0.573  Control          9.04       478.   9.61     1
#> 10 Low         0.831  Experimental    62.8       3865.  63.6      1
#> 11 Low         0.854  Experimental    72.7        419.  73.5      1
#> 12 High        0.957  Control         22.1         16.1 17.1      0