Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment
    = c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0708 Control        12.4         397.  12.5      1
#>  2 All         0.281  Experimental   22.8         188.  23.1      1
#>  3 All         0.291  Experimental   24.4         961.  24.7      1
#>  4 All         0.499  Control         6.44        864.   6.94     1
#>  5 All         0.645  Control        12.9        1708.  13.6      1
#>  6 All         0.822  Experimental   61.5         312.  62.3      1
#>  7 All         0.868  Control         3.98         65.3  4.84     1
#>  8 All         0.959  Experimental    9.66       1980.  10.6      1
#>  9 All         1.23   Experimental   29.3        1631.  30.5      1
#> 10 All         1.27   Control        13.7         372.  15.0      1
#> 11 All         1.36   Experimental   30.6        1424.  31.9      1
#> 12 All         1.43   Control        21.9        1206.  23.3      1
#> 13 All         1.45   Control        28.2         603.  29.6      1
#> 14 All         1.55   Control        15.9        1440.  17.5      1
#> 15 All         1.72   Experimental   17.4         838.  19.2      1
#> 16 All         1.74   Experimental   70.1         930.  71.8      1
#> 17 All         2.03   Experimental   33.6        1552.  35.6      1
#> 18 All         2.24   Control         0.175      3659.   2.41     1
#> 19 All         2.56   Control         2.68        546.   5.24     1
#> 20 All         2.60   Experimental    1.49        257.   4.09     1
 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.164 Experimental   21.5         126.  21.7      1
#>  2 All          0.177 Experimental   33.1         607.  33.3      1
#>  3 All          0.207 Experimental   67.4         758.  67.6      1
#>  4 All          0.304 Control        15.2        3965.  15.5      1
#>  5 All          0.338 Experimental   10.7         790.  11.1      1
#>  6 All          0.341 Control         5.24       4031.   5.58     1
#>  7 All          0.497 Experimental   14.6        2065.  15.1      1
#>  8 All          0.579 Experimental   28.1         552.  28.7      1
#>  9 All          0.776 Experimental    0.879       951.   1.66     1
#> 10 All          0.901 Experimental   18.5         591.  19.4      1
#> 11 All          0.977 Control        11.9         114.  12.9      1
#> 12 All          1.07  Experimental    0.886        66.1  1.95     1
#> 13 All          1.13  Control         4.16       2076.   5.29     1
#> 14 All          1.26  Experimental    3.10        478.   4.36     1
#> 15 All          1.35  Experimental   12.8         330.  14.1      1
#> 16 All          1.39  Experimental   15.3        2828.  16.6      1
#> 17 All          1.41  Experimental    2.94       1953.   4.35     1
#> 18 All          1.46  Experimental    0.403       451.   1.87     1
#> 19 All          1.57  Control        17.8         147.  19.3      1
#> 20 All          1.59  Experimental   29.9        1103.  31.5      1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low         0.0155 Experimental    66.6         40.3 40.3      0
#>  2 High        0.148  Experimental    39.1        367.  39.2      1
#>  3 Low         0.175  Experimental    18.3         48.6 18.5      1
#>  4 Low         0.270  Control         11.2        486.  11.4      1
#>  5 Low         0.314  Control         10.6       1751.  10.9      1
#>  6 High        0.434  Control          4.20       640.   4.63     1
#>  7 Low         0.588  Experimental    34.8       1193.  35.3      1
#>  8 High        0.688  Experimental     6.38       284.   7.07     1
#>  9 Low         0.735  Control         40.5        186.  41.2      1
#> 10 High        0.781  Control          7.69      1191.   8.47     1
#> 11 Low         1.04   Control         10.6        186.  11.7      1
#> 12 High        1.14   Experimental    59.7         69.4 60.9      1
#> 13 High        1.21   Control          2.24       964.   3.45     1
#> 14 Low         1.30   Experimental     2.09      1263.   3.40     1
#> 15 High        1.31   Experimental    74.9       1669.  76.2      1
#> 16 High        1.58   Control          3.04      1361.   4.62     1
#> 17 Low         1.61   Control         40.1        413.  41.7      1
#> 18 Low         1.68   Experimental    59.8        785.  61.5      1
#> 19 Low         1.92   Control          9.43       181.  11.4      1
#> 20 High        2.13   Experimental    95.8        921.  98.0      1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High        0.0465 Experimental     6.30       690.    6.35     1
#>  2 Low         0.252  Experimental    11.1        533.   11.4      1
#>  3 High        0.258  Control          7.26      1579.    7.52     1
#>  4 High        0.271  Control         29.4       1175.   29.6      1
#>  5 High        0.272  Experimental     2.47       595.    2.74     1
#>  6 Low         0.705  Experimental   175.        1055.  176.       1
#>  7 High        0.738  Control         53.0       2471.   53.7      1
#>  8 High        0.959  Experimental    37.1        468.   38.0      1
#>  9 High        1.05   Experimental    16.0       1843.   17.0      1
#> 10 High        1.09   Control          4.97       774.    6.06     1
#> 11 Low         1.15   Control         43.7       1662.   44.8      1
#> 12 Low         1.22   Control          7.67        98.8   8.89     1