Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment
    = c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n
Number of observations. If length(n) > 1, the length is taken to be the number required.

strata
A tibble with strata specified in `Stratum`, probability (incidence) of each stratum in `p`

block
Vector of treatments to be included in each block

enrollRates
Enrollment rates; see details and examples

failRates
Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates
Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)

#> # A tibble: 20 x 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.127 Experimental    72.0         9.07   9.19     0
#>  2 All          0.133 Control         10.6      1284.    10.7      1
#>  3 All          0.397 Control          1.17       36.5    1.56     1
#>  4 All          0.575 Experimental     3.47     1947.     4.05     1
#>  5 All          0.596 Control         10.2       873.    10.8      1
#>  6 All          0.798 Experimental    18.7      2327.    19.5      1
#>  7 All          0.831 Experimental    18.2       618.    19.0      1
#>  8 All          0.850 Control          3.14     2649.     3.99     1
#>  9 All          0.985 Experimental    23.3      1167.    24.3      1
#> 10 All          1.02  Experimental    15.3       899.    16.3      1
#> 11 All          1.04  Control         23.1       274.    24.2      1
#> 12 All          1.15  Control         10.6       155.    11.8      1
#> 13 All          1.20  Control         37.3      1508.    38.5      1
#> 14 All          1.52  Control         25.3      3416.    26.8      1
#> 15 All          1.55  Experimental   107.        739.   109.       1
#> 16 All          1.56  Experimental    25.6      1005.    27.2      1
#> 17 All          1.64  Experimental    40.1      2037.    41.7      1
#> 18 All          1.78  Control          1.30      346.     3.08     1
#> 19 All          1.79  Control         44.6      2163.    46.4      1
#> 20 All          1.98  Experimental    11.2      1335.    13.2      1

 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))

#> # A tibble: 20 x 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.144 Experimental    60.1         537. 60.3      1
#>  2 All          0.192 Experimental    23.9         292. 24.1      1
#>  3 All          0.379 Control          7.70        960.  8.08     1
#>  4 All          0.436 Experimental     7.58        131.  8.01     1
#>  5 All          1.09  Control         46.0         855. 47.1      1
#>  6 All          1.14  Experimental     9.58        545. 10.7      1
#>  7 All          1.31  Experimental    41.8        1777. 43.1      1
#>  8 All          1.46  Experimental     6.12       1536.  7.57     1
#>  9 All          1.61  Experimental     1.40        327.  3.01     1
#> 10 All          1.68  Experimental    22.4        1645. 24.1      1
#> 11 All          1.71  Experimental     9.27        118. 11.0      1
#> 12 All          1.72  Control         19.7         512. 21.5      1
#> 13 All          1.74  Control          5.28        510.  7.02     1
#> 14 All          1.74  Experimental     3.22       1709.  4.96     1
#> 15 All          1.90  Experimental    31.3        1148. 33.2      1
#> 16 All          2.03  Experimental    28.5         938. 30.5      1
#> 17 All          2.30  Experimental     2.55       1725.  4.85     1
#> 18 All          2.39  Experimental     3.51       2093.  5.90     1
#> 19 All          2.43  Control         11.2         814. 13.6      1
#> 20 All          2.48  Experimental     3.88       1770.  6.36     1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)

#> # A tibble: 20 x 7
#> # Groups:   Stratum, Treatment [3]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High        0.0193 Control         7.35        534.   7.37      1
#>  2 High        0.137  Experimental    7.76        939.   7.90      1
#>  3 High        0.369  Control         5.54        651.   5.91      1
#>  4 High        0.407  Experimental    0.112      1525.   0.519     1
#>  5 High        0.435  Control        19.5         439.  19.9       1
#>  6 High        0.633  Experimental   60.9          63.0 61.6       1
#>  7 High        0.674  Experimental   27.1          22.8 23.5       0
#>  8 High        0.720  Control         0.787      1219.   1.51      1
#>  9 High        0.791  Experimental    8.41         33.1  9.20      1
#> 10 High        1.13   Experimental   34.0          59.0 35.1       1
#> 11 High        1.15   Control         8.60        230.   9.75      1
#> 12 High        1.19   Control         2.93        677.   4.12      1
#> 13 High        1.55   Control         4.50         74.1  6.05      1
#> 14 Low         1.89   Control         7.72       1245.   9.62      1
#> 15 High        1.98   Experimental   34.9         811.  36.9       1
#> 16 High        2.09   Control         4.91       1244.   7.00      1
#> 17 High        2.19   Experimental   18.1        4649.  20.3       1
#> 18 High        2.32   Experimental   21.6        1187.  23.9       1
#> 19 High        2.34   Control        35.3         460.  37.7       1
#> 20 High        2.39   Control         3.60        750.   5.99      1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)

#> # A tibble: 12 x 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High        0.0761 Control       16.9         1338.  16.9       1
#>  2 High        0.203  Experimental  33.5         1113.  33.7       1
#>  3 High        0.421  Experimental  11.8          407.  12.3       1
#>  4 Low         0.555  Experimental  40.8         1378.  41.3       1
#>  5 High        0.699  Control        3.11         504.   3.81      1
#>  6 High        0.841  Experimental   0.0395      1426.   0.881     1
#>  7 Low         1.15   Experimental  48.6          251.  49.8       1
#>  8 High        1.35   Control       22.3         1587.  23.6       1
#>  9 High        1.41   Experimental  68.0          161.  69.4       1
#> 10 High        1.46   Control       28.0          376.  29.5       1
#> 11 Low         1.68   Control       17.5          879.  19.2       1
#> 12 High        1.76   Control       32.5           78.5 34.3       1

>>>>>>> 8da6ac0d6f0042026e5db4aae1934a972cd63c44