Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment
    = c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n
Number of observations. If length(n) > 1, the length is taken to be the number required.

strata
A tibble with strata specified in `Stratum`, probability (incidence) of each stratum in `p`

block
Vector of treatments to be included in each block

enrollRates
Enrollment rates; see details and examples

failRates
Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates
Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)

#> # A tibble: 20 x 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0284 Experimental    2.83        2024.  2.86     1
#>  2 All         0.246  Experimental   37.1          967. 37.3      1
#>  3 All         0.502  Control         4.98        1634.  5.48     1
#>  4 All         0.757  Control         1.79         884.  2.54     1
#>  5 All         0.767  Control         1.85         207.  2.62     1
#>  6 All         0.777  Experimental   30.4         1972. 31.2      1
#>  7 All         0.861  Control         4.66         310.  5.53     1
#>  8 All         1.04   Experimental    0.676       2931.  1.71     1
#>  9 All         1.48   Experimental   13.7          288. 15.2      1
#> 10 All         1.49   Experimental   11.1         1282. 12.6      1
#> 11 All         1.55   Control         9.61        1736. 11.2      1
#> 12 All         1.73   Control        20.2          166. 21.9      1
#> 13 All         2.33   Experimental    6.81        1137.  9.13     1
#> 14 All         2.33   Control         3.13         386.  5.46     1
#> 15 All         2.36   Experimental   20.5         1240. 22.9      1
#> 16 All         2.57   Control        76.6          687. 79.1      1
#> 17 All         2.62   Experimental    2.46         451.  5.08     1
#> 18 All         2.74   Control         7.80        1028. 10.5      1
#> 19 All         2.74   Experimental   11.9          589. 14.6      1
#> 20 All         2.87   Control         3.81        1206.  6.68     1

 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))

#> # A tibble: 20 x 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.129 Experimental    12.2        220.   12.3      1
#>  2 All          0.130 Experimental    45.0       1166.   45.1      1
#>  3 All          0.132 Experimental    61.0        221.   61.1      1
#>  4 All          0.134 Control         17.3         90.3  17.4      1
#>  5 All          0.167 Experimental     9.66       259.    9.83     1
#>  6 All          0.215 Experimental    11.1        637.   11.3      1
#>  7 All          0.276 Control         44.3       1668.   44.5      1
#>  8 All          0.694 Experimental     2.44       102.    3.13     1
#>  9 All          0.710 Experimental   100.         523.  101.       1
#> 10 All          0.905 Experimental    17.9        284.   18.8      1
#> 11 All          0.926 Control          5.69       387.    6.62     1
#> 12 All          1.02  Experimental    40.2        953.   41.2      1
#> 13 All          1.08  Experimental     6.39        61.0   7.47     1
#> 14 All          1.10  Control          7.97      1310.    9.07     1
#> 15 All          1.39  Experimental    17.1        980.   18.5      1
#> 16 All          1.51  Experimental    10.7         96.0  12.2      1
#> 17 All          1.55  Experimental     5.43      1348.    6.98     1
#> 18 All          1.60  Experimental    23.0        329.   24.6      1
#> 19 All          1.71  Experimental    23.2       1445.   25.0      1
#> 20 All          1.79  Control         11.7        876.   13.5      1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)

#> # A tibble: 20 x 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low          0.214 Control        51.0        425.   51.2      1
#>  2 High         0.499 Experimental   67.9        776.   68.4      1
#>  3 High         0.522 Control         4.98      1611.    5.50     1
#>  4 Low          0.635 Experimental   25.4       3510.   26.1      1
#>  5 Low          0.678 Control        21.4        324.   22.1      1
#>  6 High         0.740 Control         3.85      1123.    4.59     1
#>  7 High         1.05  Experimental   31.4        418.   32.4      1
#>  8 Low          1.56  Experimental   33.2       3583.   34.8      1
#>  9 Low          1.58  Experimental   74.7          7.82  9.40     0
#> 10 High         1.67  Control         9.36       708.   11.0      1
#> 11 High         1.72  Control        21.8         75.2  23.5      1
#> 12 Low          1.76  Experimental   61.9        498.   63.6      1
#> 13 High         1.77  Experimental    0.319      228.    2.09     1
#> 14 Low          1.98  Control        19.4       2012.   21.4      1
#> 15 High         2.13  Experimental    0.470      350.    2.60     1
#> 16 Low          2.33  Control        53.4       2780.   55.7      1
#> 17 Low          2.38  Experimental   26.7       1515.   29.0      1
#> 18 High         2.44  Experimental   58.3       1577.   60.8      1
#> 19 High         2.58  Control        15.4        390.   17.9      1
#> 20 Low          2.69  Control         5.18       304.    7.87     1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)

#> # A tibble: 12 x 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High        0.0664 Experimental    2.12       553.    2.18     1
#>  2 High        0.171  Control        18.2        149.   18.3      1
#>  3 High        0.301  Control        64.6        182.   65.0      1
#>  4 Low         0.322  Experimental    7.56      3880.    7.89     1
#>  5 High        0.378  Experimental    1.58       158.    1.96     1
#>  6 High        0.427  Experimental    0.914     3270.    1.34     1
#>  7 High        0.638  Experimental   14.8        268.   15.5      1
#>  8 High        0.670  Control        48.6        166.   49.3      1
#>  9 High        0.883  Control         1.20      1480.    2.08     1
#> 10 High        0.927  Control        30.6        977.   31.5      1
#> 11 Low         1.15   Experimental    0.334     1116.    1.48     1
#> 12 Low         1.18   Control        32.7          9.13 10.3      0