Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.178 Experimental  47.2         5523.  47.3      1
#>  2 All          0.425 Control       10.7          971.  11.2      1
#>  3 All          0.504 Experimental   1.12         469.   1.62     1
#>  4 All          0.787 Control        3.33         166.   4.12     1
#>  5 All          0.789 Experimental   2.49          39.2  3.28     1
#>  6 All          0.806 Experimental   5.56         250.   6.37     1
#>  7 All          0.876 Control       13.9         1262.  14.8      1
#>  8 All          0.889 Control       80.0          739.  80.9      1
#>  9 All          0.927 Experimental   0.453       2574.   1.38     1
#> 10 All          1.02  Control       15.0          613.  16.1      1
#> 11 All          1.62  Control       20.0          650.  21.6      1
#> 12 All          1.88  Experimental  43.1          261.  45.0      1
#> 13 All          2.33  Control       13.2           11.2 13.5      0
#> 14 All          2.35  Experimental   8.44        1219.  10.8      1
#> 15 All          2.41  Experimental  43.8          574.  46.2      1
#> 16 All          2.46  Control       15.9          927.  18.4      1
#> 17 All          2.59  Control       11.3           22.4 13.9      1
#> 18 All          2.60  Control        0.0932        28.3  2.69     1
#> 19 All          2.67  Experimental  26.8          543.  29.4      1
#> 20 All          2.89  Experimental   2.49         610.   5.38     1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0173 Control        20.9        1641.  20.9      1
#>  2 All         0.0315 Experimental   36.1          69.1 36.1      1
#>  3 All         0.154  Experimental   21.0        1556.  21.2      1
#>  4 All         0.190  Experimental   10.6         195.  10.8      1
#>  5 All         0.322  Control         1.54        313.   1.86     1
#>  6 All         0.505  Experimental    5.33        515.   5.83     1
#>  7 All         0.570  Experimental    3.31        737.   3.88     1
#>  8 All         0.889  Experimental    9.34       1712.  10.2      1
#>  9 All         0.912  Experimental   25.2        1471.  26.2      1
#> 10 All         0.916  Control         8.14        136.   9.06     1
#> 11 All         0.935  Experimental   12.9        1034.  13.8      1
#> 12 All         0.984  Experimental    3.48       3134.   4.47     1
#> 13 All         1.00   Experimental    1.52       2312.   2.53     1
#> 14 All         1.09   Experimental   12.9         440.  14.0      1
#> 15 All         1.20   Experimental   31.8          65.3 33.0      1
#> 16 All         1.23   Control        16.5        2102.  17.7      1
#> 17 All         1.34   Experimental    7.32        434.   8.66     1
#> 18 All         2.03   Experimental   28.2        1956.  30.2      1
#> 19 All         2.17   Experimental    1.25       1226.   3.42     1
#> 20 All         2.19   Control         0.413        49.2  2.61     1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High        0.0372 Experimental     3.83       875.    3.86     1
#>  2 High        0.394  Control         13.3        256.   13.7      1
#>  3 Low         0.443  Control         34.2       2975.   34.7      1
#>  4 Low         0.569  Experimental    65.6       1030.   66.2      1
#>  5 Low         0.572  Experimental   143.          83.3  83.9      0
#>  6 High        0.604  Experimental     7.83      1680.    8.43     1
#>  7 High        0.700  Control         10.6       1923.   11.3      1
#>  8 Low         0.814  Control         11.0        576.   11.9      1
#>  9 Low         0.883  Control          7.06       398.    7.95     1
#> 10 Low         0.903  Experimental    67.8        413.   68.7      1
#> 11 Low         1.12   Experimental    69.9        541.   71.0      1
#> 12 Low         1.21   Control         46.1        150.   47.3      1
#> 13 Low         1.33   Experimental     9.96        10.2  11.3      1
#> 14 Low         1.34   Experimental    17.1         62.9  18.4      1
#> 15 High        1.43   Experimental   130.        1356.  131.       1
#> 16 High        1.53   Control         17.2       1188.   18.8      1
#> 17 High        1.53   Control          3.79       939.    5.33     1
#> 18 High        1.70   Experimental    30.4       4124.   32.1      1
#> 19 High        1.85   Control          3.01      1978.    4.87     1
#> 20 High        2.07   Control          6.65        67.0   8.73     1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 High         0.213 Control        23.5          897.  23.7       1
#>  2 Low          0.336 Control        16.8          654.  17.2       1
#>  3 High         0.361 Experimental   22.5          152.  22.9       1
#>  4 High         0.671 Experimental    0.270        200.   0.940     1
#>  5 High         0.739 Control        14.9         3033.  15.6       1
#>  6 High         0.777 Control         0.794        278.   1.57      1
#>  7 Low          0.790 Experimental  107.          2241. 108.        1
#>  8 High         0.829 Experimental   26.7          334.  27.5       1
#>  9 Low          0.857 Control        12.1          612.  12.9       1
#> 10 High         0.999 Control        17.6         3557.  18.6       1
#> 11 High         1.06  Experimental   40.0         3433.  41.0       1
#> 12 High         1.22  Experimental    3.09         374.   4.30      1