Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All        0.00476 Control         6.27        207.   6.27     1
#>  2 All        0.0281  Experimental   13.9         690.  13.9      1
#>  3 All        0.330   Experimental    2.91        616.   3.24     1
#>  4 All        0.361   Control         4.87         88.0  5.23     1
#>  5 All        0.758   Control         0.349       636.   1.11     1
#>  6 All        1.10    Experimental    2.52        834.   3.62     1
#>  7 All        1.14    Control         6.42       1288.   7.56     1
#>  8 All        1.23    Experimental   17.3         111.  18.5      1
#>  9 All        1.39    Control        17.6         286.  19.0      1
#> 10 All        1.55    Control         0.539       486.   2.09     1
#> 11 All        1.85    Experimental   43.2         253.  45.1      1
#> 12 All        1.88    Experimental   53.7        1187.  55.6      1
#> 13 All        2.02    Control         3.34        242.   5.36     1
#> 14 All        2.25    Experimental   13.1         347.  15.3      1
#> 15 All        2.28    Experimental    5.97       1953.   8.25     1
#> 16 All        2.28    Control        23.0        1172.  25.3      1
#> 17 All        2.44    Experimental    3.98        552.   6.42     1
#> 18 All        2.48    Control        33.0         721.  35.5      1
#> 19 All        2.51    Control        18.5         150.  21.0      1
#> 20 All        2.54    Experimental   31.1          89.6 33.7      1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0121 Experimental   57.6        1223.  57.6      1
#>  2 All         0.102  Experimental   19.1        1945.  19.2      1
#>  3 All         0.157  Experimental    2.22       1106.   2.38     1
#>  4 All         0.228  Control         3.28        594.   3.50     1
#>  5 All         0.588  Experimental    0.911       953.   1.50     1
#>  6 All         0.591  Experimental   20.7        1145.  21.3      1
#>  7 All         0.763  Experimental   61.0          63.5 61.7      1
#>  8 All         0.823  Control         8.14        385.   8.97     1
#>  9 All         0.853  Experimental   38.3         660.  39.1      1
#> 10 All         0.863  Experimental   19.4         937.  20.3      1
#> 11 All         0.893  Experimental    2.55        631.   3.44     1
#> 12 All         1.02   Control         2.34       1611.   3.36     1
#> 13 All         1.03   Experimental    0.873       766.   1.91     1
#> 14 All         1.68   Experimental   15.6        2134.  17.3      1
#> 15 All         1.69   Experimental   13.4         379.  15.1      1
#> 16 All         1.78   Control        19.1         624.  20.9      1
#> 17 All         1.79   Control         3.99       1506.   5.78     1
#> 18 All         2.01   Experimental   27.7        1007.  29.7      1
#> 19 All         2.04   Experimental   50.7        1533.  52.8      1
#> 20 All         2.23   Experimental   23.9        1142.  26.2      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 Low         0.0856 Experimental    0.812       242.    0.897     1
#>  2 High        0.142  Control        10.2         372.   10.3       1
#>  3 High        0.440  Control         3.55       3738.    3.99      1
#>  4 Low         0.482  Control         1.05        397.    1.53      1
#>  5 High        0.550  Experimental   29.4        2877.   30.0       1
#>  6 High        0.570  Experimental    1.61        590.    2.18      1
#>  7 High        0.688  Control         8.30       1662.    8.99      1
#>  8 High        0.832  Experimental    0.846       775.    1.68      1
#>  9 Low         0.849  Control        18.5         690.   19.4       1
#> 10 Low         0.889  Experimental   12.6         170.   13.5       1
#> 11 High        0.940  Experimental   37.0         256.   38.0       1
#> 12 Low         0.981  Experimental   89.4         607.   90.4       1
#> 13 Low         1.00   Control        22.1        3532.   23.1       1
#> 14 High        1.05   Control         2.90         24.9   3.94      1
#> 15 High        1.07   Experimental   25.5         147.   26.6       1
#> 16 High        1.17   Experimental   15.9         251.   17.1       1
#> 17 High        1.34   Control         1.25        740.    2.59      1
#> 18 High        1.47   Control        13.9        4171.   15.4       1
#> 19 High        2.01   Experimental  122.          288.  124.        1
#> 20 High        2.02   Control         2.60       4077.    4.63      1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High         0.201 Experimental   9.04          64.6  9.24      1
#>  2 Low          0.283 Experimental  13.2         1704.  13.4       1
#>  3 High         0.300 Control        0.0986       809.   0.399     1
#>  4 High         0.622 Experimental   6.15        1218.   6.77      1
#>  5 High         0.816 Control        2.68        1361.   3.49      1
#>  6 High         0.817 Experimental  46.9          471.  47.7       1
#>  7 High         0.839 Control        1.86        1797.   2.70      1
#>  8 High         0.853 Control       16.2          589.  17.0       1
#>  9 High         0.894 Experimental  31.1          394.  32.0       1
#> 10 Low          0.936 Control        1.61         575.   2.55      1
#> 11 High         0.952 Experimental  31.3          402.  32.2       1
#> 12 High         0.955 Experimental   6.50         301.   7.46      1