Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0277 Experimental   29.2          10.8 10.9      0
#>  2 All         0.0489 Experimental   10.9        3309.  10.9      1
#>  3 All         0.387  Control         9.71        827.  10.1      1
#>  4 All         0.484  Control         1.89       1844.   2.37     1
#>  5 All         0.736  Control        16.5        2139.  17.2      1
#>  6 All         0.939  Experimental   18.4        2169.  19.4      1
#>  7 All         0.966  Control        17.6        1555.  18.6      1
#>  8 All         1.01   Experimental    0.214      1622.   1.22     1
#>  9 All         1.11   Control         1.33        922.   2.44     1
#> 10 All         1.17   Control         3.50       1014.   4.67     1
#> 11 All         1.27   Experimental    4.16       2424.   5.43     1
#> 12 All         1.29   Experimental   21.5        1458.  22.8      1
#> 13 All         1.35   Control         1.61        197.   2.96     1
#> 14 All         1.36   Experimental    7.41       1539.   8.77     1
#> 15 All         1.39   Control         1.02       1129.   2.41     1
#> 16 All         1.44   Experimental    3.40        773.   4.83     1
#> 17 All         1.49   Control        18.7         233.  20.1      1
#> 18 All         1.65   Control         0.324       161.   1.98     1
#> 19 All         1.72   Experimental   61.7         716.  63.5      1
#> 20 All         1.74   Experimental   30.9        1701.  32.6      1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0368 Control       10.0          166.  10.1      1
#>  2 All         0.0555 Experimental  22.6          827.  22.7      1
#>  3 All         0.0834 Experimental  37.2         1397.  37.3      1
#>  4 All         0.258  Experimental 113.            31.4 31.7      0
#>  5 All         0.424  Control        0.949       1220.   1.37     1
#>  6 All         0.426  Experimental  11.9          357.  12.3      1
#>  7 All         0.548  Experimental  59.9         2556.  60.5      1
#>  8 All         0.699  Experimental   5.57         657.   6.27     1
#>  9 All         0.813  Experimental  40.9         1006.  41.7      1
#> 10 All         0.820  Control        5.78         587.   6.60     1
#> 11 All         0.892  Experimental   9.20         404.  10.1      1
#> 12 All         0.933  Experimental  25.3         1316.  26.2      1
#> 13 All         0.955  Control        8.24         498.   9.20     1
#> 14 All         0.973  Experimental  14.3          208.  15.3      1
#> 15 All         1.05   Experimental   0.167       3361.   1.21     1
#> 16 All         1.19   Experimental   0.0974      1514.   1.29     1
#> 17 All         1.23   Experimental  13.2         1429.  14.4      1
#> 18 All         1.46   Experimental  18.6          410.  20.1      1
#> 19 All         1.59   Experimental   2.16        1098.   3.75     1
#> 20 All         1.69   Control       12.8          474.  14.4      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High         0.162 Control        25.1         109.  25.3       1
#>  2 High         0.181 Experimental    6.43       1326.   6.61      1
#>  3 High         0.350 Experimental    0.294       248.   0.644     1
#>  4 High         0.394 Control        12.9         279.  13.3       1
#>  5 High         0.443 Experimental    8.34        599.   8.79      1
#>  6 Low          0.488 Experimental   14.2        1109.  14.7       1
#>  7 High         0.656 Control        10.5          14.2 11.1       1
#>  8 High         0.676 Experimental   34.7         323.  35.3       1
#>  9 High         0.811 Control         8.44        747.   9.25      1
#> 10 High         1.11  Control         7.43       4275.   8.54      1
#> 11 High         1.18  Experimental    1.07        167.   2.25      1
#> 12 High         1.19  Experimental    3.25        993.   4.44      1
#> 13 High         1.24  Control         3.61        573.   4.86      1
#> 14 Low          1.34  Experimental   42.3        1564.  43.6       1
#> 15 High         1.38  Control         6.04        528.   7.42      1
#> 16 Low          1.72  Control        38.8        1616.  40.5       1
#> 17 High         1.73  Control         5.49       1363.   7.22      1
#> 18 Low          1.75  Control        14.9         450.  16.7       1
#> 19 High         1.88  Experimental   16.3         515.  18.1       1
#> 20 High         1.95  Experimental   11.8         237.  13.7       1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [3]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High       0.00479 Control         21.1        2922.  21.1      1
#>  2 High       0.0371  Experimental   103.          881. 103.       1
#>  3 High       0.154   Experimental     3.65        508.   3.81     1
#>  4 High       0.184   Control         13.6         293.  13.8      1
#>  5 High       0.317   Experimental    36.5         530.  36.9      1
#>  6 High       0.546   Experimental    29.6        2462.  30.1      1
#>  7 Low        0.564   Experimental   101.          391. 101.       1
#>  8 High       0.730   Control         30.4         168.  31.2      1
#>  9 High       0.734   Control         13.8         194.  14.5      1
#> 10 High       0.814   Experimental     4.96       1244.   5.77     1
#> 11 High       0.854   Experimental    89.2        1330.  90.0      1
#> 12 High       1.20    Control         11.4         285.  12.6      1