Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
simPWSurv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.262 Control        4.85         374.   5.11      1
#>  2 All          0.510 Control       31.6           77.5 32.1       1
#>  3 All          0.666 Experimental   5.28         895.   5.94      1
#>  4 All          0.693 Experimental  49.4          147.  50.1       1
#>  5 All          0.722 Control       28.4         1990.  29.1       1
#>  6 All          0.827 Experimental   4.69         366.   5.51      1
#>  7 All          0.832 Control        0.745       2598.   1.58      1
#>  8 All          0.833 Experimental   1.45        1493.   2.28      1
#>  9 All          0.834 Control       31.7         1084.  32.5       1
#> 10 All          0.849 Experimental   0.392        264.   1.24      1
#> 11 All          0.864 Experimental  93.0          261.  93.9       1
#> 12 All          0.902 Control        0.0657        61.6  0.968     1
#> 13 All          0.927 Experimental   9.20        1022.  10.1       1
#> 14 All          0.971 Control       25.7          240.  26.7       1
#> 15 All          1.17  Control       23.8          349.  24.9       1
#> 16 All          1.25  Experimental  28.9          310.  30.2       1
#> 17 All          1.30  Experimental  18.2         2096.  19.6       1
#> 18 All          1.33  Experimental  49.4           40.6 41.9       0
#> 19 All          1.37  Control       37.6          331.  39.0       1
#> 20 All          1.55  Control        1.05        3707.   2.59      1

# example 2
# 3:1 randomization
simPWSurv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All         0.0955 Experimental    4.32       1483.   4.41      1
#>  2 All         0.120  Experimental   50.1         979.  50.2       1
#>  3 All         0.176  Control         0.216       231.   0.392     1
#>  4 All         0.273  Experimental    4.33       2361.   4.61      1
#>  5 All         0.300  Control         4.09       1585.   4.39      1
#>  6 All         0.309  Experimental    1.88       1232.   2.19      1
#>  7 All         0.311  Experimental   45.6         208.  45.9       1
#>  8 All         0.340  Experimental    1.23       2063.   1.57      1
#>  9 All         0.601  Experimental    3.60        350.   4.20      1
#> 10 All         0.934  Experimental   55.3         421.  56.3       1
#> 11 All         0.953  Experimental    1.17        901.   2.12      1
#> 12 All         1.28   Control         1.85       1045.   3.13      1
#> 13 All         1.44   Control         5.22       3260.   6.66      1
#> 14 All         1.55   Experimental    9.23         25.8 10.8       1
#> 15 All         1.67   Experimental    1.43        611.   3.10      1
#> 16 All         1.72   Experimental   23.3         398.  25.0       1
#> 17 All         1.96   Experimental   65.0         759.  67.0       1
#> 18 All         1.98   Experimental    1.18       1690.   3.16      1
#> 19 All         2.00   Experimental    8.75        764.  10.8       1
#> 20 All         2.28   Control         3.01        257.   5.30      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.204 Control          2.99       115.   3.19     1
#>  2 Low          0.226 Experimental    38.1       4148.  38.3      1
#>  3 Low          0.237 Control          3.62       684.   3.86     1
#>  4 Low          0.316 Experimental     1.71       123.   2.02     1
#>  5 High         0.611 Experimental    10.0        761.  10.6      1
#>  6 High         0.733 Control         26.1         76.7 26.8      1
#>  7 Low          0.754 Control          4.94       369.   5.69     1
#>  8 High         0.779 Experimental    54.9        746.  55.7      1
#>  9 Low          0.855 Control         20.0       2199.  20.8      1
#> 10 High         1.32  Control          8.07        32.9  9.39     1
#> 11 High         1.45  Experimental     6.75      3416.   8.20     1
#> 12 High         1.55  Experimental    10.4        459.  11.9      1
#> 13 High         1.59  Control         15.0        409.  16.6      1
#> 14 Low          1.60  Experimental    94.5       4413.  96.1      1
#> 15 High         1.70  Experimental    66.0        543.  67.7      1
#> 16 High         1.88  Control          8.84       779.  10.7      1
#> 17 High         1.99  Experimental    22.4       1942.  24.4      1
#> 18 Low          2.16  Experimental    18.2       1323.  20.3      1
#> 19 High         2.30  Control         15.7         42.0 18.0      1
#> 20 Low          2.36  Control         46.4         28.5 30.9      0
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

simPWSurv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High         0.199 Control        27.3         984.   27.5      1
#>  2 High         0.368 Control         1.85        905.    2.22     1
#>  3 High         0.729 Experimental    0.581       812.    1.31     1
#>  4 Low          0.889 Experimental    7.09         87.3   7.98     1
#>  5 Low          0.957 Control         1.39       3628.    2.35     1
#>  6 High         0.994 Experimental    2.74       1646.    3.74     1
#>  7 High         1.16  Experimental  137.         1428.  138.       1
#>  8 High         1.21  Control         4.98        584.    6.19     1
#>  9 High         1.23  Control        10.9        1736.   12.2      1
#> 10 High         1.63  Experimental    8.89        105.   10.5      1
#> 11 High         1.69  Experimental    8.92       2880.   10.6      1
#> 12 Low          1.85  Control        19.8         478.   21.7      1