Simulate a stratified time-to-event outcome randomized trial

simPWSurv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simPWSurv(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All         0.0126 Experimental  73.9        1041.    73.9      1
#>  2 All         0.190  Control       38.7         150.    38.9      1
#>  3 All         0.252  Control       19.4         250.    19.7      1
#>  4 All         0.259  Experimental  18.2         670.    18.5      1
#>  5 All         0.405  Control       27.7         177.    28.1      1
#>  6 All         0.560  Experimental  34.5         110.    35.1      1
#>  7 All         0.566  Experimental  37.9         325.    38.5      1
#>  8 All         0.783  Control        1.81        262.     2.60     1
#>  9 All         0.850  Experimental   2.69        102.     3.54     1
#> 10 All         0.865  Control       46.6         523.    47.4      1
#> 11 All         1.00   Control        6.74       2173.     7.74     1
#> 12 All         1.07   Experimental  12.3         206.    13.3      1
#> 13 All         1.12   Control        9.47        242.    10.6      1
#> 14 All         1.29   Experimental  29.9         263.    31.1      1
#> 15 All         1.65   Control        2.91        521.     4.56     1
#> 16 All         1.73   Experimental   2.88         96.8    4.61     1
#> 17 All         1.79   Experimental  14.2        1093.    16.0      1
#> 18 All         1.80   Control        0.0223      667.     1.83     1
#> 19 All         1.84   Experimental 102.         2481.   104.       1
#> 20 All         2.01   Control        6.88          1.18   3.19     0
 # 3:1 randomization
 simPWSurv(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.329 Experimental   22.7        5846.  23.0       1
#>  2 All          0.385 Experimental    1.00        203.   1.39      1
#>  3 All          0.458 Experimental    0.535        55.5  0.993     1
#>  4 All          0.573 Control        26.3         114.  26.9       1
#>  5 All          0.587 Control        39.1          34.2 34.8       0
#>  6 All          0.595 Experimental    1.32       1422.   1.92      1
#>  7 All          0.696 Experimental    8.87         79.7  9.57      1
#>  8 All          0.715 Experimental   48.4          50.1 49.1       1
#>  9 All          0.728 Experimental   41.5          33.3 34.1       0
#> 10 All          0.746 Experimental   16.1         446.  16.8       1
#> 11 All          0.788 Control        53.9         899.  54.7       1
#> 12 All          0.815 Experimental    8.60        817.   9.42      1
#> 13 All          0.833 Experimental   71.7         520.  72.5       1
#> 14 All          1.08  Control         4.43        239.   5.50      1
#> 15 All          1.09  Experimental    3.60       1104.   4.69      1
#> 16 All          1.09  Experimental    6.68        616.   7.77      1
#> 17 All          1.18  Experimental   50.2         642.  51.3       1
#> 18 All          1.39  Experimental   41.3         415.  42.6       1
#> 19 All          1.39  Control         1.42       2505.   2.81      1
#> 20 All          1.44  Experimental   48.7         806.  50.1       1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simPWSurv(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low         0.0346 Control         4.64        211.   4.68     1
#>  2 High        0.0359 Experimental    1.84        435.   1.88     1
#>  3 High        0.146  Control         8.23       2153.   8.37     1
#>  4 Low         0.227  Control        21.8         668.  22.1      1
#>  5 Low         0.241  Experimental   29.1          64.1 29.3      1
#>  6 High        0.297  Control         1.06        883.   1.36     1
#>  7 High        0.438  Experimental    1.73       1481.   2.16     1
#>  8 High        0.441  Experimental   81.1         999.  81.5      1
#>  9 Low         0.467  Experimental    0.716      1226.   1.18     1
#> 10 High        0.635  Experimental   13.8         176.  14.4      1
#> 11 Low         0.744  Control        32.2        1210.  33.0      1
#> 12 Low         0.884  Experimental   58.0        1010.  58.9      1
#> 13 High        1.05   Control         4.68        127.   5.73     1
#> 14 High        1.06   Control         7.33        698.   8.40     1
#> 15 Low         1.14   Control         5.84       1703.   6.98     1
#> 16 High        1.20   Control         3.86        171.   5.06     1
#> 17 High        1.37   Control         7.76       1813.   9.14     1
#> 18 High        1.46   Experimental   24.1        3368.  25.6      1
#> 19 Low         1.53   Experimental   10.1         271.  11.7      1
#> 20 Low         1.81   Experimental   79.0        1366.  80.8      1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simPWSurv(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High       0.00859 Experimental    13.3        2897. 13.3      1
#>  2 High       0.0764  Control          3.13        629.  3.20     1
#>  3 Low        0.694   Experimental    40.9         778. 41.6      1
#>  4 Low        0.696   Experimental    23.6         605. 24.3      1
#>  5 High       0.784   Control          2.83        482.  3.61     1
#>  6 High       0.789   Experimental    46.4         932. 47.2      1
#>  7 Low        1.08    Control         21.3        4101. 22.3      1
#>  8 Low        1.23    Control         33.0         746. 34.2      1
#>  9 High       1.30    Experimental    14.4         212. 15.7      1
#> 10 Low        1.34    Experimental    13.0        2163. 14.3      1
#> 11 Low        1.65    Experimental    58.4         137. 60.1      1
#> 12 High       1.67    Control         24.7        1531. 26.4      1