Simulate a stratified time-to-event outcome randomized trial

simPWSurv_ enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv_(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simtrial:::simPWSurv_(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.227 Experimental    64.2       517.    64.4      1
#>  2 All          0.340 Experimental    80.0      1019.    80.4      1
#>  3 All          0.426 Control         20.7       743.    21.1      1
#>  4 All          0.478 Control         41.2       884.    41.7      1
#>  5 All          0.528 Experimental     3.21     2950.     3.74     1
#>  6 All          0.537 Control          4.08      440.     4.62     1
#>  7 All          0.559 Control          3.34      239.     3.90     1
#>  8 All          0.688 Experimental     7.19      284.     7.88     1
#>  9 All          0.784 Control         13.6      2477.    14.4      1
#> 10 All          0.859 Control         16.4         9.92  10.8      0
#> 11 All          0.925 Experimental     1.04      867.     1.97     1
#> 12 All          0.950 Experimental    22.7      1103.    23.6      1
#> 13 All          0.950 Experimental     2.15      264.     3.10     1
#> 14 All          1.00  Control         16.2       196.    17.2      1
#> 15 All          1.03  Control          1.72      621.     2.75     1
#> 16 All          1.36  Experimental    99.9      2983.   101.       1
#> 17 All          1.73  Control          2.38     1189.     4.11     1
#> 18 All          2.03  Control          8.22     1573.    10.2      1
#> 19 All          2.07  Experimental    39.4       330.    41.5      1
#> 20 All          2.08  Experimental    25.4      1136.    27.5      1
 # 3:1 randomization
 simtrial:::simPWSurv_(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.197 Control         29.9         103. 30.1      1
#>  2 All          0.266 Experimental     4.63        440.  4.90     1
#>  3 All          0.336 Experimental    22.6        1095. 22.9      1
#>  4 All          0.514 Experimental     2.63       2741.  3.14     1
#>  5 All          0.557 Control         19.0         285. 19.5      1
#>  6 All          0.691 Experimental    21.7         629. 22.4      1
#>  7 All          0.741 Experimental    13.9         461. 14.7      1
#>  8 All          0.834 Experimental    39.6         276. 40.4      1
#>  9 All          1.14  Experimental     7.72        317.  8.87     1
#> 10 All          1.18  Experimental    33.3        1616. 34.5      1
#> 11 All          1.19  Control          1.39        791.  2.57     1
#> 12 All          1.21  Experimental     5.46        564.  6.66     1
#> 13 All          1.23  Experimental     9.60        486. 10.8      1
#> 14 All          1.30  Experimental    37.1        1532. 38.4      1
#> 15 All          1.35  Experimental    31.0        1847. 32.3      1
#> 16 All          1.45  Control          5.40        164.  6.84     1
#> 17 All          1.59  Experimental     2.77       1545.  4.36     1
#> 18 All          1.69  Control         21.2        2859. 22.9      1
#> 19 All          1.75  Experimental     6.14        294.  7.90     1
#> 20 All          1.82  Experimental     8.82        574. 10.6      1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simtrial:::simPWSurv_(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime     cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 High        0.0295 Control         0.422      1560.    0.451     1
#>  2 High        0.195  Experimental   47.1         513.   47.3       1
#>  3 High        0.259  Experimental    9.62        612.    9.88      1
#>  4 High        0.346  Control         0.108        89.1   0.454     1
#>  5 High        0.386  Control         6.37       2033.    6.75      1
#>  6 Low         0.439  Control        13.9        1469.   14.4       1
#>  7 High        0.488  Experimental    7.37        575.    7.85      1
#>  8 Low         0.635  Control        13.7         209.   14.4       1
#>  9 High        0.855  Experimental    4.52        669.    5.37      1
#> 10 High        0.922  Control         0.134      1009.    1.06      1
#> 11 High        1.04   Control        30.7         466.   31.8       1
#> 12 High        1.06   Control        13.9         297.   14.9       1
#> 13 High        1.41   Experimental    7.39        676.    8.80      1
#> 14 Low         1.43   Experimental  119.          361.  121.        1
#> 15 High        1.55   Experimental   18.4          21.4  20.0       1
#> 16 High        1.61   Experimental   67.3        2766.   68.9       1
#> 17 Low         1.68   Experimental    6.41         64.0   8.09      1
#> 18 Low         1.70   Control         8.70        538.   10.4       1
#> 19 Low         1.75   Control         9.12        133.   10.9       1
#> 20 High        1.76   Experimental   38.6        1809.   40.3       1

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simtrial:::simPWSurv_(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime    cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 Low         0.0602 Control        35.0         3610. 35.1       1
#>  2 High        0.151  Control         0.387        281.  0.539     1
#>  3 High        0.299  Control        25.8          795. 26.1       1
#>  4 High        0.495  Experimental   73.7          110. 74.2       1
#>  5 High        0.499  Experimental    2.05        1110.  2.55      1
#>  6 High        0.706  Control        13.8         1512. 14.5       1
#>  7 High        0.813  Control        47.9         4795. 48.7       1
#>  8 Low         0.874  Experimental   41.3         2626. 42.2       1
#>  9 High        1.02   Experimental   28.7         1043. 29.7       1
#> 10 High        1.12   Experimental   12.0          758. 13.1       1
#> 11 High        1.31   Control        16.1          401. 17.4       1
#> 12 High        1.36   Experimental    3.47        1698.  4.83      1