Simulate a stratified time-to-event outcome randomized trial

simPWSurv_ enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

simPWSurv_(
  n = 100,
  strata = tibble::tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enrollRates = tibble::tibble(rate = 9, duration = 1),
  failRates = tibble::tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble::tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enrollRates

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enrollTime (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
# Tests
 simtrial:::simPWSurv_(n=20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0464 Control         1.85        229.   1.90     1
#>  2 All         0.195  Experimental    2.21        294.   2.41     1
#>  3 All         0.230  Experimental   12.7         216.  13.0      1
#>  4 All         0.326  Control        13.7        2207.  14.0      1
#>  5 All         0.474  Control        40.7         798.  41.1      1
#>  6 All         0.490  Experimental    3.44       1199.   3.93     1
#>  7 All         0.581  Control        34.7         458.  35.2      1
#>  8 All         0.689  Experimental   28.2         457.  28.9      1
#>  9 All         1.09   Experimental    6.82         69.4  7.91     1
#> 10 All         1.20   Control        13.8         885.  15.1      1
#> 11 All         1.27   Control         6.18       1806.   7.45     1
#> 12 All         1.45   Experimental   26.3          74.0 27.8      1
#> 13 All         1.57   Control        43.6        4634.  45.2      1
#> 14 All         1.64   Experimental    0.723       523.   2.36     1
#> 15 All         1.72   Control         7.85        528.   9.56     1
#> 16 All         1.85   Experimental    9.78        652.  11.6      1
#> 17 All         1.90   Control         3.30       1260.   5.21     1
#> 18 All         1.92   Experimental   17.9        3190.  19.8      1
#> 19 All         2.01   Control        26.1         234.  28.1      1
#> 20 All         2.01   Experimental    1.50        573.   3.52     1
 # 3:1 randomization
 simtrial:::simPWSurv_(n=20,block=c(rep("Experimental",3),"Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All         0.0327 Experimental   61.4         531.  61.5      1
#>  2 All         0.0801 Control        22.4         129.  22.5      1
#>  3 All         0.251  Experimental   15.4         197.  15.7      1
#>  4 All         0.329  Experimental   17.2         528.  17.6      1
#>  5 All         0.561  Control        12.0         222.  12.6      1
#>  6 All         0.720  Experimental   69.0         786.  69.8      1
#>  7 All         0.836  Experimental   13.7          98.0 14.5      1
#>  8 All         1.16   Experimental    0.682       363.   1.84     1
#>  9 All         1.18   Control        13.2         261.  14.4      1
#> 10 All         1.19   Experimental    1.69        455.   2.88     1
#> 11 All         1.27   Experimental   34.2         988.  35.4      1
#> 12 All         1.38   Experimental   12.8         730.  14.2      1
#> 13 All         1.69   Experimental    5.66       1673.   7.35     1
#> 14 All         1.81   Experimental   10.8          22.6 12.6      1
#> 15 All         1.82   Experimental   15.4          58.2 17.2      1
#> 16 All         1.89   Control        15.4         837.  17.3      1
#> 17 All         1.91   Experimental   14.0         275.  15.9      1
#> 18 All         2.05   Control         0.893        86.0  2.94     1
#> 19 All         2.07   Experimental   37.5         208.  39.6      1
#> 20 All         2.16   Experimental    1.29       1644.   3.45     1

# Simulate 2 strata; will use defaults for blocking and enrollRates
simtrial:::simPWSurv_(n=20,
          # 2 strata,30% and 70% prevalence
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.3,.7)),
          failRates=tibble::tibble(Stratum=c(rep("Low",4),rep("High",4)),
                                   period=rep(1:2,4),
                                   Treatment=rep(c(rep("Control",2),rep("Experimental",2)),2),
                                   duration=rep(c(3,1),4),
                                   rate=c(.03,.05,.03,.03,.05,.08,.07,.04)),
          dropoutRates=tibble::tibble(Stratum=c(rep("Low",2),rep("High",2)),
                                      period=rep(1,4),
                                      Treatment=rep(c("Control","Experimental"),2),
                                      duration=rep(1,4),
                                      rate=rep(.001,4))
)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low          0.105 Control        24.3        1880.  24.4      1
#>  2 High         0.192 Experimental   25.1          75.4 25.3      1
#>  3 High         0.240 Control         3.60        424.   3.84     1
#>  4 Low          0.277 Experimental   33.8        1609.  34.0      1
#>  5 High         0.744 Experimental    6.15        423.   6.89     1
#>  6 High         0.759 Control        15.7        1925.  16.4      1
#>  7 Low          0.829 Control         9.09       2912.   9.92     1
#>  8 High         0.941 Experimental   33.2         721.  34.1      1
#>  9 Low          1.03  Experimental    0.363       697.   1.39     1
#> 10 High         1.04  Experimental   27.1         957.  28.1      1
#> 11 High         1.64  Control         4.58       1204.   6.22     1
#> 12 High         1.75  Control         0.237       720.   1.98     1
#> 13 High         1.79  Control        14.2         261.  16.0      1
#> 14 Low          1.94  Control        26.7          50.3 28.7      1
#> 15 High         2.06  Experimental   30.7         480.  32.7      1
#> 16 High         2.13  Experimental   12.3         792.  14.4      1
#> 17 High         2.25  Control         9.62        595.  11.9      1
#> 18 High         2.37  Control         5.72       1797.   8.09     1
#> 19 High         2.51  Experimental   34.6        3408.  37.1      1
#> 20 High         2.59  Experimental  123.           10.4 13.0      0

# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.03),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.03),
   tibble(Stratum="Low" ,period=2,Treatment="Experimental",duration=3,rate=.02),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.05),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.06),
   tibble(Stratum="High",period=2,Treatment="Experimental",duration=3,rate=.03)
)
dropoutRates <- bind_rows(
   tibble(Stratum="Low" ,period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="Low" ,period=1,Treatment="Experimental",duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Control"     ,duration=3,rate=.001),
   tibble(Stratum="High",period=1,Treatment="Experimental",duration=3,rate=.001)
)
simtrial:::simPWSurv_(n=12,strata=tibble(Stratum=c("Low","High"),p=c(.3,.7)),
         failRates=failRates,dropoutRates=dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enrollTime Treatment    failTime dropoutTime   cte  fail
#>    <chr>        <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low         0.0289 Experimental    138.         809. 138.      1
#>  2 High        0.0353 Control          28.8       3167.  28.8     1
#>  3 Low         0.124  Control          13.3       2010.  13.4     1
#>  4 High        0.470  Control          43.9        565.  44.4     1
#>  5 Low         0.549  Control          16.4       1264.  16.9     1
#>  6 High        0.745  Experimental     10.9       2616.  11.6     1
#>  7 High        0.766  Experimental     21.4        718.  22.1     1
#>  8 High        1.09   Control          21.6        125.  22.7     1
#>  9 Low         1.14   Experimental     21.1       2456.  22.2     1
#> 10 High        1.58   Control          12.8       3108.  14.4     1
#> 11 High        1.60   Experimental     82.6        603.  84.2     1
#> 12 High        1.68   Experimental     36.5        505.  38.2     1