Simulation of fixed sample size design for time-to-event endpoint

sim_fixed_n() provides simulations of a single endpoint two-arm trial where the enrollment, hazard ratio, and failure and dropout rates change over time.

Usage

sim_fixed_n(
  n_sim = 1000,
  sample_size = 500,
  target_event = 350,
  stratum = data.frame(stratum = "All", p = 1),
  enroll_rate = data.frame(duration = c(2, 2, 10), rate = c(3, 6, 9)),
  fail_rate = data.frame(stratum = "All", duration = c(3, 100), fail_rate = log(2)/c(9,
    18), hr = c(0.9, 0.6), dropout_rate = rep(0.001, 2)),
  total_duration = 30,
  block = rep(c("experimental", "control"), 2),
  timing_type = 1:5,
  rho_gamma = data.frame(rho = 0, gamma = 0)
)

Arguments

n_sim

Number of simulations to perform.

sample_size

Total sample size per simulation.

target_event

Targeted event count for analysis.

stratum

A data frame with stratum specified in stratum, probability (incidence) of each stratum in p.

enroll_rate

Piecewise constant enrollment rates by time period. Note that these are overall population enrollment rates and the stratum argument controls the random distribution between stratum.

fail_rate

Piecewise constant control group failure rates, hazard ratio for experimental vs. control, and dropout rates by stratum and time period.

total_duration

Total follow-up from start of enrollment to data cutoff.

block

As in sim_pw_surv(). Vector of treatments to be included in each block.

timing_type

A numeric vector determining data cutoffs used; see details. Default is to include all available cutoff methods.

rho_gamma

As in fh_weight(). A data frame with variables rho and gamma, both greater than equal to zero, to specify one Fleming-Harrington weighted logrank test per row.

Value

A data frame including columns:

event: Event count.

ln_hr: Log-hazard ratio.

z: Normal test statistic; < 0 favors experimental.

cut: Text describing cutoff used.

duration: Duration of trial at cutoff for analysis.

sim: Sequential simulation ID.

One row per simulated dataset per cutoff specified in timing_type, per test statistic specified. If multiple Fleming-Harrington tests are specified in rho_gamma, then columns rho and gamma are also included.

Details

timing_type has up to 5 elements indicating different options for data cutoff:

1: Uses the planned study duration.

2: The time the targeted event count is achieved.

3: The planned minimum follow-up after enrollment is complete.

4: The maximum of planned study duration and targeted event count cuts (1 and 2).

5: The maximum of targeted event count and minimum follow-up cuts (2 and 3).

Examples

library(dplyr)
library(future)

# Example 1
# Show output structure
sim_fixed_n(n_sim = 3)
#> Backend uses sequential processing.
#>    event       ln_hr          z                              cut duration sim
#> 1     80 -0.25977214 -1.1581124                 Planned duration 30.00000   1
#> 2    350 -0.34607737 -3.2204702                  Targeted events 73.43979   1
#> 3    358 -0.35204853 -3.3133948                Minimum follow-up 75.55607   1
#> 4    350 -0.34607737 -3.2204702 Max(planned duration, event cut) 73.43979   1
#> 5    358 -0.35204853 -3.3133948    Max(min follow-up, event cut) 75.55607   1
#> 6    122 -0.40777904 -2.2269506                 Planned duration 30.00000   2
#> 7    350 -0.45626994 -4.2131329                  Targeted events 66.15573   2
#> 8    371 -0.47244478 -4.4890293                Minimum follow-up 71.26657   2
#> 9    350 -0.45626994 -4.2131329 Max(planned duration, event cut) 66.15573   2
#> 10   371 -0.47244478 -4.4890293    Max(min follow-up, event cut) 71.26657   2
#> 11    92 -0.03928754 -0.1882937                 Planned duration 30.00000   3
#> 12   350 -0.26817412 -2.4941988                  Targeted events 74.17321   3
#> 13   345 -0.24830301 -2.2938155                Minimum follow-up 72.94623   3
#> 14   350 -0.26817412 -2.4941988 Max(planned duration, event cut) 74.17321   3
#> 15   350 -0.26817412 -2.4941988    Max(min follow-up, event cut) 74.17321   3

# Example 2
# Example with 2 tests: logrank and FH(0,1)
sim_fixed_n(n_sim = 1, rho_gamma = data.frame(rho = 0, gamma = c(0, 1)))
#> Backend uses sequential processing.
#>    event      ln_hr rho gamma          z        v1        v2
#> 1    112 -0.1139817   0     0 -0.6023687 1.0000000 0.8353650
#> 2    112 -0.1139817   0     1 -0.2854116 0.8353650 1.0000000
#> 3    350 -0.3450155   0     0 -3.2027797 1.0000000 0.8524985
#> 4    350 -0.3450155   0     1 -3.8242328 0.8524985 1.0000000
#> 5    367 -0.3557917   0     0 -3.3856754 1.0000000 0.8546851
#> 6    367 -0.3557917   0     1 -4.0102954 0.8546851 1.0000000
#> 7    350 -0.3450155   0     0 -3.2027797 1.0000000 0.8524985
#> 8    350 -0.3450155   0     1 -3.8242328 0.8524985 1.0000000
#> 9    367 -0.3557917   0     0 -3.3856754 1.0000000 0.8546851
#> 10   367 -0.3557917   0     1 -4.0102954 0.8546851 1.0000000
#>                                 cut duration sim
#> 1                  Planned duration 30.00000   1
#> 2                  Planned duration 30.00000   1
#> 3                   Targeted events 69.21747   1
#> 4                   Targeted events 69.21747   1
#> 5                 Minimum follow-up 73.47841   1
#> 6                 Minimum follow-up 73.47841   1
#> 7  Max(planned duration, event cut) 69.21747   1
#> 8  Max(planned duration, event cut) 69.21747   1
#> 9     Max(min follow-up, event cut) 73.47841   1
#> 10    Max(min follow-up, event cut) 73.47841   1

# \donttest{
# Example 3
# Power by test
# Only use cuts for events, events + min follow-up
xx <- sim_fixed_n(
  n_sim = 100,
  timing_type = c(2, 5),
  rho_gamma = data.frame(rho = 0, gamma = c(0, 1))
)
#> Backend uses sequential processing.
# Get power approximation for FH, data cutoff combination
xx |>
  group_by(cut, rho, gamma) |>
  summarize(mean(z <= qnorm(.025)))
#> `summarise()` has grouped output by 'cut', 'rho'. You can override using the
#> `.groups` argument.
#> # A tibble: 4 × 4
#> # Groups:   cut, rho [2]
#>   cut                             rho gamma `mean(z <= qnorm(0.025))`
#>   <chr>                         <dbl> <dbl>                     <dbl>
#> 1 Max(min follow-up, event cut)     0     0                      0.96
#> 2 Max(min follow-up, event cut)     0     1                      0.99
#> 3 Targeted events                   0     0                      0.94
#> 4 Targeted events                   0     1                      0.96

# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx |>
  filter(cut != "Targeted events") |>
  group_by(sim) |>
  group_map(~ pvalue_maxcombo(.x)) |>
  unlist()

mean(p < .025)
#> [1] 0.97

# MaxCombo estimate for targeted events cutoff
p <- xx |>
  filter(cut == "Targeted events") |>
  group_by(sim) |>
  group_map(~ pvalue_maxcombo(.x)) |>
  unlist()

mean(p < .025)
#> [1] 0.97

# Example 4
# Use two cores
set.seed(2023)
plan("multisession", workers = 2)
sim_fixed_n(n_sim = 10)
#> Using 2 cores with backend multisession
#>    event      ln_hr          z                              cut duration sim
#> 1    109 -0.3342849 -1.7348832                 Planned duration 30.00000   1
#> 2    350 -0.4793575 -4.4607622                  Targeted events 65.23187   1
#> 3    376 -0.4498374 -4.3411518                Minimum follow-up 73.17268   1
#> 4    350 -0.4793575 -4.4607622 Max(planned duration, event cut) 65.23187   1
#> 5    376 -0.4498374 -4.3411518    Max(min follow-up, event cut) 73.17268   1
#> 6    103 -0.2047494 -1.0356297                 Planned duration 30.00000   2
#> 7    350 -0.3873648 -3.5981660                  Targeted events 67.64612   2
#> 8    367 -0.3953215 -3.7665623                Minimum follow-up 72.56793   2
#> 9    350 -0.3873648 -3.5981660 Max(planned duration, event cut) 67.64612   2
#> 10   367 -0.3953215 -3.7665623    Max(min follow-up, event cut) 72.56793   2
#> 11    95 -0.8181040 -3.8898935                 Planned duration 30.00000   3
#> 12   350 -0.6662357 -6.1228581                  Targeted events 74.95821   3
#> 13   349 -0.6719858 -6.1652657                Minimum follow-up 74.72225   3
#> 14   350 -0.6662357 -6.1228581 Max(planned duration, event cut) 74.95821   3
#> 15   350 -0.6662357 -6.1228581    Max(min follow-up, event cut) 74.95821   3
#> 16   110 -0.1120785 -0.5861109                 Planned duration 30.00000   4
#> 17   350 -0.3939024 -3.6658988                  Targeted events 69.16631   4
#> 18   372 -0.3781111 -3.6289642                Minimum follow-up 73.68989   4
#> 19   350 -0.3939024 -3.6658988 Max(planned duration, event cut) 69.16631   4
#> 20   372 -0.3781111 -3.6289642    Max(min follow-up, event cut) 73.68989   4
#> 21    94 -0.6069906 -2.8943845                 Planned duration 30.00000   5
#> 22   350 -0.4410650 -4.1082391                  Targeted events 77.05035   5
#> 23   348 -0.4296748 -3.9894083                Minimum follow-up 76.68470   5
#> 24   350 -0.4410650 -4.1082391 Max(planned duration, event cut) 77.05035   5
#> 25   350 -0.4410650 -4.1082391    Max(min follow-up, event cut) 77.05035   5
#> 26   103 -0.5638412 -2.8176983                 Planned duration 30.00000   6
#> 27   350 -0.4227156 -3.9027737                  Targeted events 71.51268   6
#> 28   351 -0.4245379 -3.9272445                Minimum follow-up 72.34254   6
#> 29   350 -0.4227156 -3.9027737 Max(planned duration, event cut) 71.51268   6
#> 30   351 -0.4245379 -3.9272445    Max(min follow-up, event cut) 72.34254   6
#> 31   108 -0.1666295 -0.8633360                 Planned duration 30.00000   7
#> 32   350 -0.3453421 -3.2209226                  Targeted events 75.61250   7
#> 33   338 -0.3305844 -3.0299119                Minimum follow-up 72.99192   7
#> 34   350 -0.3453421 -3.2209226 Max(planned duration, event cut) 75.61250   7
#> 35   350 -0.3453421 -3.2209226    Max(min follow-up, event cut) 75.61250   7
#> 36   124 -0.2803523 -1.5523243                 Planned duration 30.00000   8
#> 37   350 -0.3478454 -3.2482694                  Targeted events 66.62160   8
#> 38   367 -0.3690357 -3.5288351                Minimum follow-up 70.19488   8
#> 39   350 -0.3478454 -3.2482694 Max(planned duration, event cut) 66.62160   8
#> 40   367 -0.3690357 -3.5288351    Max(min follow-up, event cut) 70.19488   8
#> 41   104 -0.1525416 -0.7765436                 Planned duration 30.00000   9
#> 42   350 -0.3069406 -2.8670291                  Targeted events 70.92746   9
#> 43   351 -0.3023229 -2.8280363                Minimum follow-up 71.28078   9
#> 44   350 -0.3069406 -2.8670291 Max(planned duration, event cut) 70.92746   9
#> 45   351 -0.3023229 -2.8280363    Max(min follow-up, event cut) 71.28078   9
#> 46   106 -0.3551148 -1.8105400                 Planned duration 30.00000  10
#> 47   350 -0.4077490 -3.8029263                  Targeted events 75.17856  10
#> 48   360 -0.4138911 -3.9105095                Minimum follow-up 76.94235  10
#> 49   350 -0.4077490 -3.8029263 Max(planned duration, event cut) 75.17856  10
#> 50   360 -0.4138911 -3.9105095    Max(min follow-up, event cut) 76.94235  10
plan("sequential")
# }