Simulate group sequential designs with fixed sample size

WARNING: This experimental function is a work-in-progress. The function arguments will change as we add additional features.

Usage

sim_gs_n(
  n_sim = 1000,
  sample_size = 500,
  stratum = data.frame(stratum = "All", p = 1),
  enroll_rate = data.frame(duration = c(2, 2, 10), rate = c(3, 6, 9)),
  fail_rate = data.frame(stratum = "All", duration = c(3, 100), fail_rate = log(2)/c(9,
    18), hr = c(0.9, 0.6), dropout_rate = rep(0.001, 2)),
  block = rep(c("experimental", "control"), 2),
  test = wlr,
  cutting = NULL,
  seed = 2024,
  ...
)

Arguments

n_sim

Number of simulations to perform.

sample_size

Total sample size per simulation.

stratum

A data frame with stratum specified in stratum, probability (incidence) of each stratum in p.

enroll_rate

Piecewise constant enrollment rates by time period. Note that these are overall population enrollment rates and the stratum argument controls the random distribution between stratum.

fail_rate

Piecewise constant control group failure rates, hazard ratio for experimental vs. control, and dropout rates by stratum and time period.

block

As in sim_pw_surv(). Vector of treatments to be included in each block.

test

A test function such as wlr(), maxcombo(), or rmst(). The simulated data set is passed as the first positional argument to the test function provided.

cutting

A list of cutting functions created by create_cutting(), see examples.

seed

Random seed.

...

Arguments passed to the test function provided by the argument test.

Value

A data frame summarizing the simulation ID, analysis date, z statistics or p-values.

Examples

library(gsDesign2)

# Parameters for enrollment
enroll_rampup_duration <- 4 # Duration for enrollment ramp up
enroll_duration <- 16 # Total enrollment duration
enroll_rate <- define_enroll_rate(
  duration = c(
    enroll_rampup_duration,
    enroll_duration - enroll_rampup_duration
  ),
  rate = c(10, 30)
)

# Parameters for treatment effect
delay_effect_duration <- 3 # Delay treatment effect in months
median_ctrl <- 9 # Survival median of the control arm
median_exp <- c(9, 14) # Survival median of the experimental arm
dropout_rate <- 0.001
fail_rate <- define_fail_rate(
  duration = c(delay_effect_duration, 100),
  fail_rate = log(2) / median_ctrl,
  hr = median_ctrl / median_exp,
  dropout_rate = dropout_rate
)

# Other related parameters
alpha <- 0.025 # Type I error
beta <- 0.1 # Type II error
ratio <- 1 # Randomization ratio (experimental:control)

# Define cuttings of 2 IAs and 1 FA
# IA1
# The 1st interim analysis will occur at the later of the following 3 conditions:
# - At least 20 months have passed since the start of the study.
# - At least 100 events have occurred.
# - At least 20 months have elapsed after enrolling 200/400 subjects, with a
#   minimum of 20 months follow-up.
# However, if events accumulation is slow, we will wait for a maximum of 24 months.
ia1 <- create_cutting(
  planned_calendar_time = 20,
  target_event_overall = 100,
  max_extension_for_target_event = 24,
  min_n_overall = 200,
  min_followup = 20
)

# IA2
# The 2nd interim analysis will occur at the later of the following 3 conditions:
# - At least 32 months have passed since the start of the study.
# - At least 250 events have occurred.
# - At least 10 months after IA1.
# However, if events accumulation is slow, we will wait for a maximum of 34 months.
ia2 <- create_cutting(
  planned_calendar_time = 32,
  target_event_overall = 200,
  max_extension_for_target_event = 34,
  min_time_after_previous_analysis = 10
)

# FA
# The final analysis will occur at the later of the following 2 conditions:
# - At least 45 months have passed since the start of the study.
# - At least 300 events have occurred.
fa <- create_cutting(
  planned_calendar_time = 45,
  target_event_overall = 350
)

# Test 1: regular logrank test
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = wlr,
  cutting = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  weight = fh(rho = 0, gamma = 0)
)
#>   rho gamma         z analysis cut_date sim_id   n event
#> 1   0     0 -3.748605        1 24.00000      1 400   229
#> 2   0     0 -4.530340        2 32.00000      1 400   295
#> 3   0     0 -4.316453        3 45.00000      1 400   355
#> 4   0     0 -3.477144        1 24.00000      2 400   241
#> 5   0     0 -3.863150        2 32.00000      2 400   290
#> 6   0     0 -3.277778        3 46.21933      2 400   350
#> 7   0     0 -3.075863        1 24.00000      3 400   226
#> 8   0     0 -3.619345        2 32.00000      3 400   282
#> 9   0     0 -4.222592        3 50.86585      3 400   350

# Test 2: weighted logrank test by FH(0, 0.5)
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = wlr,
  cutting = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  weight = fh(rho = 0, gamma = 0.5)
)
#>   rho gamma         z analysis cut_date sim_id   n event
#> 1   0   0.5 -4.149161        1 24.00000      1 400   229
#> 2   0   0.5 -4.778108        2 32.00000      1 400   295
#> 3   0   0.5 -4.260730        3 45.00000      1 400   355
#> 4   0   0.5 -3.605093        1 24.00000      2 400   241
#> 5   0   0.5 -3.945081        2 32.00000      2 400   290
#> 6   0   0.5 -2.919180        3 46.21933      2 400   350
#> 7   0   0.5 -3.143228        1 24.00000      3 400   226
#> 8   0   0.5 -3.640459        2 32.00000      3 400   282
#> 9   0   0.5 -4.243289        3 50.86585      3 400   350

# Test 3: weighted logrank test by MB(3)
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = wlr,
  cutting = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  weight = mb(delay = 3)
)
#>           z analysis cut_date sim_id   n event
#> 1 -3.797134        1 24.00000      1 400   229
#> 2 -4.581331        2 32.00000      1 400   295
#> 3 -4.349644        3 45.00000      1 400   355
#> 4 -3.501131        1 24.00000      2 400   241
#> 5 -3.886542        2 32.00000      2 400   290
#> 6 -3.279286        3 46.21933      2 400   350
#> 7 -3.114079        1 24.00000      3 400   226
#> 8 -3.658715        2 32.00000      3 400   282
#> 9 -4.263279        3 50.86585      3 400   350

# Test 4: weighted logrank test by early zero (6)
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = wlr,
  cutting = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  weight = early_zero(6)
)
#>           z analysis cut_date sim_id   n event
#> 1 -4.552617        1 24.00000      1 400   229
#> 2 -5.188573        2 32.00000      1 400   295
#> 3 -4.686074        3 45.00000      1 400   355
#> 4 -3.185533        1 24.00000      2 400   241
#> 5 -3.597503        2 32.00000      2 400   290
#> 6 -2.786930        3 46.21933      2 400   350
#> 7 -2.367344        1 24.00000      3 400   226
#> 8 -3.063054        2 32.00000      3 400   282
#> 9 -3.781619        3 50.86585      3 400   350

# Test 5: RMST
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = rmst,
  cutting = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  tau = 20
)
#>   rmst_arm1 rmst_arm0 rmst_diff        z analysis cut_date sim_id   n event
#> 1  12.46626  9.585108  2.881152 3.789982        1 24.00000      1 400   229
#> 2  12.44420  9.591074  2.853131 3.991863        2 32.00000      1 400   295
#> 3  12.42510  9.590593  2.834508 3.980101        3 45.00000      1 400   355
#> 4  12.39211  9.824722  2.567390 3.474869        1 24.00000      2 400   241
#> 5  12.49696 10.097271  2.399692 3.295021        2 32.00000      2 400   290
#> 6  12.47912 10.110784  2.368335 3.254115        3 46.21933      2 400   350
#> 7  12.62769 10.340196  2.287498 2.980534        1 24.00000      3 400   226
#> 8  12.73792 10.289798  2.448117 3.300952        2 32.00000      3 400   282
#> 9  12.74024 10.261300  2.478942 3.350430        3 50.86585      3 400   350

# Test 6: Milestone
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = milestone,
  cutting = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  ms_time = 10
)
#>      method         z ms_time surv_ctrl  surv_exp surv_diff std_err_ctrl
#> 1 milestone  9.252619      10 0.4080041 0.5689750 0.1609709   0.03693588
#> 2 milestone 12.078381      10 0.4097269 0.5850000 0.1752731   0.03495270
#> 3 milestone 12.078381      10 0.4097269 0.5850000 0.1752731   0.03495270
#> 4 milestone  5.565741      10 0.4718269 0.5922854 0.1204585   0.03614098
#> 5 milestone  5.457930      10 0.4667007 0.5840901 0.1173894   0.03543263
#> 6 milestone  5.457930      10 0.4667007 0.5840901 0.1173894   0.03543263
#> 7 milestone  9.051773      10 0.4614961 0.6150543 0.1535582   0.03581573
#> 8 milestone  9.054983      10 0.4620000 0.6139773 0.1519773   0.03540131
#> 9 milestone  9.054983      10 0.4620000 0.6139773 0.1519773   0.03540131
#>   std_err_exp analysis cut_date sim_id   n event
#> 1  0.03662190        1 24.00000      1 400   229
#> 2  0.03484071        2 32.00000      1 400   295
#> 3  0.03484071        3 45.00000      1 400   355
#> 4  0.03531267        1 24.00000      2 400   241
#> 5  0.03491216        2 32.00000      2 400   290
#> 6  0.03491216        3 46.21933      2 400   350
#> 7  0.03505127        1 24.00000      3 400   226
#> 8  0.03473824        2 32.00000      3 400   282
#> 9  0.03473824        3 50.86585      3 400   350

# Test 7: MaxCombo (WLR-FH(0,0) + WLR-FH(0, 0.5))
# for all analyses
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = maxcombo,
  cutting = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  rho = c(0, 0),
  gamma = c(0, 0.5)
)
#>        p_value analysis cut_date sim_id   n event
#> 1 2.615539e-05        1 24.00000      1 400   229
#> 2 1.433049e-06        2 32.00000      1 400   295
#> 3 1.247802e-05        3 45.00000      1 400   355
#> 4 2.358380e-04        1 24.00000      2 400   241
#> 5 6.130078e-05        2 32.00000      2 400   290
#> 6 7.667834e-04        3 46.21933      2 400   350
#> 7 1.216230e-03        1 24.00000      3 400   226
#> 8 2.047186e-04        2 32.00000      3 400   282
#> 9 1.724936e-05        3 50.86585      3 400   350

# Test 8: MaxCombo (WLR-FH(0,0.5) + milestone(10))
# for all analyses
if (FALSE) {
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = maxcombo(test1 = wlr, test2 = milestone),
  cutting = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  test1_par = list(weight = fh(rho = 0, gamma = 0.5)),
  test2_par = list(ms_time = 10)
)
}

# Test 9: MaxCombo (WLR-FH(0,0) at IAs
# and WLR-FH(0,0) + milestone(10) + WLR-MB(4,2) at FA)
if (FALSE) {
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = list(ia1 = wlr, ia2 = wlr, fa = maxcombo),
  cutting = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  test_par = list(
    ia1 = list(weight = fh(rho = 0, gamma = 0)),
    ia2 = list(weight = fh(rho = 0, gamma = 0)),
    ia3 = list(
      test1_par = list(weight = fh(rho = 0, gamma = 0)),
      test2_par = list(ms_time = 10),
      test3_par = list(delay = 4, w_max = 2)
    )
  )
)
}