Simulate group sequential designs with fixed sample size

WARNING: This experimental function is a work-in-progress. The function arguments will change as we add additional features.

Usage

sim_gs_n(
  n_sim = 1000,
  sample_size = 500,
  stratum = data.frame(stratum = "All", p = 1),
  enroll_rate = data.frame(duration = c(2, 2, 10), rate = c(3, 6, 9)),
  fail_rate = data.frame(stratum = "All", duration = c(3, 100), fail_rate = log(2)/c(9,
    18), hr = c(0.9, 0.6), dropout_rate = rep(0.001, 2)),
  block = rep(c("experimental", "control"), 2),
  test = wlr,
  cut = NULL,
  seed = 2024,
  ...
)

Arguments

n_sim

Number of simulations to perform.

sample_size

Total sample size per simulation.

stratum

A data frame with stratum specified in stratum, probability (incidence) of each stratum in p.

enroll_rate

Piecewise constant enrollment rates by time period. Note that these are overall population enrollment rates and the stratum argument controls the random distribution between stratum.

fail_rate

Piecewise constant control group failure rates, hazard ratio for experimental vs. control, and dropout rates by stratum and time period.

block

As in sim_pw_surv(). Vector of treatments to be included in each block.

test

One or more test functions such as wlr(), maxcombo(), or rmst(). If a single test function is provided, it will be applied at each cut. Alternatively a list of functions created by create_test(). The list form is experimental and currently limited. It only accepts one test per cutting (in the future multiple tests may be accepted), and all the tests must consistently return the same exact results (again this may be more flexible in the future). Importantly, note that the simulated data set is always passed as the first positional argument to each test function provided.

cut

A list of cutting functions created by create_cut(), see examples.

seed

Random seed.

...

Arguments passed to the test function(s) provided by the argument test.

Value

A data frame summarizing the simulation ID, analysis date, z statistics or p-values.

Examples

library(gsDesign2)

# Parameters for enrollment
enroll_rampup_duration <- 4 # Duration for enrollment ramp up
enroll_duration <- 16 # Total enrollment duration
enroll_rate <- define_enroll_rate(
  duration = c(
    enroll_rampup_duration,
    enroll_duration - enroll_rampup_duration
  ),
  rate = c(10, 30)
)

# Parameters for treatment effect
delay_effect_duration <- 3 # Delay treatment effect in months
median_ctrl <- 9 # Survival median of the control arm
median_exp <- c(9, 14) # Survival median of the experimental arm
dropout_rate <- 0.001
fail_rate <- define_fail_rate(
  duration = c(delay_effect_duration, 100),
  fail_rate = log(2) / median_ctrl,
  hr = median_ctrl / median_exp,
  dropout_rate = dropout_rate
)

# Other related parameters
alpha <- 0.025 # Type I error
beta <- 0.1 # Type II error
ratio <- 1 # Randomization ratio (experimental:control)

# Define cuttings of 2 IAs and 1 FA
# IA1
# The 1st interim analysis will occur at the later of the following 3 conditions:
# - At least 20 months have passed since the start of the study.
# - At least 100 events have occurred.
# - At least 20 months have elapsed after enrolling 200/400 subjects, with a
#   minimum of 20 months follow-up.
# However, if events accumulation is slow, we will wait for a maximum of 24 months.
ia1 <- create_cut(
  planned_calendar_time = 20,
  target_event_overall = 100,
  max_extension_for_target_event = 24,
  min_n_overall = 200,
  min_followup = 20
)

# IA2
# The 2nd interim analysis will occur at the later of the following 3 conditions:
# - At least 32 months have passed since the start of the study.
# - At least 250 events have occurred.
# - At least 10 months after IA1.
# However, if events accumulation is slow, we will wait for a maximum of 34 months.
ia2 <- create_cut(
  planned_calendar_time = 32,
  target_event_overall = 200,
  max_extension_for_target_event = 34,
  min_time_after_previous_analysis = 10
)

# FA
# The final analysis will occur at the later of the following 2 conditions:
# - At least 45 months have passed since the start of the study.
# - At least 300 events have occurred.
fa <- create_cut(
  planned_calendar_time = 45,
  target_event_overall = 350
)

# Test 1: regular logrank test
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = wlr,
  cut = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  weight = fh(rho = 0, gamma = 0)
)
#>              method parameter            estimation se       z         analysis
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0)" -28.19483  7.521418 -3.748605 1       
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0)" -38.32581  8.459808 -4.53034  2       
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0)" -39.4923   9.149248 -4.316453 3       
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0)" -26.84871  7.721484 -3.477144 1       
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0)" -32.54824  8.42531  -3.86315  2       
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0)" -30.06631  9.172772 -3.277778 3       
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0)" -23.06302  7.498065 -3.075863 1       
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0)" -30.1633   8.33391  -3.619345 2       
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0)" -38.75506  9.178027 -4.222592 3       
#>              cut_date sim_id n   event
#> ans_1sim_new 24       1      400 229  
#> ans_1sim_new 32       1      400 295  
#> ans_1sim_new 45       1      400 355  
#> ans_1sim_new 24       2      400 241  
#> ans_1sim_new 32       2      400 290  
#> ans_1sim_new 46.21933 2      400 350  
#> ans_1sim_new 24       3      400 226  
#> ans_1sim_new 32       3      400 282  
#> ans_1sim_new 50.86585 3      400 350  

# Test 2: weighted logrank test by FH(0, 0.5)
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = wlr,
  cut = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  weight = fh(rho = 0, gamma = 0.5)
)
#>              method parameter              estimation se       z        
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0.5)" -16.93422  4.081359 -4.149161
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0.5)" -24.44818  5.116707 -4.778108
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0.5)" -25.51076  5.987416 -4.26073 
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0.5)" -15.50024  4.299539 -3.605093
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0.5)" -19.96769  5.061414 -3.945081
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0.5)" -17.53906  6.008214 -2.91918 
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0.5)" -12.66462  4.029178 -3.143228
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0.5)" -18.05051  4.958307 -3.640459
#> ans_1sim_new "WLR"  "FH(rho=0, gamma=0.5)" -25.59217  6.031211 -4.243289
#>              analysis cut_date sim_id n   event
#> ans_1sim_new 1        24       1      400 229  
#> ans_1sim_new 2        32       1      400 295  
#> ans_1sim_new 3        45       1      400 355  
#> ans_1sim_new 1        24       2      400 241  
#> ans_1sim_new 2        32       2      400 290  
#> ans_1sim_new 3        46.21933 2      400 350  
#> ans_1sim_new 1        24       3      400 226  
#> ans_1sim_new 2        32       3      400 282  
#> ans_1sim_new 3        50.86585 3      400 350  

# Test 3: weighted logrank test by MB(3)
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = wlr,
  cut = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  weight = mb(delay = 3)
)
#>              method parameter                         estimate  se      
#> ans_1sim_new "WLR"  "MB(delay = 3, max_weight = Inf)" -34.19243 9.004801
#> ans_1sim_new "WLR"  "MB(delay = 3, max_weight = Inf)" -46.74548 10.20347
#> ans_1sim_new "WLR"  "MB(delay = 3, max_weight = Inf)" -48.19085 11.07926
#> ans_1sim_new "WLR"  "MB(delay = 3, max_weight = Inf)" -32.19277 9.194961
#> ans_1sim_new "WLR"  "MB(delay = 3, max_weight = Inf)" -39.18612 10.08251
#> ans_1sim_new "WLR"  "MB(delay = 3, max_weight = Inf)" -36.14078 11.02093
#> ans_1sim_new "WLR"  "MB(delay = 3, max_weight = Inf)" -27.89564 8.957908
#> ans_1sim_new "WLR"  "MB(delay = 3, max_weight = Inf)" -36.66946 10.0225 
#> ans_1sim_new "WLR"  "MB(delay = 3, max_weight = Inf)" -47.28631 11.09153
#>              z         analysis cut_date sim_id n   event
#> ans_1sim_new -3.797134 1        24       1      400 229  
#> ans_1sim_new -4.581331 2        32       1      400 295  
#> ans_1sim_new -4.349644 3        45       1      400 355  
#> ans_1sim_new -3.501131 1        24       2      400 241  
#> ans_1sim_new -3.886542 2        32       2      400 290  
#> ans_1sim_new -3.279286 3        46.21933 2      400 350  
#> ans_1sim_new -3.114079 1        24       3      400 226  
#> ans_1sim_new -3.658715 2        32       3      400 282  
#> ans_1sim_new -4.263279 3        50.86585 3      400 350  

# Test 4: weighted logrank test by early zero (6)
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = wlr,
  cut = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  weight = early_zero(6)
)
#>              method parameter                                  estimate 
#> ans_1sim_new "WLR"  "Xu 2017 with first 6 months of 0 weights" -21.99899
#> ans_1sim_new "WLR"  "Xu 2017 with first 6 months of 0 weights" -32.12997
#> ans_1sim_new "WLR"  "Xu 2017 with first 6 months of 0 weights" -33.29646
#> ans_1sim_new "WLR"  "Xu 2017 with first 6 months of 0 weights" -17.19941
#> ans_1sim_new "WLR"  "Xu 2017 with first 6 months of 0 weights" -22.89893
#> ans_1sim_new "WLR"  "Xu 2017 with first 6 months of 0 weights" -20.41701
#> ans_1sim_new "WLR"  "Xu 2017 with first 6 months of 0 weights" -11.77606
#> ans_1sim_new "WLR"  "Xu 2017 with first 6 months of 0 weights" -18.87634
#> ans_1sim_new "WLR"  "Xu 2017 with first 6 months of 0 weights" -27.4681 
#>              se       z         analysis cut_date sim_id n   event
#> ans_1sim_new 4.832164 -4.552617 1        24       1      400 229  
#> ans_1sim_new 6.192449 -5.188573 2        32       1      400 295  
#> ans_1sim_new 7.105407 -4.686074 3        45       1      400 355  
#> ans_1sim_new 5.399223 -3.185533 1        24       2      400 241  
#> ans_1sim_new 6.36523  -3.597503 2        32       2      400 290  
#> ans_1sim_new 7.325985 -2.78693  3        46.21933 2      400 350  
#> ans_1sim_new 4.974376 -2.367344 1        24       3      400 226  
#> ans_1sim_new 6.162588 -3.063054 2        32       3      400 282  
#> ans_1sim_new 7.263581 -3.781619 3        50.86585 3      400 350  

# Test 5: RMST
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = rmst,
  cut = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  tau = 20
)
#>              method parameter estimation se        z        analysis cut_date
#> ans_1sim_new "RMST" 20        2.881152   0.760202  3.789982 1        24      
#> ans_1sim_new "RMST" 20        2.853131   0.7147367 3.991863 2        32      
#> ans_1sim_new "RMST" 20        2.834508   0.7121699 3.980101 3        45      
#> ans_1sim_new "RMST" 20        2.56739    0.7388451 3.474869 1        24      
#> ans_1sim_new "RMST" 20        2.399692   0.7282783 3.295021 2        32      
#> ans_1sim_new "RMST" 20        2.368335   0.727797  3.254115 3        46.21933
#> ans_1sim_new "RMST" 20        2.287498   0.7674791 2.980534 1        24      
#> ans_1sim_new "RMST" 20        2.448117   0.7416398 3.300952 2        32      
#> ans_1sim_new "RMST" 20        2.478942   0.7398877 3.35043  3        50.86585
#>              sim_id n   event
#> ans_1sim_new 1      400 229  
#> ans_1sim_new 1      400 295  
#> ans_1sim_new 1      400 355  
#> ans_1sim_new 2      400 241  
#> ans_1sim_new 2      400 290  
#> ans_1sim_new 2      400 350  
#> ans_1sim_new 3      400 226  
#> ans_1sim_new 3      400 282  
#> ans_1sim_new 3      400 350  

# Test 6: Milestone
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = milestone,
  cut = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  ms_time = 10
)
#>              method      parameter estimation se         z        analysis
#> ans_1sim_new "milestone" 10        -0.4635636 -0.2151203 2.154904 1       
#> ans_1sim_new "milestone" 10        -0.5093608 -0.2066913 2.464355 2       
#> ans_1sim_new "milestone" 10        -0.5093608 -0.2066913 2.464355 3       
#> ans_1sim_new "milestone" 10        -0.3605499 -0.2158065 1.670709 1       
#> ans_1sim_new "milestone" 10        -0.3487338 -0.2107879 1.65443  2       
#> ans_1sim_new "milestone" 10        -0.3487338 -0.2107879 1.65443  3       
#> ans_1sim_new "milestone" 10        -0.4643428 -0.217612  2.13381  1       
#> ans_1sim_new "milestone" 10        -0.4593313 -0.2152211 2.13423  2       
#> ans_1sim_new "milestone" 10        -0.4593313 -0.2152211 2.13423  3       
#>              cut_date sim_id n   event
#> ans_1sim_new 24       1      400 229  
#> ans_1sim_new 32       1      400 295  
#> ans_1sim_new 45       1      400 355  
#> ans_1sim_new 24       2      400 241  
#> ans_1sim_new 32       2      400 290  
#> ans_1sim_new 46.21933 2      400 350  
#> ans_1sim_new 24       3      400 226  
#> ans_1sim_new 32       3      400 282  
#> ans_1sim_new 50.86585 3      400 350  

# Test 7: MaxCombo (WLR-FH(0,0) + WLR-FH(0, 0.5))
# for all analyses
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = maxcombo,
  cut = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  rho = c(0, 0),
  gamma = c(0, 0.5)
)
#>              method     parameter               z         p_value      analysis
#> ans_1sim_new "MaxCombo" "FH(0, 0) + FH(0, 0.5)" numeric,2 2.615539e-05 1       
#> ans_1sim_new "MaxCombo" "FH(0, 0) + FH(0, 0.5)" numeric,2 1.433049e-06 2       
#> ans_1sim_new "MaxCombo" "FH(0, 0) + FH(0, 0.5)" numeric,2 1.247802e-05 3       
#> ans_1sim_new "MaxCombo" "FH(0, 0) + FH(0, 0.5)" numeric,2 0.000235838  1       
#> ans_1sim_new "MaxCombo" "FH(0, 0) + FH(0, 0.5)" numeric,2 6.130078e-05 2       
#> ans_1sim_new "MaxCombo" "FH(0, 0) + FH(0, 0.5)" numeric,2 0.0007667834 3       
#> ans_1sim_new "MaxCombo" "FH(0, 0) + FH(0, 0.5)" numeric,2 0.00121623   1       
#> ans_1sim_new "MaxCombo" "FH(0, 0) + FH(0, 0.5)" numeric,2 0.0002047186 2       
#> ans_1sim_new "MaxCombo" "FH(0, 0) + FH(0, 0.5)" numeric,2 1.724936e-05 3       
#>              cut_date sim_id n   event
#> ans_1sim_new 24       1      400 229  
#> ans_1sim_new 32       1      400 295  
#> ans_1sim_new 45       1      400 355  
#> ans_1sim_new 24       2      400 241  
#> ans_1sim_new 32       2      400 290  
#> ans_1sim_new 46.21933 2      400 350  
#> ans_1sim_new 24       3      400 226  
#> ans_1sim_new 32       3      400 282  
#> ans_1sim_new 50.86585 3      400 350  

# Test 8: MaxCombo (WLR-FH(0,0.5) + milestone(10))
# for all analyses
if (FALSE) {
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = maxcombo(test1 = wlr, test2 = milestone),
  cut = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  test1_par = list(weight = fh(rho = 0, gamma = 0.5)),
  test2_par = list(ms_time = 10)
)
}

# Test 9: MaxCombo (WLR-FH(0,0) at IAs
# and WLR-FH(0,0) + milestone(10) + WLR-MB(4,2) at FA)
if (FALSE) {
sim_gs_n(
  n_sim = 3,
  sample_size = 400,
  enroll_rate = enroll_rate,
  fail_rate = fail_rate,
  test = list(ia1 = wlr, ia2 = wlr, fa = maxcombo),
  cut = list(ia1 = ia1, ia2 = ia2, fa = fa),
  seed = 2024,
  test_par = list(
    ia1 = list(weight = fh(rho = 0, gamma = 0)),
    ia2 = list(weight = fh(rho = 0, gamma = 0)),
    ia3 = list(
      test1_par = list(weight = fh(rho = 0, gamma = 0)),
      test2_par = list(ms_time = 10),
      test3_par = list(delay = 4, w_max = 2)
    )
  )
)
}