Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv() enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribution and enrollment pattern where the enrollment, failure, and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure, and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p.

block

Vector of treatments to be included in each block.

enroll_rate

Enrollment rates; see details and examples.

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block.

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block.

Value

A tibble with the following variables for each observation:

stratum.

enroll_time: Enrollment time for the observation.

Treatment: Treatment group; this will be one of the values in the input block.

fail_time: Failure time generated using rpwexp().

dropout_time: Dropout time generated using rpwexp().

cte: Calendar time of enrollment plot the minimum of failure time and dropout time.

fail: Indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout.

Examples

library(tibble)
library(dplyr)

# Example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 All          0.0143 experimental   21.3           127.  21.3       1
#>  2 All          0.0512 experimental   49.9           619.  50.0       1
#>  3 All          0.217  control         1.07          652.   1.29      1
#>  4 All          0.221  control         0.0356        526.   0.257     1
#>  5 All          0.236  control        50.7          1490.  50.9       1
#>  6 All          0.312  experimental   23.8          1562.  24.1       1
#>  7 All          0.400  control         0.658        2970.   1.06      1
#>  8 All          0.412  experimental   14.6           362.  15.0       1
#>  9 All          0.490  experimental   43.7          1500.  44.2       1
#> 10 All          0.528  control         7.43           26.2  7.95      1
#> 11 All          0.558  experimental   19.0           204.  19.5       1
#> 12 All          0.648  control         0.235         793.   0.883     1
#> 13 All          0.806  experimental    2.88         3127.   3.68      1
#> 14 All          0.826  control         2.82         3787.   3.65      1
#> 15 All          1.05   control        14.1           325.  15.1       1
#> 16 All          1.12   experimental   70.4           446.  71.5       1
#> 17 All          1.13   experimental   22.0           252.  23.1       1
#> 18 All          1.16   control         3.87           26.0  5.03      1
#> 19 All          1.16   experimental   10.5            69.3 11.7       1
#> 20 All          1.19   control         3.11          994.   4.30      1

# Example 2
# 3:1 randomization
sim_pw_surv(
  n = 20,
  block = c(rep("experimental", 3), "control")
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 All          0.0383 experimental   13.2          1191.  13.3       1
#>  2 All          0.156  experimental    0.0719        566.   0.228     1
#>  3 All          0.196  control         1.04         1261.   1.24      1
#>  4 All          0.421  experimental   12.8           596.  13.2       1
#>  5 All          0.442  experimental   19.2           330.  19.7       1
#>  6 All          0.557  control         9.22          515.   9.78      1
#>  7 All          0.631  experimental   37.2            27.0 27.6       0
#>  8 All          0.670  experimental   30.9          2148.  31.6       1
#>  9 All          0.810  experimental    5.43          661.   6.24      1
#> 10 All          0.876  control         6.89          787.   7.77      1
#> 11 All          0.898  experimental   17.0          4110.  17.9       1
#> 12 All          0.934  experimental   41.7           835.  42.7       1
#> 13 All          1.26   control        27.3           660.  28.6       1
#> 14 All          1.31   experimental   16.2           312.  17.6       1
#> 15 All          1.59   experimental   66.2           529.  67.8       1
#> 16 All          1.95   experimental   14.1            50.0 16.0       1
#> 17 All          2.15   experimental    0.823         218.   2.98      1
#> 18 All          2.18   control         7.66          230.   9.84      1
#> 19 All          2.36   experimental    1.11          360.   3.47      1
#> 20 All          2.50   experimental   19.8           313.  22.3       1

# Example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(
  n = 20,
  # 2 stratum,30% and 70% prevalence
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = tibble(
    stratum = c(rep("Low", 4), rep("High", 4)),
    period = rep(1:2, 4),
    treatment = rep(c(
      rep("control", 2),
      rep("experimental", 2)
    ), 2),
    duration = rep(c(3, 1), 4),
    rate = c(.03, .05, .03, .03, .05, .08, .07, .04)
  ),
  dropout_rate = tibble(
    stratum = c(rep("Low", 2), rep("High", 2)),
    period = rep(1, 4),
    treatment = rep(c("control", "experimental"), 2),
    duration = rep(1, 4),
    rate = rep(.001, 4)
  )
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time     cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>   <dbl> <dbl>
#>  1 High         0.0186 control         0.0538        353.   0.0724     1
#>  2 High         0.802  experimental    2.23          891.   3.03       1
#>  3 High         0.805  control        47.4          3400.  48.2        1
#>  4 High         0.879  experimental    6.86          528.   7.74       1
#>  5 High         0.972  control        21.0            85.5 22.0        1
#>  6 Low          1.25   experimental    5.27         6867.   6.52       1
#>  7 High         1.40   experimental    7.36         1703.   8.75       1
#>  8 High         1.41   experimental    7.15           57.6  8.56       1
#>  9 High         1.43   control         1.04         1785.   2.47       1
#> 10 High         1.46   control        25.6          1229.  27.1        1
#> 11 High         1.47   experimental   31.7           582.  33.1        1
#> 12 High         1.52   control        17.0           599.  18.5        1
#> 13 High         1.67   experimental   34.6           664.  36.3        1
#> 14 Low          1.75   experimental    2.02          647.   3.77       1
#> 15 High         1.92   experimental   40.1           323.  42.0        1
#> 16 Low          1.95   control        39.5           412.  41.5        1
#> 17 High         2.05   control        13.6           777.  15.7        1
#> 18 Low          2.11   control        79.4           756.  81.5        1
#> 19 High         2.14   experimental    8.63           73.9 10.8        1
#> 20 High         2.28   control         3.17          139.   5.44       1
# Example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03)
)

dropout_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .001)
)

sim_pw_surv(
  n = 12,
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = fail_rate,
  dropout_rate = dropout_rate
)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time     cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>   <dbl> <dbl>
#>  1 High         0.0195 control          0.336         55.4   0.356     1
#>  2 High         0.0240 control          1.11        3340.    1.14      1
#>  3 High         0.0766 experimental     9.01         149.    9.08      1
#>  4 High         0.370  experimental    41.4          329.   41.7       1
#>  5 High         0.463  control         15.0          392.   15.5       1
#>  6 High         0.489  experimental    75.5          815.   75.9       1
#>  7 Low          0.512  control         17.0           79.9  17.5       1
#>  8 High         0.519  experimental    30.7          147.   31.3       1
#>  9 High         0.605  control          0.552        423.    1.16      1
#> 10 High         0.926  experimental    25.6         2204.   26.5       1
#> 11 Low          0.999  experimental   123.          4176.  124.        1
#> 12 High         1.34   control         25.4          315.   26.7       1