Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All           0.224 experimental     2.71        602.   2.94     1
#>  2 All           0.311 control         23.4        1855.  23.7      1
#>  3 All           0.320 experimental    19.5         761.  19.8      1
#>  4 All           0.754 control          1.46        547.   2.21     1
#>  5 All           1.00  experimental    89.6         538.  90.6      1
#>  6 All           1.64  control         40.5        1816.  42.1      1
#>  7 All           1.88  control          6.29       1273.   8.17     1
#>  8 All           2.20  experimental     0.828       945.   3.03     1
#>  9 All           2.37  experimental    31.9         244.  34.3      1
#> 10 All           2.44  control         22.5        2577.  25.0      1
#> 11 All           2.59  control         21.2         443.  23.8      1
#> 12 All           2.61  experimental    28.0         496.  30.6      1
#> 13 All           2.68  experimental    10.9          95.9 13.6      1
#> 14 All           2.75  control          7.16        133.   9.91     1
#> 15 All           2.78  experimental    16.8         932.  19.5      1
#> 16 All           2.85  control          7.18        282.  10.0      1
#> 17 All           3.04  experimental    11.8        1818.  14.9      1
#> 18 All           3.33  control         11.0         724.  14.3      1
#> 19 All           3.38  control         22.6          19.1 22.4      0
#> 20 All           3.40  experimental     6.48        546.   9.88     1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("experimental",3), "control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All           0.268 experimental     9.02        595.   9.29     1
#>  2 All           0.378 control          8.33         56.5  8.71     1
#>  3 All           0.416 experimental     6.99        150.   7.40     1
#>  4 All           0.445 experimental     6.70        381.   7.14     1
#>  5 All           0.813 experimental     1.14       2699.   1.95     1
#>  6 All           0.881 control         12.7         589.  13.6      1
#>  7 All           0.901 experimental    34.7        1190.  35.6      1
#>  8 All           1.18  experimental    78.3          67.2 68.4      0
#>  9 All           1.32  experimental     1.59        401.   2.91     1
#> 10 All           1.32  control         47.2        2716.  48.6      1
#> 11 All           1.36  experimental     2.90       2374.   4.26     1
#> 12 All           1.42  experimental    30.1        1827.  31.6      1
#> 13 All           1.45  control          5.01       1278.   6.45     1
#> 14 All           1.90  experimental     5.35       4275.   7.25     1
#> 15 All           1.92  experimental    22.6          90.7 24.5      1
#> 16 All           1.98  experimental     0.702      1438.   2.68     1
#> 17 All           2.01  experimental    14.1        1186.  16.1      1
#> 18 All           2.08  control          0.349       275.   2.43     1
#> 19 All           2.10  experimental    64.7        1439.  66.8      1
#> 20 All           2.39  experimental     5.36        542.   7.75     1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             treatment= rep(c(rep("control", 2),
                                         rep("experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropout_rate = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                treatment = rep(c("control", "experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.0505 control          16.7        533.  16.7      1
#>  2 High         0.215  control          36.5         24.3 24.5      0
#>  3 High         0.256  experimental      1.28       111.   1.54     1
#>  4 Low          0.274  experimental     96.5       1202.  96.8      1
#>  5 High         0.329  experimental     24.9       1503.  25.2      1
#>  6 High         0.383  control           8.67      1336.   9.05     1
#>  7 High         0.475  experimental      2.21       910.   2.69     1
#>  8 High         0.690  experimental      3.68       735.   4.37     1
#>  9 Low          0.714  control           6.23       389.   6.94     1
#> 10 High         0.838  control           3.12        76.0  3.96     1
#> 11 Low          0.865  experimental     18.3        918.  19.2      1
#> 12 Low          0.987  control          52.2        746.  53.2      1
#> 13 Low          1.03   control          84.1       1167.  85.1      1
#> 14 High         1.04   experimental      1.15       139.   2.20     1
#> 15 High         1.06   control          18.2        255.  19.3      1
#> 16 High         1.12   experimental     46.4       1241.  47.5      1
#> 17 High         1.28   control           8.98       621.  10.3      1
#> 18 High         1.60   experimental     73.6        880.  75.2      1
#> 19 High         1.70   control           4.71      1878.   6.41     1
#> 20 High         1.80   control           9.73       273.  11.5      1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, treatment = "control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03))

dropout_rate <- bind_rows(
  tibble(stratum = "Low" , period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropout_rate = dropout_rate)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [3]
#>    stratum enroll_time treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 High          0.147 experimental     18.7        1545.  18.8      1
#>  2 High          0.364 control          19.3        1133.  19.7      1
#>  3 High          0.713 control          11.6         636.  12.3      1
#>  4 High          0.714 experimental     38.6         140.  39.3      1
#>  5 High          0.766 experimental      1.08       1870.   1.85     1
#>  6 High          0.869 control          23.0         925.  23.9      1
#>  7 High          0.996 experimental     31.6         197.  32.6      1
#>  8 High          1.02  control          27.3        1076.  28.3      1
#>  9 High          1.15  experimental      9.78        737.  10.9      1
#> 10 High          1.24  control          50.7         750.  52.0      1
#> 11 High          1.25  experimental    161.         2050. 162.       1
#> 12 Low           1.29  control         124.          526. 125.       1