Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, Treatment [2]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All           0.280 Experimental     6.78         301.  7.06     1
#>  2 All           0.806 Experimental    64.2          569. 65.0      1
#>  3 All           0.832 Control          5.98        2340.  6.81     1
#>  4 All           1.09  Control          7.66         395.  8.75     1
#>  5 All           1.12  Experimental    36.1          875. 37.3      1
#>  6 All           1.52  Control          9.12         119. 10.6      1
#>  7 All           1.58  Control         11.5         1141. 13.1      1
#>  8 All           1.65  Experimental    80.1          483. 81.8      1
#>  9 All           1.74  Control          1.27        2013.  3.01     1
#> 10 All           1.77  Control         27.0         1370. 28.8      1
#> 11 All           1.98  Experimental     0.508        949.  2.49     1
#> 12 All           2.03  Experimental    50.4          475. 52.4      1
#> 13 All           2.04  Control          2.98        6192.  5.03     1
#> 14 All           2.05  Experimental    75.8          946. 77.9      1
#> 15 All           2.17  Experimental    29.9         1216. 32.1      1
#> 16 All           2.17  Control         13.5         1538. 15.7      1
#> 17 All           2.22  Experimental    21.1         1435. 23.3      1
#> 18 All           2.31  Control          6.72        3350.  9.03     1
#> 19 All           2.41  Control          8.14         277. 10.6      1
#> 20 All           2.43  Experimental     1.91         391.  4.34     1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, Treatment [2]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0233 Experimental    58.6         970.  58.6      1
#>  2 All          0.297  Control          4.58        246.   4.88     1
#>  3 All          0.333  Experimental    35.7        3100.  36.0      1
#>  4 All          0.337  Experimental     8.22       2493.   8.55     1
#>  5 All          0.554  Experimental    12.7        2390.  13.2      1
#>  6 All          0.773  Experimental    45.4         758.  46.2      1
#>  7 All          0.810  Experimental    30.0        1017.  30.8      1
#>  8 All          1.03   Control          4.42        265.   5.45     1
#>  9 All          1.09   Experimental    13.0          36.7 14.0      1
#> 10 All          1.10   Experimental     8.17        238.   9.27     1
#> 11 All          1.16   Experimental     2.97       1007.   4.13     1
#> 12 All          1.32   Control          9.70        418.  11.0      1
#> 13 All          1.50   Experimental     6.67        849.   8.17     1
#> 14 All          1.58   Experimental     2.78         43.3  4.36     1
#> 15 All          1.82   Experimental    30.2         370.  32.0      1
#> 16 All          1.98   Control          2.01       2425.   3.99     1
#> 17 All          2.00   Experimental     9.74        958.  11.7      1
#> 18 All          2.00   Experimental    13.4         478.  15.4      1
#> 19 All          2.17   Experimental     0.536       960.   2.71     1
#> 20 All          2.94   Control          1.65       1543.   4.58     1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, Treatment [4]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 Low           0.113 Experimental      4.43      1833.   4.54     1
#>  2 High          0.201 Control          11.5        107.  11.7      1
#>  3 Low           0.283 Control          31.2        612.  31.5      1
#>  4 Low           0.343 Control          64.7       3604.  65.1      1
#>  5 High          0.421 Control          16.7       1778.  17.1      1
#>  6 High          0.460 Experimental      7.15       194.   7.61     1
#>  7 High          0.654 Experimental     40.3        200.  40.9      1
#>  8 High          0.668 Experimental     59.0        616.  59.7      1
#>  9 Low           0.756 Experimental      5.15      1968.   5.90     1
#> 10 High          0.888 Control           6.96       342.   7.85     1
#> 11 High          1.04  Control          19.3        455.  20.3      1
#> 12 High          1.13  Experimental      8.57       662.   9.70     1
#> 13 Low           1.16  Control          45.6       2506.  46.8      1
#> 14 High          1.23  Control          51.7         94.2 52.9      1
#> 15 Low           1.40  Experimental      2.42       306.   3.82     1
#> 16 Low           1.41  Experimental     30.8        154.  32.2      1
#> 17 Low           1.73  Control          51.1       3159.  52.8      1
#> 18 High          1.86  Experimental     27.1        470.  29.0      1
#> 19 High          1.91  Experimental     14.5         86.1 16.4      1
#> 20 High          1.93  Control           3.09       364.   5.02     1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   stratum, Treatment [4]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 High          0.175 Control          33.6        196.  33.8      1
#>  2 Low           0.468 Experimental      7.76        46.4  8.23     1
#>  3 High          1.00  Experimental      1.61       150.   2.61     1
#>  4 High          1.18  Experimental     28.8        166.  30.0      1
#>  5 High          1.18  Control           5.25       619.   6.43     1
#>  6 Low           1.32  Control          22.6         19.0 20.3      0
#>  7 High          1.44  Experimental      2.01       160.   3.46     1
#>  8 Low           1.80  Experimental     13.7        479.  15.5      1
#>  9 High          1.86  Experimental     23.2       3696.  25.1      1
#> 10 High          1.89  Control          35.5        373.  37.3      1
#> 11 High          2.11  Control          21.5        468.  23.6      1
#> 12 High          2.33  Control          15.1        728.  17.4      1