Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All           0.159 control         23.8        2040.  24.0      1
#>  2 All           0.184 experimental    15.8         322.  16.0      1
#>  3 All           0.225 control          2.71        817.   2.93     1
#>  4 All           0.470 experimental    12.4        1391.  12.9      1
#>  5 All           0.516 control          1.59         49.8  2.11     1
#>  6 All           0.529 control         37.9         196.  38.5      1
#>  7 All           0.718 experimental     9.02       1659.   9.74     1
#>  8 All           0.727 experimental     7.01         18.8  7.74     1
#>  9 All           0.878 experimental     5.33        209.   6.20     1
#> 10 All           1.08  control         28.8        2838.  29.9      1
#> 11 All           1.22  experimental    67.2         180.  68.4      1
#> 12 All           1.35  control          7.34        113.   8.68     1
#> 13 All           1.66  control         16.8         587.  18.5      1
#> 14 All           1.82  control          4.53        289.   6.35     1
#> 15 All           1.84  experimental    85.0         926.  86.8      1
#> 16 All           2.03  experimental    12.2        1126.  14.3      1
#> 17 All           2.18  experimental    16.5         179.  18.7      1
#> 18 All           2.20  control         74.8        2709.  77.0      1
#> 19 All           2.27  control          2.75        212.   5.02     1
#> 20 All           2.57  experimental     0.150       380.   2.72     1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("experimental",3), "control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All           0.102 experimental      8.37      1099.   8.48     1
#>  2 All           0.440 control          10.6        762.  11.0      1
#>  3 All           0.473 experimental     62.2        125.  62.7      1
#>  4 All           0.592 experimental      2.26      2043.   2.85     1
#>  5 All           0.721 experimental     51.8        362.  52.5      1
#>  6 All           0.749 experimental     21.7       2739.  22.4      1
#>  7 All           0.908 control          36.1        138.  37.0      1
#>  8 All           0.991 experimental     45.3        424.  46.3      1
#>  9 All           1.05  experimental     43.4        467.  44.4      1
#> 10 All           1.23  control           4.34      1002.   5.57     1
#> 11 All           1.24  experimental      4.25       988.   5.50     1
#> 12 All           1.64  experimental      3.60      1129.   5.24     1
#> 13 All           1.84  experimental      1.36       316.   3.21     1
#> 14 All           1.93  experimental      7.07       127.   9.00     1
#> 15 All           1.98  experimental     41.1       1100.  43.1      1
#> 16 All           2.00  control          36.0         77.9 38.0      1
#> 17 All           2.05  experimental      9.65      3297.  11.7      1
#> 18 All           2.05  experimental     29.1         72.4 31.1      1
#> 19 All           2.11  experimental     30.0        607.  32.2      1
#> 20 All           2.22  control           7.72       439.   9.94     1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             treatment= rep(c(rep("control", 2),
                                         rep("experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropout_rate = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                treatment = rep(c("control", "experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.0603 control           6.94      3344.   7.00     1
#>  2 High         0.157  control           6.46       138.   6.61     1
#>  3 High         0.312  experimental     12.3       3563.  12.6      1
#>  4 High         0.345  experimental     35.6        917.  36.0      1
#>  5 High         0.347  experimental     41.1         31.2 31.6      0
#>  6 High         0.642  control          13.2       1379.  13.9      1
#>  7 High         0.689  experimental      7.32      3021.   8.01     1
#>  8 High         0.756  control          13.4       1569.  14.2      1
#>  9 High         0.796  experimental      8.13       387.   8.92     1
#> 10 High         0.928  experimental      8.75      2601.   9.68     1
#> 11 High         1.02   control           5.23      1743.   6.25     1
#> 12 Low          1.03   control          11.0       4608.  12.1      1
#> 13 High         1.03   control           9.66        28.1 10.7      1
#> 14 High         1.06   experimental     76.0        169.  77.0      1
#> 15 High         1.83   control           6.65      3351.   8.48     1
#> 16 High         2.02   control           6.09       598.   8.11     1
#> 17 High         2.12   experimental     12.0        587.  14.2      1
#> 18 High         2.30   experimental      2.83       504.   5.13     1
#> 19 High         2.56   control           9.48       310.  12.0      1
#> 20 Low          2.64   experimental     43.6        818.  46.2      1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, treatment = "control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03))

dropout_rate <- bind_rows(
  tibble(stratum = "Low" , period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropout_rate = dropout_rate)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 Low          0.0476 experimental      1.24       963.   1.29     1
#>  2 Low          0.0852 control          23.1        369.  23.2      1
#>  3 Low          0.111  control          54.0        410.  54.1      1
#>  4 High         0.169  experimental     74.2        760.  74.4      1
#>  5 High         0.374  experimental     10.4       1534.  10.8      1
#>  6 High         0.515  control          13.4        537.  13.9      1
#>  7 High         0.738  control          38.4        804.  39.1      1
#>  8 Low          0.894  experimental      2.44      1858.   3.33     1
#>  9 High         1.16   experimental      9.80        98.0 11.0      1
#> 10 High         1.40   control          28.8         53.7 30.2      1
#> 11 Low          1.44   experimental     14.2        452.  15.6      1
#> 12 Low          1.95   control          16.6       1928.  18.6      1