Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 All         0.00592 Control         0.342        155.   0.348     1
#>  2 All         0.332   Experimental   97.4           65.5 65.9       0
#>  3 All         0.586   Control        16.9          314.  17.5       1
#>  4 All         0.738   Experimental   13.1          916.  13.9       1
#>  5 All         0.766   Control        13.7          970.  14.4       1
#>  6 All         0.806   Experimental    1.96        1552.   2.77      1
#>  7 All         0.838   Experimental   23.8         3403.  24.7       1
#>  8 All         0.946   Control         0.0728      3704.   1.02      1
#>  9 All         1.08    Experimental    2.79        1464.   3.87      1
#> 10 All         1.26    Experimental   16.9          637.  18.1       1
#> 11 All         1.67    Control        24.7          124.  26.4       1
#> 12 All         1.95    Control        17.2          359.  19.2       1
#> 13 All         2.01    Control         8.95        1874.  11.0       1
#> 14 All         2.11    Experimental    9.92         933.  12.0       1
#> 15 All         2.15    Control         4.57         281.   6.72      1
#> 16 All         2.20    Experimental   34.4          334.  36.6       1
#> 17 All         2.24    Control         0.586         31.5  2.82      1
#> 18 All         2.30    Experimental    2.99         450.   5.29      1
#> 19 All         2.37    Control        32.7          107.  35.0       1
#> 20 All         2.48    Experimental   51.0         1531.  53.5       1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0103 Experimental    25.1          85.3 25.1      1
#>  2 All          0.417  Control         29.6         886.  30.1      1
#>  3 All          0.427  Experimental     1.18        196.   1.60     1
#>  4 All          0.806  Experimental     1.64       1944.   2.44     1
#>  5 All          1.01   Experimental    25.8         538.  26.8      1
#>  6 All          1.06   Experimental    23.1        1624.  24.2      1
#>  7 All          1.11   Control         19.4         730.  20.5      1
#>  8 All          1.11   Experimental    77.2         111.  78.3      1
#>  9 All          1.12   Control          3.21        175.   4.33     1
#> 10 All          1.20   Experimental    10.1         492.  11.3      1
#> 11 All          1.30   Experimental     6.38       1125.   7.68     1
#> 12 All          1.59   Experimental    24.6         873.  26.2      1
#> 13 All          1.65   Experimental     2.18        141.   3.83     1
#> 14 All          1.65   Experimental     0.836       357.   2.49     1
#> 15 All          1.66   Experimental    35.2          14.6 16.2      0
#> 16 All          1.84   Control          8.63       1526.  10.5      1
#> 17 All          1.85   Experimental    29.0         165.  30.9      1
#> 18 All          1.96   Control          2.16       1113.   4.12     1
#> 19 All          2.04   Experimental    42.9        1682.  45.0      1
#> 20 All          2.05   Experimental    14.7         904.  16.8      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 High          0.343 Experimental     0.172        26.8  0.515     1
#>  2 High          0.473 Control         11.2         126.  11.7       1
#>  3 Low           0.548 Control         11.3        2474.  11.8       1
#>  4 High          0.582 Experimental     0.580       730.   1.16      1
#>  5 Low           0.633 Experimental     3.19        226.   3.82      1
#>  6 High          0.745 Control         21.9        2000.  22.6       1
#>  7 Low           1.61  Experimental    38.1         137.  39.7       1
#>  8 High          1.64  Control         10.4        1044.  12.0       1
#>  9 Low           1.77  Control         12.5         721.  14.3       1
#> 10 High          1.81  Experimental    12.4        4121.  14.2       1
#> 11 High          1.82  Control          5.77        664.   7.59      1
#> 12 Low           2.01  Control         14.2          14.9 16.2       1
#> 13 Low           2.14  Experimental    48.1        1035.  50.2       1
#> 14 High          2.25  Experimental    55.3         850.  57.5       1
#> 15 Low           2.28  Experimental     8.70       2936.  11.0       1
#> 16 High          2.53  Experimental     2.12        250.   4.65      1
#> 17 High          2.65  Control         39.4        1355.  42.1       1
#> 18 High          2.77  Experimental     3.66        794.   6.43      1
#> 19 High          2.79  Control         16.7        4852.  19.5       1
#> 20 High          2.85  Control          3.99       1268.   6.84      1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 High          0.197 Experimental     17.2        664.  17.4      1
#>  2 High          0.202 Control          13.3         49.0 13.5      1
#>  3 High          0.341 Control          23.9        374.  24.3      1
#>  4 Low           0.403 Experimental     80.1       1938.  80.6      1
#>  5 High          0.722 Experimental     25.2        958.  25.9      1
#>  6 High          0.846 Control          67.8        342.  68.7      1
#>  7 Low           1.09  Control          13.2       1012.  14.3      1
#>  8 High          1.16  Experimental      1.55      1068.   2.70     1
#>  9 High          1.28  Control           2.49       370.   3.77     1
#> 10 Low           1.41  Experimental     19.5        459.  20.9      1
#> 11 High          1.79  Experimental     86.5        152.  88.3      1
#> 12 High          1.94  Experimental     41.1       1605.  43.1      1