Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, Treatment [2]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0726 Experimental     8.67        722.   8.74     1
#>  2 All          0.0851 Experimental    63.6        2994.  63.7      1
#>  3 All          0.0962 Control         14.2         417.  14.3      1
#>  4 All          0.156  Control          4.59       1137.   4.75     1
#>  5 All          0.235  Experimental     1.42       1627.   1.65     1
#>  6 All          0.338  Control          1.19       1645.   1.53     1
#>  7 All          0.366  Control          1.57        839.   1.93     1
#>  8 All          0.485  Experimental     2.57         18.4  3.05     1
#>  9 All          0.688  Control         26.8        1219.  27.5      1
#> 10 All          0.733  Control          3.22        219.   3.95     1
#> 11 All          0.793  Experimental     3.04        290.   3.84     1
#> 12 All          0.799  Experimental    33.2        1948.  34.0      1
#> 13 All          0.823  Control         39.5        1761.  40.3      1
#> 14 All          0.831  Control         20.4        1192.  21.2      1
#> 15 All          0.853  Experimental     1.49       3527.   2.34     1
#> 16 All          1.25   Experimental    33.0         708.  34.3      1
#> 17 All          1.61   Control          2.28        641.   3.89     1
#> 18 All          1.70   Control          2.42        833.   4.11     1
#> 19 All          1.94   Experimental     0.517       190.   2.45     1
#> 20 All          1.95   Experimental     9.39       1973.  11.3      1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, Treatment [2]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0251 Experimental     10.3        687.  10.3      1
#>  2 All          0.116  Experimental     10.0       2271.  10.2      1
#>  3 All          0.202  Control          11.7       1038.  11.9      1
#>  4 All          0.234  Experimental     53.7        155.  53.9      1
#>  5 All          0.271  Control           8.86        88.0  9.13     1
#>  6 All          0.411  Experimental      2.82      1328.   3.23     1
#>  7 All          0.436  Experimental      4.22       800.   4.65     1
#>  8 All          0.467  Experimental     16.6        374.  17.1      1
#>  9 All          0.522  Experimental     53.4        942.  53.9      1
#> 10 All          0.529  Experimental     18.0       1254.  18.6      1
#> 11 All          0.529  Experimental     36.2       2109.  36.7      1
#> 12 All          0.578  Control           3.50      2195.   4.08     1
#> 13 All          0.924  Experimental     27.9        574.  28.8      1
#> 14 All          1.03   Control          21.3       1189.  22.3      1
#> 15 All          1.06   Experimental     30.9       1200.  31.9      1
#> 16 All          1.44   Experimental     53.8        613.  55.3      1
#> 17 All          1.46   Control          37.4       2660.  38.9      1
#> 18 All          1.51   Experimental      1.24      1090.   2.75     1
#> 19 All          1.86   Experimental     82.2        271.  84.1      1
#> 20 All          1.87   Experimental     71.3        518.  73.2      1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, Treatment [4]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.0536 Control         13.4        1275.  13.4      1
#>  2 High         0.0789 Experimental     7.50        434.   7.58     1
#>  3 Low          0.0825 Experimental     5.44       1101.   5.53     1
#>  4 High         0.115  Experimental    15.8          81.5 16.0      1
#>  5 High         0.513  Control          3.41        808.   3.92     1
#>  6 Low          0.721  Control          5.19       3046.   5.91     1
#>  7 Low          0.806  Control         24.4         309.  25.2      1
#>  8 High         0.830  Control         28.3        1712.  29.1      1
#>  9 High         0.888  Experimental    19.0        1169.  19.9      1
#> 10 High         1.09   Experimental     0.990       110.   2.08     1
#> 11 High         1.22   Control         14.7        2406.  15.9      1
#> 12 Low          1.27   Experimental     0.470       395.   1.74     1
#> 13 Low          1.28   Experimental    77.8         826.  79.1      1
#> 14 High         1.34   Experimental    30.8          14.3 15.7      0
#> 15 Low          1.37   Control         30.7         328.  32.1      1
#> 16 High         1.40   Control         21.9         144.  23.3      1
#> 17 Low          1.41   Experimental    28.3          27.0 28.4      0
#> 18 High         1.43   Experimental    56.2         225.  57.7      1
#> 19 High         1.49   Control         10.4        1894.  11.9      1
#> 20 Low          1.52   Control          6.21        398.   7.73     1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   stratum, Treatment [4]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.0600 Control          12.7       2069.  12.8      1
#>  2 High         0.269  Experimental      4.72      1608.   4.99     1
#>  3 High         0.489  Control          61.7         41.6 42.1      0
#>  4 High         0.671  Experimental     28.6        964.  29.3      1
#>  5 High         0.695  Experimental      1.12        67.4  1.81     1
#>  6 High         0.714  Control          81.6        315.  82.3      1
#>  7 High         0.840  Experimental     16.6       3487.  17.4      1
#>  8 High         0.991  Control          37.5        488.  38.5      1
#>  9 High         1.32   Control          10.0        687.  11.3      1
#> 10 Low          1.37   Control          51.9       2553.  53.3      1
#> 11 High         1.46   Experimental     86.9        233.  88.3      1
#> 12 Low          1.54   Experimental     38.2       1119.  39.7      1