Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime      cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>    <dbl> <dbl>
#>  1 All          0.0377 control         0.0448        327.   0.0825     1
#>  2 All          0.117  control        22.0           123.  22.2        1
#>  3 All          0.257  experimental   14.6          2021.  14.9        1
#>  4 All          0.521  experimental    0.993         602.   1.51       1
#>  5 All          0.767  experimental    6.88          733.   7.65       1
#>  6 All          0.923  control        12.5           535.  13.4        1
#>  7 All          1.06   experimental    2.59         1711.   3.65       1
#>  8 All          1.22   control         0.0865        162.   1.30       1
#>  9 All          1.30   experimental   83.4           685.  84.7        1
#> 10 All          1.62   control         9.93         1779.  11.5        1
#> 11 All          1.64   experimental    6.77          485.   8.41       1
#> 12 All          1.69   control         7.80          367.   9.49       1
#> 13 All          1.71   experimental   13.2          2291.  14.9        1
#> 14 All          2.17   control         3.42          723.   5.59       1
#> 15 All          2.32   experimental  106.            472. 108.         1
#> 16 All          2.35   control        16.9           135.  19.3        1
#> 17 All          2.46   experimental    8.80          891.  11.3        1
#> 18 All          2.60   experimental    0.355         196.   2.96       1
#> 19 All          2.76   control         0.826         405.   3.58       1
#> 20 All          2.81   control         6.97         3339.   9.77       1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("experimental",3), "control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0515 experimental     2.17       4126.   2.23     1
#>  2 All          0.0569 experimental    14.8          73.2 14.9      1
#>  3 All          0.139  experimental     7.45       3789.   7.59     1
#>  4 All          0.160  control          5.22        382.   5.38     1
#>  5 All          0.383  experimental    30.7         317.  31.0      1
#>  6 All          0.402  experimental     0.742        56.3  1.14     1
#>  7 All          0.453  experimental    77.6          85.5 78.0      1
#>  8 All          0.479  control          2.86        386.   3.34     1
#>  9 All          0.592  experimental     0.566       949.   1.16     1
#> 10 All          0.744  control         16.7        2869.  17.4      1
#> 11 All          0.747  experimental    33.1        1148.  33.9      1
#> 12 All          0.876  experimental     9.58       4374.  10.5      1
#> 13 All          0.961  experimental    74.2         209.  75.1      1
#> 14 All          1.07   experimental     1.39         95.7  2.46     1
#> 15 All          1.14   control         28.7        1250.  29.8      1
#> 16 All          1.32   experimental    18.1        2538.  19.4      1
#> 17 All          1.46   experimental     5.27        272.   6.73     1
#> 18 All          1.55   experimental    43.3         975.  44.8      1
#> 19 All          1.72   experimental    55.7        3496.  57.4      1
#> 20 All          1.76   control         11.4        1935.  13.1      1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             treatment= rep(c(rep("control", 2),
                                         rep("experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropout_rate = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                treatment = rep(c("control", "experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 Low          0.0179 control          2.50        583.   2.51      1
#>  2 High         0.0464 control          0.519      1221.   0.565     1
#>  3 High         0.0949 experimental     6.08        980.   6.17      1
#>  4 High         0.261  control         24.8        1476.  25.1       1
#>  5 High         0.284  experimental     2.14       2231.   2.42      1
#>  6 High         0.910  experimental    27.1         542.  28.0       1
#>  7 Low          0.982  experimental    32.9        1771.  33.9       1
#>  8 Low          1.15   control         16.5        2138.  17.7       1
#>  9 High         1.19   control          3.69        683.   4.88      1
#> 10 Low          1.32   experimental    17.1        1477.  18.4       1
#> 11 High         1.36   control         14.7         882.  16.0       1
#> 12 Low          1.52   experimental    28.1         882.  29.7       1
#> 13 High         1.53   experimental     9.54       3546.  11.1       1
#> 14 Low          1.62   control         53.6          90.0 55.3       1
#> 15 Low          1.66   experimental     7.84         99.4  9.51      1
#> 16 High         1.67   control         14.9        3219.  16.6       1
#> 17 High         1.78   control          6.23         55.3  8.01      1
#> 18 Low          1.81   control          2.50        803.   4.31      1
#> 19 High         1.90   experimental    10.2         274.  12.1       1
#> 20 High         1.93   experimental    44.5         182.  46.4       1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, treatment = "control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03))

dropout_rate <- bind_rows(
  tibble(stratum = "Low" , period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropout_rate = dropout_rate)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 Low           0.210 experimental      3.85      1758.    4.06     1
#>  2 High          0.222 control          30.4       1303.   30.7      1
#>  3 High          0.270 experimental    106.         319.  106.       1
#>  4 High          0.318 control          89.3        578.   89.6      1
#>  5 High          0.332 experimental     10.4       1652.   10.7      1
#>  6 High          0.446 experimental     36.2       1377.   36.7      1
#>  7 High          0.536 control          11.3        350.   11.8      1
#>  8 Low           0.595 control          72.3         84.2  72.9      1
#>  9 Low           0.654 experimental     17.3       2456.   18.0      1
#> 10 Low           0.702 control          49.1        203.   49.8      1
#> 11 High          0.743 control           7.41       813.    8.15     1
#> 12 Low           0.941 control          30.7        373.   31.6      1