Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv() enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribution and enrollment pattern where the enrollment, failure, and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure, and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p.

block

Vector of treatments to be included in each block.

enroll_rate

Enrollment rates; see details and examples.

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block.

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block.

Value

A tibble with the following variables for each observation:

stratum.

enroll_time: Enrollment time for the observation.

Treatment: Treatment group; this will be one of the values in the input block.

fail_time: Failure time generated using rpwexp().

dropout_time: Dropout time generated using rpwexp().

cte: Calendar time of enrollment plot the minimum of failure time and dropout time.

fail: Indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout.

Examples

library(tibble)
library(dplyr)

# Example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All           0.368 experimental   22.4          2055.  22.8      1
#>  2 All           0.602 experimental   75.9           685.  76.5      1
#>  3 All           0.638 control         1.18          825.   1.82     1
#>  4 All           0.764 control        26.9           236.  27.6      1
#>  5 All           0.771 control        22.3          1785.  23.0      1
#>  6 All           0.819 control        16.7          3003.  17.5      1
#>  7 All           1.17  experimental    3.44           67.9  4.61     1
#>  8 All           1.25  experimental    1.21          447.   2.46     1
#>  9 All           1.42  control         1.08          208.   2.50     1
#> 10 All           1.45  experimental   38.2           553.  39.6      1
#> 11 All           1.62  control         3.62          423.   5.25     1
#> 12 All           1.81  experimental   11.4          2270.  13.2      1
#> 13 All           1.86  experimental   25.7          1518.  27.5      1
#> 14 All           2.05  experimental    6.09          971.   8.14     1
#> 15 All           2.07  control         1.19         2135.   3.27     1
#> 16 All           2.45  control        21.0           116.  23.4      1
#> 17 All           2.52  control         6.08         1296.   8.60     1
#> 18 All           2.54  control         0.900         396.   3.44     1
#> 19 All           2.59  experimental    1.40          649.   3.99     1
#> 20 All           2.63  experimental    0.0820       1814.   2.72     1

# Example 2
# 3:1 randomization
sim_pw_surv(
  n = 20,
  block = c(rep("experimental", 3), "control")
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All           0.115 experimental     15.3        1556.  15.4      1
#>  2 All           0.148 experimental     74.5        1011.  74.6      1
#>  3 All           0.164 control          22.3         563.  22.5      1
#>  4 All           0.311 experimental     12.1        2202.  12.4      1
#>  5 All           0.348 experimental     96.9        1133.  97.3      1
#>  6 All           0.364 experimental      6.14       1374.   6.50     1
#>  7 All           0.495 control           6.41       1286.   6.91     1
#>  8 All           0.536 experimental     28.1        1846.  28.6      1
#>  9 All           0.551 experimental     99.0          31.7 32.3      0
#> 10 All           0.804 control          11.6        1083.  12.4      1
#> 11 All           0.807 experimental      8.38        609.   9.19     1
#> 12 All           0.839 experimental     15.4        1489.  16.2      1
#> 13 All           0.976 experimental     34.2        1105.  35.1      1
#> 14 All           1.07  experimental      2.43        792.   3.50     1
#> 15 All           1.10  experimental     51.2         489.  52.3      1
#> 16 All           1.24  control           9.82        635.  11.1      1
#> 17 All           1.28  experimental     31.7         108.  33.0      1
#> 18 All           1.32  experimental      1.93        489.   3.25     1
#> 19 All           1.37  control          11.1         189.  12.5      1
#> 20 All           1.41  experimental      7.20        177.   8.61     1

# Example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(
  n = 20,
  # 2 stratum,30% and 70% prevalence
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = tibble(
    stratum = c(rep("Low", 4), rep("High", 4)),
    period = rep(1:2, 4),
    treatment = rep(c(
      rep("control", 2),
      rep("experimental", 2)
    ), 2),
    duration = rep(c(3, 1), 4),
    rate = c(.03, .05, .03, .03, .05, .08, .07, .04)
  ),
  dropout_rate = tibble(
    stratum = c(rep("Low", 2), rep("High", 2)),
    period = rep(1, 4),
    treatment = rep(c("control", "experimental"), 2),
    duration = rep(1, 4),
    rate = rep(.001, 4)
  )
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 High         0.0796 experimental    78.1         1259.  78.2       1
#>  2 High         0.208  control          0.173       1846.   0.381     1
#>  3 High         0.224  control          1.77         308.   2.00      1
#>  4 Low          0.231  control          4.71         445.   4.94      1
#>  5 High         0.275  experimental    16.8           95.2 17.1       1
#>  6 High         0.320  experimental    61.8          479.  62.1       1
#>  7 High         0.361  control         12.8          948.  13.1       1
#>  8 High         0.443  experimental    69.2          699.  69.6       1
#>  9 High         0.662  control          6.21         618.   6.87      1
#> 10 Low          0.865  experimental    27.1           15.3 16.1       0
#> 11 Low          0.892  experimental     7.34         598.   8.23      1
#> 12 Low          1.05   control         18.3          548.  19.4       1
#> 13 High         1.06   experimental    57.0          424.  58.0       1
#> 14 High         1.10   control          4.71        1829.   5.81      1
#> 15 Low          1.29   experimental    22.4         1690.  23.7       1
#> 16 Low          1.42   experimental    41.4          626.  42.8       1
#> 17 High         1.48   experimental    36.8         1901.  38.3       1
#> 18 High         1.55   control          2.63          35.2  4.18      1
#> 19 Low          1.79   control          0.251       2548.   2.04      1
#> 20 High         1.82   experimental    56.7          134.  58.6       1
# Example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03)
)

dropout_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .001)
)

sim_pw_surv(
  n = 12,
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = fail_rate,
  dropout_rate = dropout_rate
)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 High          0.245 control          4.44          66.2  4.69     1
#>  2 High          0.362 experimental    23.0          232.  23.4      1
#>  3 Low           0.373 experimental     8.86          48.3  9.23     1
#>  4 Low           0.459 control          2.52         137.   2.97     1
#>  5 Low           0.681 experimental    16.9         4401.  17.6      1
#>  6 High          0.712 control          0.482       2304.   1.19     1
#>  7 High          0.799 experimental     1.16        1316.   1.96     1
#>  8 Low           0.837 control          6.90        1267.   7.74     1
#>  9 High          0.850 control         21.3          864.  22.2      1
#> 10 High          0.950 experimental     2.53        2101.   3.48     1
#> 11 High          1.12  experimental    36.9         1355.  38.0      1
#> 12 High          1.16  control         54.7          294.  55.9      1