Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0616 Experimental    22.7       3783.  22.8      1
#>  2 All          0.318  Experimental     6.95        65.2  7.26     1
#>  3 All          0.427  Control         26.4        296.  26.9      1
#>  4 All          0.452  Control         83.2        861.  83.6      1
#>  5 All          0.741  Experimental     2.07       862.   2.81     1
#>  6 All          1.20   Experimental    45.7        275.  46.9      1
#>  7 All          1.26   Control          8.46       366.   9.72     1
#>  8 All          1.28   Control         15.9       5974.  17.2      1
#>  9 All          1.62   Experimental    31.9        350.  33.5      1
#> 10 All          1.62   Experimental     6.51       408.   8.13     1
#> 11 All          1.70   Control          7.08      1211.   8.78     1
#> 12 All          1.73   Control          9.28      2349.  11.0      1
#> 13 All          2.04   Control          4.33       235.   6.37     1
#> 14 All          2.58   Control          1.84       330.   4.42     1
#> 15 All          2.72   Experimental     2.32       770.   5.04     1
#> 16 All          2.86   Experimental    47.3       6662.  50.1      1
#> 17 All          2.89   Control         14.8       3023.  17.7      1
#> 18 All          2.92   Experimental    16.0        345.  18.9      1
#> 19 All          3.15   Experimental     1.79      1471.   4.94     1
#> 20 All          3.16   Control          1.74       226.   4.89     1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0573 Experimental    1.06       1962.   1.11     1
#>  2 All          0.162  Experimental    8.04       1374.   8.20     1
#>  3 All          0.284  Control         6.51        432.   6.79     1
#>  4 All          0.431  Experimental    2.34        439.   2.77     1
#>  5 All          0.475  Experimental   12.9         564.  13.4      1
#>  6 All          0.504  Experimental   13.1         673.  13.6      1
#>  7 All          0.510  Control         4.16         25.5  4.67     1
#>  8 All          0.656  Experimental   15.9         469.  16.6      1
#>  9 All          0.824  Control        36.2         813.  37.0      1
#> 10 All          0.863  Experimental   54.8         700.  55.6      1
#> 11 All          0.885  Experimental    0.623      1338.   1.51     1
#> 12 All          0.957  Experimental    1.75        556.   2.71     1
#> 13 All          1.09   Experimental   13.9          38.1 15.0      1
#> 14 All          1.13   Experimental   18.5          75.8 19.6      1
#> 15 All          1.22   Control         7.92        500.   9.14     1
#> 16 All          1.36   Experimental   26.2        1119.  27.6      1
#> 17 All          1.42   Experimental    0.357       909.   1.78     1
#> 18 All          1.46   Control        10.3         205.  11.8      1
#> 19 All          1.47   Experimental    2.88         32.0  4.35     1
#> 20 All          1.48   Experimental   17.5        1817.  18.9      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low         0.00777 Control         70.9        242.  70.9      1
#>  2 High        0.101   Experimental    25.4         60.1 25.5      1
#>  3 High        0.201   Experimental    76.6       2627.  76.8      1
#>  4 High        0.204   Control         24.2        963.  24.4      1
#>  5 High        0.209   Control         14.7       1147.  14.9      1
#>  6 High        0.328   Experimental    28.0       4923.  28.3      1
#>  7 High        0.448   Experimental    41.4        993.  41.8      1
#>  8 High        0.501   Control          1.92        81.2  2.42     1
#>  9 High        0.604   Control          6.47       499.   7.07     1
#> 10 Low         0.633   Control          7.16      1067.   7.79     1
#> 11 High        0.720   Control          2.59       193.   3.31     1
#> 12 High        0.862   Control          1.26       495.   2.12     1
#> 13 High        1.24    Experimental    13.4        598.  14.7      1
#> 14 Low         1.56    Experimental    21.5       1059.  23.1      1
#> 15 Low         1.74    Experimental    54.9       1508.  56.6      1
#> 16 High        1.80    Experimental     6.64      1047.   8.43     1
#> 17 High        1.91    Control         12.4       3255.  14.3      1
#> 18 Low         1.98    Control          8.48       115.  10.5      1
#> 19 Low         2.02    Experimental    24.5        330.  26.5      1
#> 20 High        2.07    Experimental     2.86      1515.   4.93     1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 Low           0.386 Experimental    20.3         205. 20.6      1
#>  2 High          0.556 Experimental    26.0        1032. 26.6      1
#>  3 Low           0.646 Control         54.1        2096. 54.7      1
#>  4 Low           0.747 Control          3.79        829.  4.54     1
#>  5 High          0.770 Control         11.0         108. 11.7      1
#>  6 Low           0.835 Experimental    37.1         340. 37.9      1
#>  7 Low           0.981 Experimental     8.98        610.  9.96     1
#>  8 High          1.29  Experimental    63.5        3484. 64.8      1
#>  9 Low           1.34  Experimental     8.16       2127.  9.50     1
#> 10 High          1.36  Control          5.15        113.  6.51     1
#> 11 High          1.93  Experimental     8.97        401. 10.9      1
#> 12 High          2.02  Experimental    16.5         782. 18.5      1