Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropout_time (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 All          0.0533 control          1.38        1305.   1.43      1
#>  2 All          0.0863 control         19.5          560.  19.6       1
#>  3 All          0.146  experimental     6.35        1163.   6.50      1
#>  4 All          0.204  experimental    38.2          632.  38.4       1
#>  5 All          0.285  experimental     2.13        2783.   2.41      1
#>  6 All          0.430  control          2.12         169.   2.55      1
#>  7 All          0.507  experimental    19.5           71.6 20.0       1
#>  8 All          0.545  control         22.7          167.  23.2       1
#>  9 All          0.792  experimental    37.0          924.  37.8       1
#> 10 All          0.861  control         17.8         1181.  18.6       1
#> 11 All          0.884  experimental     0.114        498.   0.999     1
#> 12 All          0.977  control          2.34         697.   3.31      1
#> 13 All          1.15   control         33.6         2047.  34.7       1
#> 14 All          1.20   control          1.46         226.   2.66      1
#> 15 All          1.42   experimental   103.            30.9 32.3       0
#> 16 All          1.57   experimental    17.8          900.  19.3       1
#> 17 All          1.65   control         28.6          919.  30.2       1
#> 18 All          1.67   experimental    12.4          129.  14.1       1
#> 19 All          1.69   experimental     1.04         818.   2.72      1
#> 20 All          1.97   control          1.79         897.   3.77      1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("experimental",3), "control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All           0.115 control          2.59        2345.   2.71     1
#>  2 All           0.131 experimental     4.42         338.   4.55     1
#>  3 All           0.220 experimental     0.893       1686.   1.11     1
#>  4 All           0.233 experimental    48.0         1281.  48.2      1
#>  5 All           0.625 experimental    43.3         1214.  44.0      1
#>  6 All           0.628 experimental     6.56        1138.   7.18     1
#>  7 All           0.703 experimental    50.7         6262.  51.4      1
#>  8 All           0.842 control         51.0         1434.  51.9      1
#>  9 All           0.916 experimental    31.4          246.  32.3      1
#> 10 All           0.980 experimental    25.2          874.  26.2      1
#> 11 All           1.14  experimental    41.0         1932.  42.1      1
#> 12 All           1.50  control         14.2         2002.  15.7      1
#> 13 All           1.62  experimental     2.78         108.   4.40     1
#> 14 All           1.68  experimental    28.4          563.  30.1      1
#> 15 All           1.70  experimental    14.2          249.  15.9      1
#> 16 All           1.73  control         13.8         1301.  15.5      1
#> 17 All           1.90  experimental     0.690        194.   2.59     1
#> 18 All           1.95  experimental    17.6           81.0 19.5      1
#> 19 All           2.01  experimental     1.28         758.   3.29     1
#> 20 All           2.04  control         21.7         2207.  23.7      1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             treatment= rep(c(rep("control", 2),
                                         rep("experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropout_rate = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                treatment = rep(c("control", "experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 High         0.0522 control           2.24        187.   2.29     1
#>  2 Low          0.338  experimental      1.00       1174.   1.34     1
#>  3 High         0.383  experimental     14.8        1322.  15.2      1
#>  4 High         0.407  experimental     28.9         353.  29.3      1
#>  5 High         0.455  control           9.14        301.   9.60     1
#>  6 High         0.564  experimental     23.3         151.  23.9      1
#>  7 High         0.770  control          24.1        1851.  24.9      1
#>  8 Low          0.775  control          10.1          95.1 10.9      1
#>  9 Low          0.897  control          40.4         645.  41.2      1
#> 10 High         0.911  experimental      4.28       2386.   5.19     1
#> 11 Low          0.991  experimental     48.7        2290.  49.7      1
#> 12 High         1.08   control           9.72        580.  10.8      1
#> 13 High         1.14   control          13.6        1506.  14.7      1
#> 14 Low          1.44   experimental     37.1        1884.  38.6      1
#> 15 High         1.65   control           8.55       1351.  10.2      1
#> 16 High         1.67   experimental     31.1        3975.  32.8      1
#> 17 High         1.73   experimental      7.69       3515.   9.41     1
#> 18 High         1.99   control          32.4         528.  34.4      1
#> 19 High         2.00   control           3.94        976.   5.94     1
#> 20 High         2.35   experimental     24.0        4185.  26.4      1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, treatment = "control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03))

dropout_rate <- bind_rows(
  tibble(stratum = "Low" , period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropout_rate = dropout_rate)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 High        0.00414 experimental    38.9          223.   38.9      1
#>  2 High        0.0839  control          2.79        2024.    2.87     1
#>  3 High        0.290   control          0.784        579.    1.07     1
#>  4 Low         0.333   experimental   104.          2303.  104.       1
#>  5 Low         0.892   control         66.7          222.   67.6      1
#>  6 High        0.907   experimental    31.9         1124.   32.8      1
#>  7 High        1.10    experimental     3.80        2613.    4.90     1
#>  8 High        1.17    experimental    12.9           27.1  14.1      1
#>  9 High        1.17    control         12.4          483.   13.6      1
#> 10 High        1.28    control         49.3          456.   50.6      1
#> 11 Low         1.32    experimental     5.26          97.7   6.59     1
#> 12 High        1.32    control          6.59         303.    7.92     1