Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, Treatment [2]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0253 Experimental      6.42       1131.  6.44     1
#>  2 All          0.143  Control           5.35        362.  5.49     1
#>  3 All          0.146  Experimental      6.92       1318.  7.07     1
#>  4 All          0.194  Control           3.11        926.  3.30     1
#>  5 All          0.313  Control          12.2         415. 12.5      1
#>  6 All          0.440  Experimental     12.4         860. 12.9      1
#>  7 All          0.545  Experimental     30.9        2034. 31.4      1
#>  8 All          0.685  Control          13.1        1751. 13.8      1
#>  9 All          0.812  Experimental     22.6         413. 23.4      1
#> 10 All          0.951  Control          17.7        2001. 18.6      1
#> 11 All          0.969  Experimental     29.6        3004. 30.5      1
#> 12 All          1.11   Control           4.65        226.  5.76     1
#> 13 All          1.24   Control          51.2        2270. 52.5      1
#> 14 All          1.49   Experimental      9.86       1331. 11.4      1
#> 15 All          1.59   Control          42.6         159. 44.2      1
#> 16 All          1.59   Experimental     21.8        1026. 23.4      1
#> 17 All          1.64   Experimental     39.3         342. 40.9      1
#> 18 All          1.64   Control           7.72       1377.  9.36     1
#> 19 All          1.73   Control           9.23       1948. 11.0      1
#> 20 All          1.75   Experimental     14.5         801. 16.2      1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, Treatment [2]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0948 Experimental    42.7         365.  42.8      1
#>  2 All          0.183  Experimental     5.41        101.   5.59     1
#>  3 All          0.395  Experimental    11.8         142.  12.2      1
#>  4 All          0.507  Control         65.7        3323.  66.2      1
#>  5 All          0.595  Experimental    74.0        1137.  74.6      1
#>  6 All          0.822  Control         28.7        1510.  29.5      1
#>  7 All          0.863  Experimental     0.548      1220.   1.41     1
#>  8 All          0.970  Experimental     0.817      1598.   1.79     1
#>  9 All          0.992  Experimental     2.39        346.   3.38     1
#> 10 All          1.41   Experimental     6.42        335.   7.83     1
#> 11 All          1.44   Control         19.7        1083.  21.2      1
#> 12 All          1.56   Experimental    57.5        1435.  59.1      1
#> 13 All          1.58   Control          6.04        592.   7.62     1
#> 14 All          1.79   Experimental     3.85       2098.   5.64     1
#> 15 All          2.01   Experimental     4.72        261.   6.73     1
#> 16 All          2.03   Experimental     9.92        393.  12.0      1
#> 17 All          2.12   Control         20.2          42.6 22.3      1
#> 18 All          2.24   Experimental     0.892       746.   3.13     1
#> 19 All          2.37   Experimental    29.8         340.  32.2      1
#> 20 All          2.38   Experimental     4.31       2771.   6.70     1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, Treatment [4]
#>    stratum enroll_time Treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 Low          0.0684 Control         29.3        365.   29.3      1
#>  2 High         0.0967 Control         15.0        911.   15.1      1
#>  3 High         0.103  Experimental    33.9       1851.   34.0      1
#>  4 High         0.107  Experimental    11.0        262.   11.1      1
#>  5 High         0.415  Control          7.33      2003.    7.74     1
#>  6 High         0.514  Control         13.7       1583.   14.2      1
#>  7 High         0.596  Experimental     0.665      242.    1.26     1
#>  8 Low          0.597  Control         65.7       1958.   66.3      1
#>  9 High         0.640  Control          9.92      1840.   10.6      1
#> 10 High         0.679  Experimental     3.72       791.    4.40     1
#> 11 High         0.829  Experimental    88.0          5.63  6.46     0
#> 12 High         0.936  Control          4.22       172.    5.16     1
#> 13 High         1.07   Experimental     2.25       226.    3.32     1
#> 14 High         1.20   Control         26.8        627.   28.0      1
#> 15 High         1.26   Experimental    25.7        300.   27.0      1
#> 16 Low          1.35   Experimental     3.67       680.    5.02     1
#> 17 High         1.40   Control          2.01       199.    3.41     1
#> 18 Low          1.50   Experimental    57.2       5011.   58.7      1
#> 19 High         1.63   Control          3.50       559.    5.13     1
#> 20 High         1.81   Experimental    30.0         61.9  31.8      1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   stratum, Treatment [4]
#>    stratum enroll_time Treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 High          0.217 Control          35.5        258.   35.7      1
#>  2 High          0.217 Control          11.9       2488.   12.1      1
#>  3 Low           0.297 Control          21.6       4550.   21.9      1
#>  4 Low           0.422 Experimental    110.        2242.  111.       1
#>  5 Low           0.453 Control          35.3        116.   35.7      1
#>  6 Low           0.499 Experimental     60.1        689.   60.6      1
#>  7 Low           0.516 Experimental      8.23        29.4   8.74     1
#>  8 High          0.703 Experimental     30.9       1569.   31.6      1
#>  9 High          0.781 Experimental     40.5       1014.   41.2      1
#> 10 High          0.844 Experimental     30.3        357.   31.2      1
#> 11 High          0.885 Control           5.88      1934.    6.76     1
#> 12 Low           0.958 Experimental    106.        2386.  107.       1