Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.0223 control         33.6         165.   33.7      1
#>  2 All          0.140  experimental     3.40        150.    3.54     1
#>  3 All          0.156  control          5.50        360.    5.65     1
#>  4 All          0.213  experimental     2.96       1891.    3.17     1
#>  5 All          0.524  experimental    30.6        1979.   31.2      1
#>  6 All          0.533  experimental    15.1          26.5  15.6      1
#>  7 All          0.613  control         18.7        1475.   19.3      1
#>  8 All          0.768  control         25.3         351.   26.0      1
#>  9 All          0.848  control          2.37        254.    3.21     1
#> 10 All          0.853  experimental    12.4         419.   13.2      1
#> 11 All          1.01   experimental     9.67        376.   10.7      1
#> 12 All          1.13   control          1.96         87.7   3.09     1
#> 13 All          1.18   experimental    29.2         684.   30.4      1
#> 14 All          1.20   control         21.3         937.   22.5      1
#> 15 All          1.34   experimental   109.         2251.  110.       1
#> 16 All          1.50   control          0.544      1009.    2.04     1
#> 17 All          1.55   experimental     2.47        389.    4.02     1
#> 18 All          1.80   control         21.3         655.   23.1      1
#> 19 All          2.06   control          1.99        456.    4.04     1
#> 20 All          2.21   experimental     1.99       1095.    4.20     1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("experimental",3), "control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All           0.196 control         11.9         440.  12.1      1
#>  2 All           0.216 experimental    64.8        2286.  65.0      1
#>  3 All           0.296 experimental    13.0         751.  13.3      1
#>  4 All           0.308 experimental    40.2        1344.  40.5      1
#>  5 All           0.510 experimental     1.53        547.   2.04     1
#>  6 All           0.543 experimental    21.3         535.  21.9      1
#>  7 All           0.647 control          0.368       128.   1.01     1
#>  8 All           0.652 experimental     0.892       317.   1.54     1
#>  9 All           0.786 experimental    41.0        2226.  41.8      1
#> 10 All           0.870 control         27.0         304.  27.9      1
#> 11 All           0.910 experimental    58.4        2849.  59.3      1
#> 12 All           0.940 experimental    45.0         570.  45.9      1
#> 13 All           1.05  experimental     1.83         25.2  2.89     1
#> 14 All           1.23  experimental     2.05       1534.   3.28     1
#> 15 All           1.30  control         22.1        4512.  23.4      1
#> 16 All           1.33  experimental    23.4          37.8 24.7      1
#> 17 All           1.37  control         51.0        2703.  52.4      1
#> 18 All           1.44  experimental     0.308       237.   1.75     1
#> 19 All           1.51  experimental     6.92        941.   8.43     1
#> 20 All           1.83  experimental     4.45        883.   6.29     1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             treatment= rep(c(rep("control", 2),
                                         rep("experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropout_rate = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                treatment = rep(c("control", "experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 Low           0.444 control          7.11       1764.   7.56     1
#>  2 High          0.526 control         24.7         112.  25.2      1
#>  3 High          0.745 experimental     4.41        295.   5.15     1
#>  4 High          0.763 experimental     2.79        123.   3.55     1
#>  5 Low           0.879 experimental    89.3         280.  90.2      1
#>  6 Low           1.00  experimental    27.1         855.  28.1      1
#>  7 Low           1.17  control          6.20        821.   7.37     1
#>  8 High          1.27  control          7.23        773.   8.50     1
#>  9 Low           1.28  experimental    17.7         318.  19.0      1
#> 10 Low           1.41  control         23.8         227.  25.2      1
#> 11 Low           1.42  control          0.451       521.   1.87     1
#> 12 High          1.44  control          8.11       1396.   9.55     1
#> 13 High          1.47  control         11.3         757.  12.7      1
#> 14 High          1.57  experimental    17.2         761.  18.8      1
#> 15 High          1.80  experimental    43.0         219.  44.8      1
#> 16 High          1.98  experimental     2.03       1040.   4.00     1
#> 17 High          2.04  control          1.16        483.   3.20     1
#> 18 High          2.06  control         12.3         667.  14.4      1
#> 19 Low           2.40  experimental     8.95        565.  11.3      1
#> 20 Low           2.40  experimental     4.72         30.6  7.12     1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, treatment = "control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03))

dropout_rate <- bind_rows(
  tibble(stratum = "Low" , period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropout_rate = dropout_rate)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 High         0.0743 control           2.51        441.   2.58     1
#>  2 High         0.298  experimental     78.6         331.  78.9      1
#>  3 High         0.350  control          50.2        1306.  50.6      1
#>  4 High         0.366  experimental     68.0         465.  68.4      1
#>  5 Low          0.483  experimental     43.7         218.  44.2      1
#>  6 Low          0.521  experimental    137.         1195. 138.       1
#>  7 High         0.607  experimental     20.6         484.  21.2      1
#>  8 Low          0.769  control          17.3        1210.  18.0      1
#>  9 High         0.846  experimental      5.37       1939.   6.22     1
#> 10 Low          0.899  control           3.78        642.   4.68     1
#> 11 High         0.915  control          32.4         708.  33.3      1
#> 12 High         1.09   control          35.6         343.  36.7      1