Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv() enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribution and enrollment pattern where the enrollment, failure, and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure, and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p.

block

Vector of treatments to be included in each block.

enroll_rate

Enrollment rates; see details and examples.

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block.

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block.

Value

A tibble with the following variables for each observation:

stratum.

enroll_time: Enrollment time for the observation.

Treatment: Treatment group; this will be one of the values in the input block.

fail_time: Failure time generated using rpwexp().

dropout_time: Dropout time generated using rpwexp().

cte: Calendar time of enrollment plot the minimum of failure time and dropout time.

fail: Indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout.

Examples

library(tibble)
library(dplyr)

# Example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All           0.113 experimental     12.4        1280.  12.5      1
#>  2 All           0.212 control           4.93       2016.   5.14     1
#>  3 All           0.269 experimental     38.9         155.  39.2      1
#>  4 All           0.400 control           2.41        651.   2.81     1
#>  5 All           0.467 experimental      2.48        555.   2.95     1
#>  6 All           0.481 experimental     34.0         578.  34.5      1
#>  7 All           0.504 control          43.2         916.  43.7      1
#>  8 All           0.626 control           1.78        315.   2.41     1
#>  9 All           0.650 control          45.3         674.  45.9      1
#> 10 All           0.669 experimental    104.           36.1 36.8      0
#> 11 All           0.681 control           2.02       2167.   2.70     1
#> 12 All           0.941 experimental      8.48         84.6  9.43     1
#> 13 All           1.01  experimental      1.36        219.   2.38     1
#> 14 All           1.14  control          27.8         287.  29.0      1
#> 15 All           1.18  control           2.33        347.   3.51     1
#> 16 All           1.19  experimental      2.42       1400.   3.61     1
#> 17 All           1.40  experimental      3.00        149.   4.40     1
#> 18 All           1.71  control           4.68       1250.   6.39     1
#> 19 All           1.77  experimental     11.4         405.  13.2      1
#> 20 All           2.02  control          16.6         446.  18.7      1

# Example 2
# 3:1 randomization
sim_pw_surv(
  n = 20,
  block = c(rep("experimental", 3), "control")
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All          0.0347 control          4.29          208.  4.32     1
#>  2 All          0.122  experimental     1.20         1696.  1.32     1
#>  3 All          0.250  experimental    18.8           166. 19.0      1
#>  4 All          0.288  experimental    39.5           355. 39.8      1
#>  5 All          0.301  experimental     3.07         1807.  3.37     1
#>  6 All          0.401  experimental     9.14          638.  9.54     1
#>  7 All          0.983  experimental    22.6           248. 23.6      1
#>  8 All          1.06   control          7.00          110.  8.06     1
#>  9 All          1.13   experimental     2.64         1163.  3.77     1
#> 10 All          1.38   experimental     0.800         197.  2.18     1
#> 11 All          1.47   experimental    37.9           203. 39.4      1
#> 12 All          1.48   control          5.23          484.  6.71     1
#> 13 All          1.48   experimental    18.1           385. 19.6      1
#> 14 All          1.48   control          6.50          920.  7.98     1
#> 15 All          1.54   experimental     5.87          265.  7.41     1
#> 16 All          1.62   experimental     2.66         3632.  4.29     1
#> 17 All          1.65   experimental    49.1          2113. 50.7      1
#> 18 All          1.69   experimental     9.31          432. 11.0      1
#> 19 All          1.74   control          4.77         2170.  6.51     1
#> 20 All          1.86   experimental    32.6           642. 34.4      1

# Example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(
  n = 20,
  # 2 stratum,30% and 70% prevalence
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = tibble(
    stratum = c(rep("Low", 4), rep("High", 4)),
    period = rep(1:2, 4),
    treatment = rep(c(
      rep("control", 2),
      rep("experimental", 2)
    ), 2),
    duration = rep(c(3, 1), 4),
    rate = c(.03, .05, .03, .03, .05, .08, .07, .04)
  ),
  dropout_rate = tibble(
    stratum = c(rep("Low", 2), rep("High", 2)),
    period = rep(1, 4),
    treatment = rep(c("control", "experimental"), 2),
    duration = rep(1, 4),
    rate = rep(.001, 4)
  )
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 High          0.115 experimental     20.9         623.  21.0      1
#>  2 High          0.132 experimental      4.76        247.   4.89     1
#>  3 Low           0.220 control           3.85       2152.   4.07     1
#>  4 High          0.363 control          13.4        1047.  13.8      1
#>  5 High          0.397 control          20.9        1007.  21.3      1
#>  6 Low           0.481 experimental     25.1         114.  25.6      1
#>  7 Low           0.665 control          34.8         529.  35.5      1
#>  8 High          0.727 experimental     48.7        1882.  49.4      1
#>  9 Low           0.771 experimental      5.72         82.9  6.49     1
#> 10 High          0.806 control           9.16        469.   9.97     1
#> 11 High          0.924 control          10.7         636.  11.6      1
#> 12 Low           0.945 control          14.5         267.  15.4      1
#> 13 High          1.15  experimental      6.85        926.   8.00     1
#> 14 High          1.30  control           2.74       2440.   4.03     1
#> 15 High          1.34  control          13.7         104.  15.0      1
#> 16 Low           1.37  experimental     23.2         805.  24.5      1
#> 17 High          1.39  experimental     22.2         180.  23.6      1
#> 18 Low           1.59  control           5.28       1415.   6.87     1
#> 19 Low           1.69  experimental      1.61        588.   3.30     1
#> 20 High          1.69  experimental     71.2         835.  72.9      1
# Example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03)
)

dropout_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .001)
)

sim_pw_surv(
  n = 12,
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = fail_rate,
  dropout_rate = dropout_rate
)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 High          0.249 experimental    37.7          892.  37.9      1
#>  2 Low           0.285 control         20.7          747.  21.0      1
#>  3 Low           0.695 experimental    18.6          583.  19.3      1
#>  4 Low           0.703 experimental    21.8          530.  22.5      1
#>  5 High          0.812 experimental     1.28        1562.   2.09     1
#>  6 High          1.09  control          7.69          65.2  8.77     1
#>  7 High          1.11  control         23.8         1056.  24.9      1
#>  8 High          1.22  experimental    35.2         1074.  36.5      1
#>  9 High          1.40  control         20.3          976.  21.7      1
#> 10 High          1.41  experimental     1.91        4818.   3.32     1
#> 11 High          1.54  control         56.5          313.  58.1      1
#> 12 Low           1.67  control          0.914        207.   2.59     1