Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.0220 experimental    0.0952       1815.  0.117     1
#>  2 All          0.152  control        10.4          3039. 10.6       1
#>  3 All          0.389  control        13.2           449. 13.5       1
#>  4 All          0.401  experimental    3.56          873.  3.96      1
#>  5 All          0.468  control        10.3          1198. 10.8       1
#>  6 All          0.701  experimental    2.87          230.  3.57      1
#>  7 All          0.726  control        18.2           553. 18.9       1
#>  8 All          0.940  experimental   24.7           548. 25.6       1
#>  9 All          1.07   control         1.33          402.  2.40      1
#> 10 All          1.11   experimental   25.3          1408. 26.4       1
#> 11 All          1.18   experimental   11.9          2561. 13.1       1
#> 12 All          1.24   control        14.0           532. 15.3       1
#> 13 All          1.34   experimental    3.61         1937.  4.95      1
#> 14 All          1.34   control        11.3          1784. 12.6       1
#> 15 All          1.38   experimental   28.0          1175. 29.4       1
#> 16 All          1.42   control         3.65         1366.  5.07      1
#> 17 All          1.54   experimental    6.73         1081.  8.27      1
#> 18 All          1.75   control         4.86          424.  6.61      1
#> 19 All          2.01   experimental   30.1          2493. 32.1       1
#> 20 All          2.05   control        11.7          1948. 13.7       1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("experimental",3), "control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All           0.189 experimental    39.5        1216.  39.7      1
#>  2 All           0.378 experimental    13.0        1894.  13.4      1
#>  3 All           0.383 control         27.7        1098.  28.1      1
#>  4 All           0.495 experimental     1.54       2432.   2.04     1
#>  5 All           0.669 experimental     1.30        165.   1.97     1
#>  6 All           0.783 experimental     6.81       3217.   7.59     1
#>  7 All           0.858 experimental     4.87       1234.   5.73     1
#>  8 All           0.900 control         33.0         397.  33.9      1
#>  9 All           1.01  experimental     1.48       2760.   2.48     1
#> 10 All           1.11  control         26.1         800.  27.2      1
#> 11 All           1.34  experimental    68.8          28.4 29.7      0
#> 12 All           1.47  experimental     8.75        533.  10.2      1
#> 13 All           1.51  experimental    24.9         437.  26.5      1
#> 14 All           1.76  control          0.868       434.   2.62     1
#> 15 All           1.97  experimental     0.668      2569.   2.64     1
#> 16 All           2.02  experimental    11.8        3381.  13.8      1
#> 17 All           2.25  control         18.3        1366.  20.6      1
#> 18 All           2.28  experimental    10.5         172.  12.8      1
#> 19 All           2.36  experimental     7.76        448.  10.1      1
#> 20 All           2.59  experimental    15.9          55.2 18.5      1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             treatment= rep(c(rep("control", 2),
                                         rep("experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropout_rate = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                treatment = rep(c("control", "experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 High         0.0189 experimental   13.4         1048.  13.4       1
#>  2 Low          0.122  control         4.67          58.0  4.80      1
#>  3 High         0.197  control        11.0          393.  11.2       1
#>  4 Low          0.229  experimental   10.2         3492.  10.5       1
#>  5 Low          0.328  control         0.0760      1614.   0.404     1
#>  6 Low          0.466  experimental   11.0          416.  11.5       1
#>  7 High         0.559  control         2.09        1089.   2.65      1
#>  8 High         0.565  experimental   18.8          754.  19.4       1
#>  9 High         0.763  experimental   14.2         1059.  15.0       1
#> 10 High         0.770  control         1.21        1096.   1.98      1
#> 11 Low          0.936  control         4.44        1206.   5.38      1
#> 12 Low          1.08   experimental   31.6         3912.  32.6       1
#> 13 High         1.14   experimental   68.2          769.  69.3       1
#> 14 Low          1.16   experimental   67.3          864.  68.4       1
#> 15 Low          1.25   control        41.2         1412.  42.4       1
#> 16 Low          1.37   experimental    1.86        3165.   3.23      1
#> 17 High         1.53   control         5.83          13.4  7.36      1
#> 18 High         1.74   control        20.2         1995.  21.9       1
#> 19 High         1.75   control        15.7          103.  17.4       1
#> 20 High         1.93   experimental   18.3         1286.  20.3       1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, treatment = "control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03))

dropout_rate <- bind_rows(
  tibble(stratum = "Low" , period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropout_rate = dropout_rate)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>  <dbl> <dbl>
#>  1 High         0.0287 experimental    33.1         129.  33.1       1
#>  2 High         0.105  experimental    21.0         402.  21.1       1
#>  3 High         0.204  control          6.94         13.7  7.15      1
#>  4 Low          0.211  control          9.08       1111.   9.29      1
#>  5 High         0.238  control          0.310       435.   0.548     1
#>  6 High         0.415  control         40.6         624.  41.0       1
#>  7 High         0.733  experimental    57.6         122.  58.4       1
#>  8 High         0.839  experimental    26.9        2583.  27.7       1
#>  9 High         0.875  control          0.759      1944.   1.63      1
#> 10 Low          1.14   experimental     0.930      1911.   2.07      1
#> 11 High         1.23   experimental    15.9        2739.  17.2       1
#> 12 High         1.44   experimental    71.0          41.0 42.5       0