Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv() enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribution and enrollment pattern where the enrollment, failure, and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure, and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p.

block

Vector of treatments to be included in each block.

enroll_rate

Enrollment rates; see details and examples.

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block.

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block.

Value

A tibble with the following variables for each observation:

stratum.

enroll_time: Enrollment time for the observation.

Treatment: Treatment group; this will be one of the values in the input block.

fail_time: Failure time generated using rpwexp().

dropout_time: Dropout time generated using rpwexp().

cte: Calendar time of enrollment plot the minimum of failure time and dropout time.

fail: Indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout.

Examples

library(tibble)
library(dplyr)

# Example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All           0.395 control          6.44          351.  6.84     1
#>  2 All           0.591 experimental    30.1          2051. 30.7      1
#>  3 All           0.815 experimental    32.8          1577. 33.6      1
#>  4 All           0.841 control          7.70         1617.  8.54     1
#>  5 All           0.880 control          4.45          228.  5.33     1
#>  6 All           0.887 experimental     9.72          573. 10.6      1
#>  7 All           0.914 experimental    53.7          2733. 54.6      1
#>  8 All           1.20  control         10.3           749. 11.5      1
#>  9 All           1.27  experimental     0.978         885.  2.25     1
#> 10 All           1.62  control          2.32          274.  3.94     1
#> 11 All           1.85  control         12.2          1095. 14.1      1
#> 12 All           1.95  experimental    16.1          1418. 18.0      1
#> 13 All           1.97  experimental    37.6          1500. 39.5      1
#> 14 All           2.01  experimental    31.2          2549. 33.2      1
#> 15 All           2.05  control          3.12          218.  5.18     1
#> 16 All           2.06  control          1.74          429.  3.80     1
#> 17 All           2.08  experimental    69.7           276. 71.8      1
#> 18 All           2.45  experimental    44.6           981. 47.0      1
#> 19 All           2.49  control          2.00         1177.  4.49     1
#> 20 All           2.53  control         34.8           488. 37.4      1

# Example 2
# 3:1 randomization
sim_pw_surv(
  n = 20,
  block = c(rep("experimental", 3), "control")
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All           0.359 experimental     7.89         577.   8.25     1
#>  2 All           0.364 experimental    23.5          395.  23.9      1
#>  3 All           0.407 control          1.12         118.   1.53     1
#>  4 All           0.766 experimental    22.7          944.  23.5      1
#>  5 All           0.820 experimental    86.3         1837.  87.2      1
#>  6 All           0.882 experimental     8.20        1103.   9.08     1
#>  7 All           0.967 experimental     2.70        1469.   3.67     1
#>  8 All           0.980 control         13.3          586.  14.2      1
#>  9 All           1.34  experimental     9.15        1243.  10.5      1
#> 10 All           1.35  control          8.03        1721.   9.38     1
#> 11 All           1.45  experimental    19.9           89.7 21.4      1
#> 12 All           1.55  experimental    19.5          192.  21.1      1
#> 13 All           1.85  control          2.95          39.7  4.80     1
#> 14 All           2.26  experimental    38.1         3527.  40.3      1
#> 15 All           2.61  experimental     4.85         915.   7.46     1
#> 16 All           2.67  experimental     0.427        748.   3.09     1
#> 17 All           2.76  experimental    31.0          190.  33.8      1
#> 18 All           2.99  experimental    80.7          938.  83.7      1
#> 19 All           3.05  experimental     9.31        2081.  12.4      1
#> 20 All           3.13  control          6.16         796.   9.28     1

# Example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(
  n = 20,
  # 2 stratum,30% and 70% prevalence
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = tibble(
    stratum = c(rep("Low", 4), rep("High", 4)),
    period = rep(1:2, 4),
    treatment = rep(c(
      rep("control", 2),
      rep("experimental", 2)
    ), 2),
    duration = rep(c(3, 1), 4),
    rate = c(.03, .05, .03, .03, .05, .08, .07, .04)
  ),
  dropout_rate = tibble(
    stratum = c(rep("Low", 2), rep("High", 2)),
    period = rep(1, 4),
    treatment = rep(c("control", "experimental"), 2),
    duration = rep(1, 4),
    rate = rep(.001, 4)
  )
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 Low         0.00386 experimental    65.0          1158. 65.0      1
#>  2 Low         0.0132  control         39.8           289. 39.8      1
#>  3 High        0.0356  experimental     0.965         804.  1.00     1
#>  4 High        0.0818  control          6.57          639.  6.65     1
#>  5 High        0.171   control         13.1          2127. 13.2      1
#>  6 High        0.309   experimental    31.2          1044. 31.5      1
#>  7 High        0.380   experimental    11.5           751. 11.9      1
#>  8 High        0.460   control          3.08          215.  3.54     1
#>  9 Low         0.582   experimental    79.4          2643. 79.9      1
#> 10 Low         0.620   control          4.01          139.  4.62     1
#> 11 High        1.08    experimental     4.27         3867.  5.35     1
#> 12 Low         1.18    control         12.7          3802. 13.9      1
#> 13 Low         1.31    experimental    11.8           251. 13.1      1
#> 14 Low         1.32    experimental    16.2           885. 17.5      1
#> 15 High        1.56    control         10.7           794. 12.2      1
#> 16 High        1.57    control          7.36         1849.  8.93     1
#> 17 High        1.60    control          2.67         1315.  4.26     1
#> 18 High        1.64    experimental    41.0           195. 42.6      1
#> 19 High        1.68    experimental     3.80         1755.  5.48     1
#> 20 Low         2.04    control         19.8           536. 21.9      1
# Example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03)
)

dropout_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .001)
)

sim_pw_surv(
  n = 12,
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = fail_rate,
  dropout_rate = dropout_rate
)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 Low           0.161 control           8.01       1274.    8.17     1
#>  2 High          0.188 experimental     25.0          36.3  25.2      1
#>  3 High          0.202 experimental     32.6         242.   32.8      1
#>  4 High          0.313 control          10.3        1470.   10.7      1
#>  5 Low           0.316 control           5.92        447.    6.23     1
#>  6 Low           0.352 experimental     17.7         482.   18.0      1
#>  7 Low           0.439 experimental    135.         1372.  136.       1
#>  8 Low           0.474 control          42.5          51.9  43.0      1
#>  9 High          0.493 control           5.47       1708.    5.96     1
#> 10 High          0.532 control          13.9         385.   14.5      1
#> 11 High          0.894 experimental     13.0         526.   13.9      1
#> 12 High          0.976 experimental     13.3         579.   14.3      1