Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv() enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribution and enrollment pattern where the enrollment, failure, and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure, and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p.

block

Vector of treatments to be included in each block.

enroll_rate

Enrollment rates; see details and examples.

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block.

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block.

Value

A tibble with the following variables for each observation:

stratum.

enroll_time: Enrollment time for the observation.

Treatment: Treatment group; this will be one of the values in the input block.

fail_time: Failure time generated using rpwexp().

dropout_time: Dropout time generated using rpwexp().

cte: Calendar time of enrollment plot the minimum of failure time and dropout time.

fail: Indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout.

Examples

library(tibble)
library(dplyr)

# Example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 All          0.0135 experimental     0.944        470.   0.958     1
#>  2 All          0.129  control          9.88         811.  10.0       1
#>  3 All          0.294  experimental    43.9          843.  44.2       1
#>  4 All          0.532  control          4.27        3486.   4.80      1
#>  5 All          0.806  experimental    12.0          118.  12.8       1
#>  6 All          0.824  control         17.3          775.  18.1       1
#>  7 All          0.983  control         22.7          566.  23.7       1
#>  8 All          0.988  experimental    13.9          351.  14.9       1
#>  9 All          0.999  experimental    20.2          280.  21.2       1
#> 10 All          1.04   control         15.2          293.  16.2       1
#> 11 All          1.09   experimental    23.9          512.  25.0       1
#> 12 All          1.13   control         34.2          818.  35.4       1
#> 13 All          1.27   control         33.5         1029.  34.7       1
#> 14 All          1.30   experimental     4.95         470.   6.24      1
#> 15 All          1.72   control          2.07         380.   3.79      1
#> 16 All          1.73   experimental     0.623        709.   2.36      1
#> 17 All          1.86   control         11.6         2007.  13.5       1
#> 18 All          2.11   experimental     0.763         47.6  2.87      1
#> 19 All          2.13   experimental    54.9          316.  57.0       1
#> 20 All          2.17   control          6.39         403.   8.56      1

# Example 2
# 3:1 randomization
sim_pw_surv(
  n = 20,
  block = c(rep("experimental", 3), "control")
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All          0.0357 experimental     44.7         599.  44.8      1
#>  2 All          0.0474 experimental     58.8        1099.  58.9      1
#>  3 All          0.128  control          14.1         327.  14.2      1
#>  4 All          0.238  experimental      9.96         10.9 10.2      1
#>  5 All          0.399  experimental     10.9        1215.  11.3      1
#>  6 All          0.416  control           7.71         43.9  8.13     1
#>  7 All          0.443  experimental     18.2         673.  18.7      1
#>  8 All          0.509  experimental     82.5         556.  83.0      1
#>  9 All          0.627  control          14.2        1036.  14.9      1
#> 10 All          0.658  experimental     19.6         149.  20.3      1
#> 11 All          0.846  experimental     27.0        1267.  27.8      1
#> 12 All          0.880  experimental     15.5        2848.  16.4      1
#> 13 All          1.05   experimental     20.2        1107.  21.3      1
#> 14 All          1.07   experimental     41.8         721.  42.8      1
#> 15 All          1.61   experimental     30.5        1668.  32.1      1
#> 16 All          1.84   control           6.59       2032.   8.42     1
#> 17 All          1.96   control          18.3         662.  20.3      1
#> 18 All          1.98   experimental     23.4        1189.  25.3      1
#> 19 All          1.98   experimental     65.8        1725.  67.8      1
#> 20 All          2.05   experimental      1.05        376.   3.10     1

# Example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(
  n = 20,
  # 2 stratum,30% and 70% prevalence
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = tibble(
    stratum = c(rep("Low", 4), rep("High", 4)),
    period = rep(1:2, 4),
    treatment = rep(c(
      rep("control", 2),
      rep("experimental", 2)
    ), 2),
    duration = rep(c(3, 1), 4),
    rate = c(.03, .05, .03, .03, .05, .08, .07, .04)
  ),
  dropout_rate = tibble(
    stratum = c(rep("Low", 2), rep("High", 2)),
    period = rep(1, 4),
    treatment = rep(c("control", "experimental"), 2),
    duration = rep(1, 4),
    rate = rep(.001, 4)
  )
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time     cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>   <dbl> <dbl>
#>  1 High        0.00801 experimental    19.2         1118.   19.2       1
#>  2 High        0.0551  control          5.15         431.    5.21      1
#>  3 High        0.0606  control         29.8          945.   29.8       1
#>  4 High        0.0937  experimental    21.0         2065.   21.1       1
#>  5 Low         0.315   experimental   100.           124.  101.        1
#>  6 High        0.416   control         19.0          576.   19.4       1
#>  7 Low         0.490   control          0.134       1078.    0.623     1
#>  8 High        0.646   experimental     2.60          82.6   3.24      1
#>  9 High        0.882   control         29.8          313.   30.7       1
#> 10 High        0.984   experimental    16.5          706.   17.5       1
#> 11 High        1.03    control         14.3           29.0  15.4       1
#> 12 High        1.20    experimental     2.16        2850.    3.36      1
#> 13 High        1.29    experimental     0.444       2224.    1.73      1
#> 14 High        1.33    control         17.0          439.   18.3       1
#> 15 High        1.63    experimental    49.6         2243.   51.2       1
#> 16 Low         1.74    experimental     1.19         339.    2.93      1
#> 17 High        2.11    control         61.8         1146.   63.9       1
#> 18 High        2.15    experimental    67.6          439.   69.7       1
#> 19 Low         2.42    control         25.3         2040.   27.7       1
#> 20 Low         2.48    experimental    29.4          244.   31.9       1
# Example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03)
)

dropout_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .001)
)

sim_pw_surv(
  n = 12,
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = fail_rate,
  dropout_rate = dropout_rate
)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 High         0.0880 experimental      1.97        356.    2.06     1
#>  2 High         0.101  control           6.31         38.6   6.41     1
#>  3 High         0.147  control           2.34        619.    2.49     1
#>  4 Low          0.148  control           7.25       2595.    7.40     1
#>  5 High         0.179  experimental     14.5        1027.   14.7      1
#>  6 High         0.327  control          66.5         506.   66.8      1
#>  7 High         0.525  experimental     13.6        1006.   14.1      1
#>  8 High         0.583  experimental    175.          679.  176.       1
#>  9 High         0.698  control           3.61        858.    4.31     1
#> 10 High         1.05   experimental     58.1        1722.   59.1      1
#> 11 Low          1.07   control           6.56       2269.    7.63     1
#> 12 Low          1.36   experimental      4.63        728.    5.99     1