Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv() enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribution and enrollment pattern where the enrollment, failure, and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure, and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p.

block

Vector of treatments to be included in each block.

enroll_rate

Enrollment rates; see details and examples.

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block.

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block.

Value

A tibble with the following variables for each observation:

stratum.

enroll_time: Enrollment time for the observation.

Treatment: Treatment group; this will be one of the values in the input block.

fail_time: Failure time generated using rpwexp().

dropout_time: Dropout time generated using rpwexp().

cte: Calendar time of enrollment plot the minimum of failure time and dropout time.

fail: Indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout.

Examples

library(tibble)
library(dplyr)

# Example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 All          0.0218 control        52.9          137.   53.0       1
#>  2 All          0.0490 experimental   26.8          478.   26.9       1
#>  3 All          0.0696 experimental    1.62           4.62  1.69      1
#>  4 All          0.0953 control        10.1         1514.   10.2       1
#>  5 All          0.197  control         0.625        406.    0.822     1
#>  6 All          0.203  experimental    3.53        1786.    3.73      1
#>  7 All          0.375  control         0.0560      2167.    0.431     1
#>  8 All          0.687  experimental    6.12         334.    6.81      1
#>  9 All          0.689  experimental   81.5          168.   82.2       1
#> 10 All          0.746  experimental   19.0          507.   19.8       1
#> 11 All          0.834  control         6.08        1470.    6.92      1
#> 12 All          1.01   control        24.5          423.   25.5       1
#> 13 All          1.11   experimental    0.988        190.    2.10      1
#> 14 All          1.17   control         5.95         459.    7.12      1
#> 15 All          1.34   control         2.46         293.    3.80      1
#> 16 All          1.40   experimental   17.9         1762.   19.4       1
#> 17 All          1.43   control        14.1          372.   15.6       1
#> 18 All          1.50   control        12.5          266.   14.0       1
#> 19 All          1.55   experimental   21.3          647.   22.9       1
#> 20 All          1.57   experimental    8.67           3.83  5.40      0

# Example 2
# 3:1 randomization
sim_pw_surv(
  n = 20,
  block = c(rep("experimental", 3), "control")
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 All           0.131 experimental     77.7        567.    77.8      1
#>  2 All           0.271 experimental     22.6        658.    22.9      1
#>  3 All           0.272 experimental    114.         162.   114.       1
#>  4 All           0.380 control          16.3        189.    16.7      1
#>  5 All           0.422 experimental     21.1          2.63   3.05     0
#>  6 All           0.468 experimental      1.11       665.     1.58     1
#>  7 All           0.821 control           4.05      1628.     4.87     1
#>  8 All           1.37  experimental      2.33      1170.     3.70     1
#>  9 All           1.70  experimental     37.5       1054.    39.2      1
#> 10 All           1.72  control          22.8        284.    24.5      1
#> 11 All           1.77  experimental     25.0       1638.    26.8      1
#> 12 All           1.90  experimental     25.9        417.    27.8      1
#> 13 All           1.97  experimental     59.5        931.    61.4      1
#> 14 All           2.07  experimental      4.06       747.     6.13     1
#> 15 All           2.07  experimental      2.69       895.     4.76     1
#> 16 All           2.26  control          22.2        634.    24.5      1
#> 17 All           2.27  control           5.08      3521.     7.35     1
#> 18 All           2.37  experimental     12.7        918.    15.1      1
#> 19 All           2.65  experimental      2.76       959.     5.41     1
#> 20 All           2.97  experimental     53.4       3201.    56.4      1

# Example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(
  n = 20,
  # 2 stratum,30% and 70% prevalence
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = tibble(
    stratum = c(rep("Low", 4), rep("High", 4)),
    period = rep(1:2, 4),
    treatment = rep(c(
      rep("control", 2),
      rep("experimental", 2)
    ), 2),
    duration = rep(c(3, 1), 4),
    rate = c(.03, .05, .03, .03, .05, .08, .07, .04)
  ),
  dropout_rate = tibble(
    stratum = c(rep("Low", 2), rep("High", 2)),
    period = rep(1, 4),
    treatment = rep(c("control", "experimental"), 2),
    duration = rep(1, 4),
    rate = rep(.001, 4)
  )
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 High        0.00104 experimental     14.9        259.   14.9      1
#>  2 Low         0.130   experimental     18.4        338.   18.5      1
#>  3 High        0.203   control          16.9        539.   17.1      1
#>  4 Low         0.342   experimental     39.4          3.45  3.79     0
#>  5 High        0.469   control           8.21       757.    8.68     1
#>  6 High        0.610   experimental     10.6       2117.   11.2      1
#>  7 High        0.622   control           4.70      1411.    5.33     1
#>  8 High        0.785   control           7.33       736.    8.12     1
#>  9 High        0.829   experimental      3.90      1757.    4.73     1
#> 10 High        0.831   experimental     38.8       1454.   39.6      1
#> 11 High        0.947   control           3.28       109.    4.23     1
#> 12 High        1.05    experimental     15.6       1589.   16.6      1
#> 13 High        1.09    experimental     36.8       2453.   37.9      1
#> 14 High        1.09    control           6.61       604.    7.71     1
#> 15 High        1.12    experimental      1.17      1162.    2.29     1
#> 16 Low         1.22    control           5.78      1607.    7.00     1
#> 17 Low         1.48    control          17.1         48.4  18.5      1
#> 18 Low         1.78    control           7.35       991.    9.12     1
#> 19 High        1.85    experimental      1.47       153.    3.32     1
#> 20 Low         2.03    experimental      3.06        34.5   5.09     1
# Example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03)
)

dropout_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .001)
)

sim_pw_surv(
  n = 12,
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = fail_rate,
  dropout_rate = dropout_rate
)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 Low         0.00680 experimental     12.1        134.   12.1      1
#>  2 High        0.124   control          32.2        605.   32.3      1
#>  3 High        0.308   control          24.5        712.   24.8      1
#>  4 High        0.324   experimental      5.21       385.    5.54     1
#>  5 Low         0.339   control          16.4       3944.   16.8      1
#>  6 High        0.416   experimental      3.60      3852.    4.01     1
#>  7 High        0.547   control          25.6          9.63 10.2      0
#>  8 High        0.561   experimental     13.0        186.   13.6      1
#>  9 High        0.586   control           8.49       258.    9.08     1
#> 10 Low         0.598   control          16.5        197.   17.1      1
#> 11 High        0.606   experimental     60.4        194.   61.0      1
#> 12 High        0.731   experimental     39.3       3804.   40.0      1