Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    failTime dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 All          0.0374 Experimental    9.31         630.  9.34      1
#>  2 All          0.0925 Control         0.252       1841.  0.344     1
#>  3 All          0.185  Experimental   20.6          291. 20.8       1
#>  4 All          0.251  Control         1.83         347.  2.08      1
#>  5 All          0.283  Experimental    8.96         829.  9.25      1
#>  6 All          0.429  Control        21.1          657. 21.6       1
#>  7 All          0.516  Control         2.63         602.  3.14      1
#>  8 All          0.590  Experimental    0.656        134.  1.25      1
#>  9 All          0.606  Control        11.2          585. 11.8       1
#> 10 All          0.754  Control        14.0          749. 14.7       1
#> 11 All          0.805  Experimental   45.2         2026. 46.0       1
#> 12 All          0.968  Experimental   15.3         1770. 16.3       1
#> 13 All          1.12   Experimental   13.3         4241. 14.4       1
#> 14 All          1.39   Control        29.6          228. 30.9       1
#> 15 All          1.47   Control        24.8         4017. 26.2       1
#> 16 All          1.77   Experimental   34.1          424. 35.8       1
#> 17 All          1.96   Experimental   37.4         1114. 39.4       1
#> 18 All          2.09   Control        45.9          976. 47.9       1
#> 19 All          2.30   Control        16.4          646. 18.7       1
#> 20 All          2.38   Experimental   20.1         1353. 22.4       1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All           0.432 Experimental    0.998       911.   1.43     1
#>  2 All           0.518 Experimental    1.30       3568.   1.82     1
#>  3 All           0.555 Control         9.24        557.   9.80     1
#>  4 All           0.587 Experimental    0.576      3207.   1.16     1
#>  5 All           0.847 Control        22.2        2064.  23.1      1
#>  6 All           1.17  Experimental    1.11        538.   2.28     1
#>  7 All           1.27  Experimental   43.8         667.  45.1      1
#>  8 All           1.37  Experimental    7.55        673.   8.92     1
#>  9 All           1.37  Experimental   51.8         438.  53.2      1
#> 10 All           1.54  Experimental   10.4         887.  11.9      1
#> 11 All           1.58  Experimental    4.88       1234.   6.46     1
#> 12 All           1.64  Control         9.36        558.  11.0      1
#> 13 All           1.67  Experimental   72.2          31.8 33.5      0
#> 14 All           1.89  Control        36.9          15.3 17.2      0
#> 15 All           1.92  Experimental    7.78       1315.   9.70     1
#> 16 All           2.04  Experimental    4.54        334.   6.58     1
#> 17 All           2.22  Experimental   13.7        2073.  15.9      1
#> 18 All           2.33  Experimental   26.5        1244.  28.9      1
#> 19 All           2.34  Control         2.76       1526.   5.10     1
#> 20 All           2.45  Experimental   48.4        1734.  50.9      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High          0.135 Control         33.5      2586.   33.6      1
#>  2 Low           0.170 Experimental    32.4      1033.   32.6      1
#>  3 High          0.204 Experimental    13.9       257.   14.1      1
#>  4 Low           0.231 Control         13.7       329.   13.9      1
#>  5 Low           0.254 Experimental    24.1       342.   24.4      1
#>  6 High          0.276 Control          7.32      301.    7.60     1
#>  7 High          0.434 Experimental     3.50      115.    3.93     1
#>  8 High          0.523 Experimental     8.88      369.    9.40     1
#>  9 High          0.575 Control          8.61     1906.    9.18     1
#> 10 High          0.685 Experimental    11.0        18.2  11.6      1
#> 11 High          0.739 Control          8.69     2959.    9.43     1
#> 12 High          0.746 Experimental    12.1       904.   12.9      1
#> 13 High          0.877 Control          2.04        5.32  2.91     1
#> 14 High          0.956 Experimental    53.2      1706.   54.2      1
#> 15 Low           1.97  Control          1.01       29.7   2.98     1
#> 16 Low           1.99  Experimental    30.3       237.   32.3      1
#> 17 High          2.00  Control          7.86     1221.    9.86     1
#> 18 High          2.13  Control         11.3        89.2  13.4      1
#> 19 High          2.33  Experimental     5.40     2748.    7.73     1
#> 20 Low           2.38  Experimental    38.1       292.   40.5      1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    failTime dropoutTime     cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl>   <dbl> <dbl>
#>  1 High          0.146 Experimental  129.         3404.  129.        1
#>  2 Low           0.212 Experimental    9.16        394.    9.37      1
#>  3 Low           0.440 Experimental   71.0        2697.   71.5       1
#>  4 High          0.493 Control        24.6         953.   25.1       1
#>  5 High          0.559 Control        18.0         609.   18.5       1
#>  6 Low           0.580 Control         7.81        594.    8.39      1
#>  7 High          0.674 Experimental    0.154      1323.    0.828     1
#>  8 High          0.810 Control         0.818        52.3   1.63      1
#>  9 Low           0.948 Control        30.7         538.   31.6       1
#> 10 High          1.16  Experimental    6.74        427.    7.90      1
#> 11 High          1.18  Control         4.97       2023.    6.16      1
#> 12 High          1.34  Experimental    0.640       202.    1.98      1