Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv() enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribution and enrollment pattern where the enrollment, failure, and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure, and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p.

block

Vector of treatments to be included in each block.

enroll_rate

Enrollment rates; see details and examples.

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block.

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block.

Value

A tibble with the following variables for each observation:

stratum.

enroll_time: Enrollment time for the observation.

Treatment: Treatment group; this will be one of the values in the input block.

fail_time: Failure time generated using rpwexp().

dropout_time: Dropout time generated using rpwexp().

cte: Calendar time of enrollment plot the minimum of failure time and dropout time.

fail: Indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout.

Examples

library(tibble)
library(dplyr)

# Example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time     cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>   <dbl> <dbl>
#>  1 All        0.000157 control         13.7          2143.  13.7       1
#>  2 All        0.141    experimental    38.8          3871.  39.0       1
#>  3 All        0.315    control          0.249         875.   0.564     1
#>  4 All        0.346    experimental    28.5          1321.  28.9       1
#>  5 All        0.395    control          9.47         1094.   9.86      1
#>  6 All        0.443    experimental    20.5           667.  20.9       1
#>  7 All        0.502    experimental     9.33          655.   9.83      1
#>  8 All        0.558    control         26.7           723.  27.3       1
#>  9 All        0.585    experimental    54.1          1085.  54.7       1
#> 10 All        0.612    control          1.87         1612.   2.48      1
#> 11 All        0.657    control          3.67         1486.   4.33      1
#> 12 All        0.662    experimental     4.74          270.   5.40      1
#> 13 All        0.752    control         40.8          1179.  41.5       1
#> 14 All        0.762    control          5.07         1148.   5.83      1
#> 15 All        0.914    experimental     8.79         1652.   9.70      1
#> 16 All        0.932    experimental    38.9          1701.  39.9       1
#> 17 All        1.16     experimental     3.96         5163.   5.12      1
#> 18 All        1.17     control         64.2           191.  65.4       1
#> 19 All        1.65     experimental   172.           1129. 174.        1
#> 20 All        1.80     control          0.870         670.   2.67      1

# Example 2
# 3:1 randomization
sim_pw_surv(
  n = 20,
  block = c(rep("experimental", 3), "control")
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All          0.0581 experimental    65.2          149.  65.3      1
#>  2 All          0.420  control         12.5          442.  12.9      1
#>  3 All          0.432  experimental     6.12         353.   6.56     1
#>  4 All          0.572  experimental     8.38        2202.   8.95     1
#>  5 All          0.866  experimental     0.950        310.   1.82     1
#>  6 All          0.941  experimental    67.4          137.  68.3      1
#>  7 All          1.13   experimental     0.357        136.   1.49     1
#>  8 All          1.53   control         23.4           86.8 25.0      1
#>  9 All          1.60   experimental     4.34        1156.   5.94     1
#> 10 All          1.77   experimental     5.46         229.   7.23     1
#> 11 All          1.81   control          7.90          82.2  9.70     1
#> 12 All          1.90   experimental    26.8          686.  28.7      1
#> 13 All          1.90   experimental    23.7           30.9 25.6      1
#> 14 All          1.92   experimental    34.1         2311.  36.0      1
#> 15 All          1.98   control         12.0           21.9 14.0      1
#> 16 All          2.20   experimental    10.0         1704.  12.2      1
#> 17 All          2.26   experimental     7.52         695.   9.78     1
#> 18 All          2.28   experimental     1.18        3967.   3.46     1
#> 19 All          2.58   experimental     3.21        1655.   5.79     1
#> 20 All          2.72   control         14.3          997.  17.0      1

# Example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(
  n = 20,
  # 2 stratum,30% and 70% prevalence
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = tibble(
    stratum = c(rep("Low", 4), rep("High", 4)),
    period = rep(1:2, 4),
    treatment = rep(c(
      rep("control", 2),
      rep("experimental", 2)
    ), 2),
    duration = rep(c(3, 1), 4),
    rate = c(.03, .05, .03, .03, .05, .08, .07, .04)
  ),
  dropout_rate = tibble(
    stratum = c(rep("Low", 2), rep("High", 2)),
    period = rep(1, 4),
    treatment = rep(c("control", "experimental"), 2),
    duration = rep(1, 4),
    rate = rep(.001, 4)
  )
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 High          0.204 experimental    61.7         2702.  61.9      1
#>  2 Low           0.741 control          5.09        1401.   5.83     1
#>  3 Low           0.910 experimental    55.7         1442.  56.6      1
#>  4 High          1.05  control         12.8          634.  13.8      1
#>  5 High          1.12  experimental     2.64          41.0  3.75     1
#>  6 High          1.27  control         20.9          423.  22.1      1
#>  7 Low           1.65  experimental    66.3          101.  68.0      1
#>  8 High          1.72  control          9.44         201.  11.2      1
#>  9 High          1.86  control         23.9          225.  25.7      1
#> 10 High          1.96  experimental    51.1           58.5 53.0      1
#> 11 Low           1.96  control         20.9         2016.  22.9      1
#> 12 High          2.44  experimental    26.8          316.  29.3      1
#> 13 High          2.45  experimental    66.1         1849.  68.6      1
#> 14 High          2.48  experimental    57.6          675.  60.1      1
#> 15 Low           2.65  experimental    47.1          545.  49.8      1
#> 16 High          3.08  control          0.314       2803.   3.39     1
#> 17 High          3.12  control         51.6          550.  54.7      1
#> 18 Low           3.18  control         16.2          707.  19.4      1
#> 19 Low           3.23  control         44.9           19.1 22.4      0
#> 20 High          3.48  experimental     1.31        1091.   4.80     1
# Example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03)
)

dropout_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .001)
)

sim_pw_surv(
  n = 12,
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = fail_rate,
  dropout_rate = dropout_rate
)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [3]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 High          0.159 experimental     27.4         1961.  27.6      1
#>  2 High          0.167 control          18.0          282.  18.2      1
#>  3 High          0.245 experimental     10.1          198.  10.4      1
#>  4 Low           0.435 control         141.          2567. 141.       1
#>  5 High          0.507 control           2.66         385.   3.17     1
#>  6 High          0.743 experimental     31.3          363.  32.1      1
#>  7 Low           0.928 control          18.6          988.  19.5      1
#>  8 High          0.944 experimental     27.6         1268.  28.5      1
#>  9 High          1.11  control           9.46         491.  10.6      1
#> 10 High          1.21  control           3.34         424.   4.56     1
#> 11 High          1.31  experimental    122.           759. 124.       1
#> 12 High          1.40  control          11.0          265.  12.4      1