Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), fail_time (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime     cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl>   <dbl> <dbl>
#>  1 All           0.194 control         12.0         1407.  12.2       1
#>  2 All           0.360 experimental   109.           649. 110.        1
#>  3 All           0.403 control          5.22         481.   5.62      1
#>  4 All           0.452 experimental     0.263        960.   0.715     1
#>  5 All           0.473 experimental    30.3         1332.  30.8       1
#>  6 All           0.552 control         18.7         2235.  19.3       1
#>  7 All           0.732 experimental    48.3         1329.  49.1       1
#>  8 All           0.851 control          1.56        2364.   2.41      1
#>  9 All           0.947 control          4.03         251.   4.98      1
#> 10 All           0.992 experimental     5.92         160.   6.91      1
#> 11 All           1.20  control          8.12        1216.   9.32      1
#> 12 All           1.20  experimental     2.28         942.   3.48      1
#> 13 All           1.55  control          2.77         134.   4.32      1
#> 14 All           1.62  experimental     6.87        2196.   8.49      1
#> 15 All           1.88  control          1.08         691.   2.96      1
#> 16 All           1.89  experimental    44.5         1788.  46.3       1
#> 17 All           1.93  control         32.3         3004.  34.3       1
#> 18 All           1.98  experimental     3.26        1311.   5.24      1
#> 19 All           2.06  control          1.54         579.   3.60      1
#> 20 All           2.07  experimental    41.0         2483.  43.0       1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("experimental",3), "control"))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0457 control         14.2         211.  14.2      1
#>  2 All          0.0505 experimental    35.4        1825.  35.4      1
#>  3 All          0.0571 experimental    13.7         660.  13.7      1
#>  4 All          0.227  experimental     1.75       2835.   1.98     1
#>  5 All          0.490  control         22.2         528.  22.7      1
#>  6 All          0.664  experimental    15.2        1046.  15.9      1
#>  7 All          0.927  experimental    31.8         196.  32.7      1
#>  8 All          1.00   experimental    20.3        1916.  21.3      1
#>  9 All          1.11   experimental    27.1          56.5 28.3      1
#> 10 All          1.29   experimental     8.71        435.  10.0      1
#> 11 All          1.53   control          0.806      1881.   2.33     1
#> 12 All          1.55   experimental    60.0        3050.  61.5      1
#> 13 All          1.60   experimental    33.5        1822.  35.1      1
#> 14 All          1.65   experimental     1.36        800.   3.01     1
#> 15 All          1.65   control          9.03        715.  10.7      1
#> 16 All          1.76   experimental    50.4         254.  52.2      1
#> 17 All          1.76   experimental    12.5        3149.  14.3      1
#> 18 All          1.88   experimental     1.68       1086.   3.56     1
#> 19 All          2.26   experimental    37.1         349.  39.4      1
#> 20 All          2.51   control         22.9         463.  25.4      1

# example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 stratum,30% and 70% prevalence
          stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
          fail_rate = tibble(stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             treatment= rep(c(rep("control", 2),
                                         rep("experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropout_rate = tibble(stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                treatment = rep(c("control", "experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 Low          0.0805 experimental     9.55       2468.   9.63     1
#>  2 High         0.120  experimental    27.3          64.6 27.4      1
#>  3 High         0.261  control         11.1         927.  11.4      1
#>  4 High         0.289  control          2.22       4246.   2.51     1
#>  5 Low          0.296  experimental     1.12       3637.   1.41     1
#>  6 Low          0.328  control         13.6         306.  13.9      1
#>  7 Low          0.347  control          3.45        944.   3.80     1
#>  8 High         0.624  experimental     1.23         12.9  1.86     1
#>  9 Low          0.658  experimental    34.9        2634.  35.5      1
#> 10 Low          0.798  experimental    51.4        1069.  52.2      1
#> 11 Low          0.873  control          3.73       3365.   4.60     1
#> 12 High         0.972  control         14.3        1382.  15.2      1
#> 13 High         1.05   experimental     0.729      3783.   1.78     1
#> 14 High         1.12   control         14.0        4211.  15.1      1
#> 15 High         1.59   experimental    18.3         425.  19.9      1
#> 16 High         1.73   control          5.78         83.8  7.51     1
#> 17 High         1.94   experimental     1.32        646.   3.26     1
#> 18 High         2.00   experimental     4.17        526.   6.17     1
#> 19 High         2.10   control         11.4        1102.  13.5      1
#> 20 High         2.12   control         12.0         955.  14.1      1
# example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low" , period = 1, treatment = "control"     , duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low" , period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control"     , duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03))

dropout_rate <- bind_rows(
  tibble(stratum = "Low" , period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "Low" , period=1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "control"     , duration = 3, rate = .001),
  tibble(stratum = "High", period=1, treatment = "experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         stratum = tibble(stratum = c("Low","High"), p = c(.3, .7)),
         fail_rate = fail_rate,
         dropout_rate = dropout_rate)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [3]
#>    stratum enroll_time treatment    fail_time dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>       <dbl> <dbl> <dbl>
#>  1 High          0.137 control          43.2       191.   43.3      1
#>  2 High          0.299 experimental     54.3         3.87  4.17     0
#>  3 High          0.406 control          39.1      1199.   39.5      1
#>  4 High          0.458 experimental      3.67      222.    4.13     1
#>  5 Low           0.516 control          44.3       657.   44.8      1
#>  6 High          0.578 experimental     51.7       102.   52.2      1
#>  7 High          0.599 control          31.3      1818.   31.9      1
#>  8 High          0.601 experimental     32.8       507.   33.4      1
#>  9 Low           0.709 control           3.74      331.    4.45     1
#> 10 High          0.926 control          12.9      1256.   13.9      1
#> 11 High          0.960 experimental     42.6       337.   43.6      1
#> 12 High          1.12  control          23.6      4854.   24.7      1