Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribtuion and enrollment pattern where the enrollment, failure and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  strata = tibble(Stratum = "All", p = 1),
  block = c(rep("Control", 2), rep("Experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  failRates = tibble(Stratum = rep("All", 4), period = rep(1:2, 2), Treatment =
    c(rep("Control", 2), rep("Experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropoutRates = tibble(Stratum = rep("All", 2), period = rep(1, 2), Treatment =
    c("Control", "Experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p

block

Vector of treatments to be included in each block

enroll_rate

Enrollment rates; see details and examples

failRates

Failure rates; see details and examples; note that treatments need to be the same as input in block

dropoutRates

Dropout rates; see details and examples; note that treatments need to be the same as input in block

Value

a tibble with the following variables for each observation Stratum, enroll_time (enrollment time for the observation), Treatment (treatment group; this will be one of the values in the input block), failTime (failure time generated using rpwexp()), dropoutTime (dropout time generated using rpwexp()), cte (calendar time of enrollment plot the minimum of failure time and dropout time), fail (indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout).

Examples

library(dplyr)
library(tibble)

# example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0334 Control        26.2         295.  26.3      1
#>  2 All          0.256  Experimental    4.01        329.   4.26     1
#>  3 All          0.296  Experimental   68.3        1482.  68.6      1
#>  4 All          0.330  Control         4.72        348.   5.05     1
#>  5 All          0.441  Experimental   45.1          11.6 12.1      0
#>  6 All          0.501  Control        23.0         236.  23.5      1
#>  7 All          0.705  Experimental   21.7        1128.  22.4      1
#>  8 All          0.854  Control         1.10        515.   1.95     1
#>  9 All          0.913  Control        30.1         108.  31.0      1
#> 10 All          0.920  Control        11.1         224.  12.0      1
#> 11 All          0.995  Experimental   42.9        2961.  43.9      1
#> 12 All          1.01   Experimental    2.39       1898.   3.41     1
#> 13 All          1.21   Experimental    2.03       1147.   3.24     1
#> 14 All          1.26   Experimental   21.5         109.  22.8      1
#> 15 All          1.54   Control         0.414      2374.   1.95     1
#> 16 All          1.60   Control         2.38       1057.   3.98     1
#> 17 All          1.61   Experimental    8.73        264.  10.3      1
#> 18 All          1.62   Experimental    9.15       1507.  10.8      1
#> 19 All          1.95   Control         0.641      2075.   2.60     1
#> 20 All          2.06   Control        10.1         171.  12.1      1

# example 2
# 3:1 randomization
sim_pw_surv(n = 20,
          block = c(rep("Experimental",3), "Control"))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [2]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 All          0.0390 Experimental    19.3       1857.  19.3      1
#>  2 All          0.208  Experimental    74.4       1440.  74.6      1
#>  3 All          0.218  Control          3.92       636.   4.14     1
#>  4 All          0.248  Experimental    12.6        902.  12.9      1
#>  5 All          0.268  Experimental    31.1        645.  31.3      1
#>  6 All          0.350  Experimental     4.01       299.   4.36     1
#>  7 All          0.501  Control         33.3       1037.  33.8      1
#>  8 All          0.548  Experimental    25.6        924.  26.1      1
#>  9 All          0.573  Control         29.2        409.  29.8      1
#> 10 All          0.663  Experimental     1.24       156.   1.90     1
#> 11 All          0.717  Experimental    12.4       1451.  13.1      1
#> 12 All          0.856  Experimental    20.5       1255.  21.4      1
#> 13 All          1.40   Control          6.52       800.   7.92     1
#> 14 All          1.42   Experimental    26.1       1818.  27.5      1
#> 15 All          1.56   Experimental     4.24      1195.   5.80     1
#> 16 All          1.60   Experimental    53.4        192.  55.0      1
#> 17 All          1.66   Experimental     2.70       730.   4.36     1
#> 18 All          1.99   Experimental    24.7        363.  26.7      1
#> 19 All          2.03   Control          2.97        28.6  5.00     1
#> 20 All          2.08   Experimental    18.3       1082.  20.4      1

# example 3
# Simulate 2 strata; will use defaults for blocking and enrollRates
sim_pw_surv(n = 20,
          # 2 strata,30% and 70% prevalence
          strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
          failRates = tibble(Stratum = c(rep("Low", 4), rep("High", 4)),
                             period = rep(1:2, 4),
                             Treatment = rep(c(rep("Control", 2),
                                         rep("Experimental", 2)), 2),
                             duration = rep(c(3,1), 4),
                             rate = c(.03, .05, .03, .03, .05, .08, .07, .04)),
          dropoutRates = tibble(Stratum = c(rep("Low", 2), rep("High", 2)),
                                period = rep(1, 4),
                                Treatment = rep(c("Control", "Experimental"), 2),
                                duration = rep(1, 4),
                                rate = rep(.001, 4)))
#> # A tibble: 20 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    failTime dropoutTime   cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl> <dbl> <dbl>
#>  1 High         0.0830 Experimental   18.2          38.0 18.3      1
#>  2 High         0.224  Control         3.90        306.   4.12     1
#>  3 High         0.395  Experimental    7.16         40.9  7.56     1
#>  4 High         0.417  Control         2.10        251.   2.51     1
#>  5 Low          0.946  Experimental    8.96       2255.   9.91     1
#>  6 High         1.12   Experimental    4.62        287.   5.74     1
#>  7 High         1.14   Control         4.53       1136.   5.68     1
#>  8 Low          1.18   Control        12.1        1507.  13.3      1
#>  9 High         1.23   Control        36.0        1013.  37.3      1
#> 10 High         1.23   Experimental    0.298      3760.   1.53     1
#> 11 High         1.56   Control         8.08       1562.   9.64     1
#> 12 High         1.56   Experimental   18.2        1075.  19.8      1
#> 13 High         1.58   Control        14.5        1183.  16.1      1
#> 14 High         1.67   Experimental   22.7         210.  24.3      1
#> 15 High         2.45   Experimental    7.38       2321.   9.82     1
#> 16 High         3.03   Experimental    0.436       819.   3.47     1
#> 17 High         3.06   Control        37.0        1311.  40.0      1
#> 18 High         3.11   Control        31.9         793.  35.0      1
#> 19 Low          3.11   Control        12.2          57.0 15.3      1
#> 20 High         3.17   Control        26.3         756.  29.5      1
# example 4
# If you want a more rectangular entry for a tibble
failRates <- bind_rows(
  tibble(Stratum = "Low" , period = 1, Treatment = "Control"     , duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 1, Treatment = "Experimental", duration = 3, rate = .03),
  tibble(Stratum = "Low" , period = 2, Treatment = "Experimental", duration = 3, rate = .02),
  tibble(Stratum = "High", period = 1, Treatment = "Control"     , duration = 3, rate = .05),
  tibble(Stratum = "High", period = 1, Treatment = "Experimental", duration = 3, rate = .06),
  tibble(Stratum = "High", period = 2, Treatment = "Experimental", duration = 3, rate = .03))

dropoutRates <- bind_rows(
  tibble(Stratum = "Low" , period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "Low" , period=1, Treatment = "Experimental", duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Control"     , duration = 3, rate = .001),
  tibble(Stratum = "High", period=1, Treatment = "Experimental", duration = 3, rate = .001))

sim_pw_surv(n = 12,
         strata = tibble(Stratum = c("Low","High"), p = c(.3, .7)),
         failRates = failRates,
         dropoutRates = dropoutRates)
#> # A tibble: 12 × 7
#> # Groups:   Stratum, Treatment [4]
#>    Stratum enroll_time Treatment    failTime dropoutTime    cte  fail
#>    <chr>         <dbl> <chr>           <dbl>       <dbl>  <dbl> <dbl>
#>  1 High          0.194 Control        15.8         1273.  16.0      1
#>  2 Low           0.538 Experimental   28.5         3496.  29.1      1
#>  3 High          0.569 Experimental  272.           281. 273.       1
#>  4 Low           0.720 Control        42.5         1268.  43.3      1
#>  5 Low           0.830 Control        48.1          415.  48.9      1
#>  6 High          0.853 Experimental   99.0          437.  99.9      1
#>  7 High          0.856 Control         6.92         445.   7.78     1
#>  8 Low           1.01  Experimental   47.5          162.  48.5      1
#>  9 Low           1.07  Control        67.5          267.  68.6      1
#> 10 Low           1.16  Experimental   86.5         1631.  87.7      1
#> 11 Low           1.26  Experimental   18.9         1578.  20.2      1
#> 12 High          1.27  Control         0.398        638.   1.67     1