Simulate a stratified time-to-event outcome randomized trial

sim_pw_surv() enables simulation of a clinical trial with essentially arbitrary patterns of enrollment, failure rates and censoring. The piecewise exponential distribution allows a simple method to specify a distribution and enrollment pattern where the enrollment, failure, and dropout rate changes over time. While the main purpose may be to generate a trial that can be analyzed at a single point in time or using group sequential methods, the routine can also be used to simulate an adaptive trial design. Enrollment, failure, and dropout rates are specified by treatment group, stratum and time period. Fixed block randomization is used; blocks must include treatments provided in failure and dropout specification. Default arguments are set up to allow very simple implementation of a non-proportional hazards assumption for an unstratified design.

Usage

sim_pw_surv(
  n = 100,
  stratum = tibble(stratum = "All", p = 1),
  block = c(rep("control", 2), rep("experimental", 2)),
  enroll_rate = tibble(rate = 9, duration = 1),
  fail_rate = tibble(stratum = rep("All", 4), period = rep(1:2, 2), treatment =
    c(rep("control", 2), rep("experimental", 2)), duration = rep(c(3, 1), 2), rate =
    log(2)/c(9, 9, 9, 18)),
  dropout_rate = tibble(stratum = rep("All", 2), period = rep(1, 2), treatment =
    c("control", "experimental"), duration = rep(100, 2), rate = rep(0.001, 2))
)

Arguments

n

Number of observations. If length(n) > 1, the length is taken to be the number required.

stratum

A tibble with stratum specified in stratum, probability (incidence) of each stratum in p.

block

Vector of treatments to be included in each block.

enroll_rate

Enrollment rates; see details and examples.

fail_rate

Failure rates; see details and examples; note that treatments need to be the same as input in block.

dropout_rate

Dropout rates; see details and examples; note that treatments need to be the same as input in block.

Value

A tibble with the following variables for each observation:

stratum.

enroll_time: Enrollment time for the observation.

Treatment: Treatment group; this will be one of the values in the input block.

fail_time: Failure time generated using rpwexp().

dropout_time: Dropout time generated using rpwexp().

cte: Calendar time of enrollment plot the minimum of failure time and dropout time.

fail: Indicator that cte was set using failure time; i.e., 1 is a failure, 0 is a dropout.

Examples

library(tibble)
library(dplyr)

# Example 1
sim_pw_surv(n = 20)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time    cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>  <dbl> <dbl>
#>  1 All           0.117 control          2.35        726.    2.47      1
#>  2 All           0.166 experimental     7.21       1375.    7.38      1
#>  3 All           0.220 control          5.28        279.    5.50      1
#>  4 All           0.459 experimental     4.58         11.3   5.04      1
#>  5 All           0.516 experimental     0.341        98.6   0.857     1
#>  6 All           0.553 experimental    48.5        3940.   49.0       1
#>  7 All           0.602 control          7.90       1507.    8.50      1
#>  8 All           0.731 control          5.35          8.86  6.08      1
#>  9 All           0.801 experimental     4.63        592.    5.43      1
#> 10 All           0.844 control          1.76         20.5   2.60      1
#> 11 All           0.892 control          5.23        152.    6.12      1
#> 12 All           0.918 experimental    25.3        1074.   26.2       1
#> 13 All           1.05  control         20.0        1421.   21.1       1
#> 14 All           1.27  experimental     8.28        461.    9.55      1
#> 15 All           1.29  experimental     1.64        481.    2.93      1
#> 16 All           1.29  control         11.3         460.   12.6       1
#> 17 All           1.39  experimental    19.0          80.2  20.4       1
#> 18 All           1.61  control         10.0         296.   11.6       1
#> 19 All           1.62  control          6.23        256.    7.85      1
#> 20 All           1.68  experimental     0.755       278.    2.43      1

# Example 2
# 3:1 randomization
sim_pw_surv(
  n = 20,
  block = c(rep("experimental", 3), "control")
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [2]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 All          0.0701 experimental    67.0         1603.  67.1      1
#>  2 All          0.314  experimental    24.0          190.  24.3      1
#>  3 All          0.482  control          8.94         848.   9.42     1
#>  4 All          0.746  experimental    20.3          700.  21.0      1
#>  5 All          0.961  experimental    53.9          707.  54.8      1
#>  6 All          1.05   control          0.172        256.   1.22     1
#>  7 All          1.11   experimental    38.2         1101.  39.3      1
#>  8 All          1.27   experimental    77.1         1447.  78.4      1
#>  9 All          1.36   control          6.53         742.   7.89     1
#> 10 All          1.37   experimental    58.8          252.  60.1      1
#> 11 All          1.37   experimental    38.8          481.  40.2      1
#> 12 All          1.45   experimental    13.4          597.  14.9      1
#> 13 All          1.47   experimental     1.31        1697.   2.78     1
#> 14 All          1.79   experimental    17.4         4614.  19.2      1
#> 15 All          1.83   experimental    10.6           86.2 12.4      1
#> 16 All          1.86   control         10.2          686.  12.1      1
#> 17 All          1.90   experimental     2.62        1819.   4.52     1
#> 18 All          1.95   experimental     3.41         290.   5.36     1
#> 19 All          1.96   control         51.0         1135.  53.0      1
#> 20 All          2.09   experimental     2.34         384.   4.42     1

# Example 3
# Simulate 2 stratum; will use defaults for blocking and enrollRates
sim_pw_surv(
  n = 20,
  # 2 stratum,30% and 70% prevalence
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = tibble(
    stratum = c(rep("Low", 4), rep("High", 4)),
    period = rep(1:2, 4),
    treatment = rep(c(
      rep("control", 2),
      rep("experimental", 2)
    ), 2),
    duration = rep(c(3, 1), 4),
    rate = c(.03, .05, .03, .03, .05, .08, .07, .04)
  ),
  dropout_rate = tibble(
    stratum = c(rep("Low", 2), rep("High", 2)),
    period = rep(1, 4),
    treatment = rep(c("control", "experimental"), 2),
    duration = rep(1, 4),
    rate = rep(.001, 4)
  )
)
#> # A tibble: 20 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time   cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl> <dbl> <dbl>
#>  1 High          0.186 experimental     8.55         437.   8.74     1
#>  2 Low           0.317 control         69.4          514.  69.7      1
#>  3 Low           0.343 experimental    23.3          485.  23.6      1
#>  4 High          0.366 control          4.10         791.   4.47     1
#>  5 Low           0.684 control         23.9         1164.  24.6      1
#>  6 High          0.708 control         17.6          398.  18.3      1
#>  7 Low           0.781 experimental    26.3          475.  27.1      1
#>  8 High          0.810 experimental     0.193         43.3  1.00     1
#>  9 High          0.812 experimental    15.0         4175.  15.9      1
#> 10 Low           1.43  experimental    12.7         1625.  14.2      1
#> 11 High          1.48  experimental    20.2          465.  21.7      1
#> 12 High          1.56  control         23.0          759.  24.6      1
#> 13 High          1.69  control         11.6         1005.  13.3      1
#> 14 High          1.76  experimental     4.11        2189.   5.87     1
#> 15 High          1.91  control          6.78         494.   8.69     1
#> 16 Low           2.11  experimental     4.47        1487.   6.58     1
#> 17 High          2.25  control         29.7           21.8 24.1      0
#> 18 High          2.50  experimental     1.83         361.   4.33     1
#> 19 Low           2.84  control         41.4         2212.  44.2      1
#> 20 High          3.05  experimental     0.817       3358.   3.86     1
# Example 4
# If you want a more rectangular entry for a tibble
fail_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .03),
  tibble(stratum = "Low", period = 2, treatment = "experimental", duration = 3, rate = .02),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .05),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .06),
  tibble(stratum = "High", period = 2, treatment = "experimental", duration = 3, rate = .03)
)

dropout_rate <- bind_rows(
  tibble(stratum = "Low", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "Low", period = 1, treatment = "experimental", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "control", duration = 3, rate = .001),
  tibble(stratum = "High", period = 1, treatment = "experimental", duration = 3, rate = .001)
)

sim_pw_surv(
  n = 12,
  stratum = tibble(stratum = c("Low", "High"), p = c(.3, .7)),
  fail_rate = fail_rate,
  dropout_rate = dropout_rate
)
#> # A tibble: 12 × 7
#> # Groups:   stratum, treatment [4]
#>    stratum enroll_time treatment    fail_time dropout_time     cte  fail
#>    <chr>         <dbl> <chr>            <dbl>        <dbl>   <dbl> <dbl>
#>  1 Low          0.0734 experimental    19.8           65.9  19.9       1
#>  2 High         0.129  control          9.46        1250.    9.59      1
#>  3 High         0.150  experimental     5.39        2506.    5.54      1
#>  4 Low          0.408  control          0.300        347.    0.707     1
#>  5 High         0.587  experimental     8.30         646.    8.89      1
#>  6 Low          0.645  experimental    68.5          876.   69.1       1
#>  7 Low          0.718  control         53.4          804.   54.1       1
#>  8 Low          1.26   experimental   181.           465.  182.        1
#>  9 High         1.32   control         31.3         2964.   32.6       1
#> 10 Low          1.44   experimental    54.9          724.   56.3       1
#> 11 High         1.64   control         27.0          404.   28.7       1
#> 12 High         1.70   experimental     8.59        1027.   10.3       1