Simulation of fixed sample size design for time-to-event endpoint

`simfix()` provide simulations of a single endpoint two-arm trial where the enrollment, hazard ratio, and failure and dropout rates change over time.

simfix(
  nsim = 1000,
  sampleSize = 500,
  targetEvents = 350,
  strata = tibble::tibble(Stratum = "All", p = 1),
  enrollRates = tibble::tibble(duration = c(2, 2, 10), rate = c(3, 6, 9)),
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2)),
  totalDuration = 30,
  block = rep(c("Experimental", "Control"), 2),
  timingType = 1:5,
  rg = tibble::tibble(rho = 0, gamma = 0)
)

Arguments

nsim
Number of simulations to perform.

sampleSize
Total sample size per simulation.

targetEvents
Targeted event count for analysis.

strata
A tibble with strata specified in `Stratum`, probability (incidence) of each stratum in `p`.

enrollRates
Piecewise constant enrollment rates by time period. Note that these are overall population enrollment rates and the `strata` argument controls the random distribution between strata.

failRates
Piecewise constant control group failure rates, hazard ratio for experimental vs control, and dropout rates by stratum and time period.

totalDuration
Total follow-up from start of enrollment to data cutoff.

block
As in `simtrial::simPWSurv()`. Vector of treatments to be included in each block.

timingType
A numeric vector determining data cutoffs used; see details. Default is to include all available cutoff methods.

rg
As in `simtrial::tenFHCorr()`. A tibble with variables rho and gamma, both greater than equal to zero, to specify one Fleming-Harrington weighted logrank test per row.

Value

A tibble including columns Events (event count), lnhr (log-hazard ratio), Z (normal test statistic; < 0 favors experimental) cut (text describing cutoff used), Duration (duration of trial at cutoff for analysis) and sim (sequential simulation id). One row per simulated dataset per cutoff specified in timingType, per test statistic specified. If multiple Fleming-Harrington tests are specified in rg, then columns rho,gamma are also included.

Details

timingType has up to 5 elements indicating different options for data cutoff. 1 uses the planned study duration, 2 the time the targeted event count is achieved, 3 the planned minimum follow-up after enrollment is complete, 4 the maximum of planned study duration and targeted event count cuts (1 and 2), 5 the maximum of targeted event count and minimum follow-up cuts (2 and 3).

Examples

library(tidyr)
library(dplyr)
# Show output structure
simfix(nsim=3)

#>    Events       lnhr         Z                              cut Duration Sim
#> 1      81 -0.3525196 -1.582442                 Planned duration 30.00000   1
#> 2     350 -0.4143356 -3.859180                  Targeted events 76.19521   1
#> 3     351 -0.4230050 -3.944232                Minimum follow-up 76.79245   1
#> 4     350 -0.4143356 -3.859180 Max(planned duration, event cut) 76.19521   1
#> 5     351 -0.4230050 -3.944232    Max(min follow-up, event cut) 76.79245   1
#> 6      99 -0.2604907 -1.293595                 Planned duration 30.00000   2
#> 7     350 -0.3553879 -3.306405                  Targeted events 70.53559   2
#> 8     355 -0.3526134 -3.303661                Minimum follow-up 71.88450   2
#> 9     350 -0.3553879 -3.306405 Max(planned duration, event cut) 70.53559   2
#> 10    355 -0.3526134 -3.303661    Max(min follow-up, event cut) 71.88450   2
#> 11    105 -0.2861868 -1.454685                 Planned duration 30.00000   3
#> 12    350 -0.4464634 -4.156319                  Targeted events 67.49890   3
#> 13    360 -0.4305428 -4.068018                Minimum follow-up 70.83052   3
#> 14    350 -0.4464634 -4.156319 Max(planned duration, event cut) 67.49890   3
#> 15    360 -0.4305428 -4.068018    Max(min follow-up, event cut) 70.83052   3

# Example with 2 tests: logrank and FH(0,1)
simfix(nsim=1,rg=tibble::tibble(rho=0,gamma=c(0,1)))

#>    Events       lnhr rho gamma         Z        V1        V2
#> 1      89 -0.6259149   0     0 -2.900541 1.0000000 0.8304269
#> 2      89 -0.6259149   0     1 -2.326387 0.8304269 1.0000000
#> 3     350 -0.5652593   0     0 -5.271915 1.0000000 0.8490496
#> 4     350 -0.5652593   0     1 -4.148967 0.8490496 1.0000000
#> 5     350 -0.5624797   0     0 -5.251611 1.0000000 0.8491038
#> 6     350 -0.5624797   0     1 -4.104058 0.8491038 1.0000000
#> 7     350 -0.5652593   0     0 -5.271915 1.0000000 0.8490496
#> 8     350 -0.5652593   0     1 -4.148967 0.8490496 1.0000000
#> 9     350 -0.5624797   0     0 -5.251611 1.0000000 0.8491038
#> 10    350 -0.5624797   0     1 -4.104058 0.8491038 1.0000000
#>                                 cut Duration Sim
#> 1                  Planned duration 30.00000   1
#> 2                  Planned duration 30.00000   1
#> 3                   Targeted events 74.33769   1
#> 4                   Targeted events 74.33769   1
#> 5                 Minimum follow-up 74.80411   1
#> 6                 Minimum follow-up 74.80411   1
#> 7  Max(planned duration, event cut) 74.33769   1
#> 8  Max(planned duration, event cut) 74.33769   1
#> 9     Max(min follow-up, event cut) 74.80411   1
#> 10    Max(min follow-up, event cut) 74.80411   1

# Power by test
# Only use cuts for events, events + min follow-up
xx <- simfix(nsim=100,timingType=c(2,5),rg=tibble::tibble(rho=0,gamma=c(0,1)))
# Get power approximation for FH, data cutoff combination
xx %>% group_by(cut,rho,gamma) %>% summarise(mean(Z<=qnorm(.025)))

#> `summarise()` has grouped output by 'cut', 'rho'. You can override using the `.groups` argument.

#> # A tibble: 4 x 4
#> # Groups:   cut, rho [2]
#>   cut                             rho gamma `mean(Z <= qnorm(0.025))`
#>   <chr>                         <dbl> <dbl>                     <dbl>
#> 1 Max(min follow-up, event cut)     0     0                      0.98
#> 2 Max(min follow-up, event cut)     0     1                      0.98
#> 3 Targeted events                   0     0                      0.96
#> 4 Targeted events                   0     1                      0.96

# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx %>%  filter(cut != "Targeted events") %>% group_by(Sim) %>% group_map(pMaxCombo)
p <- unlist(p)
mean(p<.025)

#> [1] 0.98

# MaxCombo estimate for targeted events cutoff
p <- unlist(xx %>%  filter(cut == "Targeted events") %>% group_by(Sim) %>% group_map(pMaxCombo))
mean(p<.025)

#> [1] 0.96