Simulation of fixed sample size design for time-to-event endpoint

simfix() provide simulations of a single endpoint two-arm trial where the enrollment, hazard ratio, and failure and dropout rates change over time.

Usage

simfix(
  nsim = 1000,
  sampleSize = 500,
  targetEvents = 350,
  strata = tibble::tibble(Stratum = "All", p = 1),
  enrollRates = tibble::tibble(duration = c(2, 2, 10), rate = c(3, 6, 9)),
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2)),
  totalDuration = 30,
  block = rep(c("Experimental", "Control"), 2),
  timingType = 1:5,
  rg = tibble::tibble(rho = 0, gamma = 0)
)

Arguments

nsim

Number of simulations to perform.

sampleSize

Total sample size per simulation.

targetEvents

Targeted event count for analysis.

strata

A tibble with strata specified in Stratum, probability (incidence) of each stratum in p.

enrollRates

Piecewise constant enrollment rates by time period. Note that these are overall population enrollment rates and the strata argument controls the random distribution between strata.

failRates

Piecewise constant control group failure rates, hazard ratio for experimental vs control, and dropout rates by stratum and time period.

totalDuration

Total follow-up from start of enrollment to data cutoff.

block

As in simtrial::simPWSurv(). Vector of treatments to be included in each block.

timingType

A numeric vector determining data cutoffs used; see details. Default is to include all available cutoff methods.

rg

As in simtrial::tenFHCorr(). A tibble with variables rho and gamma, both greater than equal to zero, to specify one Fleming-Harrington weighted logrank test per row.

Value

A tibble including columns Events (event count), lnhr (log-hazard ratio), Z (normal test statistic; < 0 favors experimental) cut (text describing cutoff used), Duration (duration of trial at cutoff for analysis) and sim (sequential simulation id). One row per simulated dataset per cutoff specified in timingType, per test statistic specified. If multiple Fleming-Harrington tests are specified in rg, then columns rho,gammaare also included.

Details

timingType has up to 5 elements indicating different options for data cutoff. 1 uses the planned study duration, 2 the time the targeted event count is achieved, 3 the planned minimum follow-up after enrollment is complete, 4 the maximum of planned study duration and targeted event count cuts (1 and 2), 5 the maximum of targeted event count and minimum follow-up cuts (2 and 3).

Examples

library(tidyr)
library(dplyr)
# Show output structure
simfix(nsim=3)
#>    Events       lnhr         Z                              cut Duration Sim
#> 1      92 -0.3128897 -1.490033                 Planned duration 30.00000   1
#> 2     350 -0.5021028 -4.612538                  Targeted events 72.85907   1
#> 3     355 -0.5028353 -4.657850                Minimum follow-up 74.43384   1
#> 4     350 -0.5021028 -4.612538 Max(planned duration, event cut) 72.85907   1
#> 5     355 -0.5028353 -4.657850    Max(min follow-up, event cut) 74.43384   1
#> 6     101 -0.4778714 -2.380747                 Planned duration 30.00000   2
#> 7     350 -0.4791463 -4.460631                  Targeted events 65.97012   2
#> 8     380 -0.4922374 -4.766701                Minimum follow-up 73.43063   2
#> 9     350 -0.4791463 -4.460631 Max(planned duration, event cut) 65.97012   2
#> 10    380 -0.4922374 -4.766701    Max(min follow-up, event cut) 73.43063   2
#> 11    104 -0.4804010 -2.419694                 Planned duration 30.00000   3
#> 12    350 -0.4855681 -4.511520                  Targeted events 73.54686   3
#> 13    350 -0.4854395 -4.510305                Minimum follow-up 73.98536   3
#> 14    350 -0.4855681 -4.511520 Max(planned duration, event cut) 73.54686   3
#> 15    350 -0.4854395 -4.510305    Max(min follow-up, event cut) 73.98536   3
# Example with 2 tests: logrank and FH(0,1)
simfix(nsim=1,rg=tibble::tibble(rho=0,gamma=c(0,1)))
#>    Events       lnhr rho gamma         Z        V1        V2
#> 1      95 -0.3680582   0     0 -1.777574 1.0000000 0.8388841
#> 2      95 -0.3680582   0     1 -2.119641 0.8388841 1.0000000
#> 3     350 -0.2272778   0     0 -2.120176 1.0000000 0.8535068
#> 4     350 -0.2272778   0     1 -2.571627 0.8535068 1.0000000
#> 5     357 -0.2333218   0     0 -2.199242 1.0000000 0.8540263
#> 6     357 -0.2333218   0     1 -2.684065 0.8540263 1.0000000
#> 7     350 -0.2272778   0     0 -2.120176 1.0000000 0.8535068
#> 8     350 -0.2272778   0     1 -2.571627 0.8535068 1.0000000
#> 9     357 -0.2333218   0     0 -2.199242 1.0000000 0.8540263
#> 10    357 -0.2333218   0     1 -2.684065 0.8540263 1.0000000
#>                                 cut Duration Sim
#> 1                  Planned duration 30.00000   1
#> 2                  Planned duration 30.00000   1
#> 3                   Targeted events 69.37263   1
#> 4                   Targeted events 69.37263   1
#> 5                 Minimum follow-up 70.54840   1
#> 6                 Minimum follow-up 70.54840   1
#> 7  Max(planned duration, event cut) 69.37263   1
#> 8  Max(planned duration, event cut) 69.37263   1
#> 9     Max(min follow-up, event cut) 70.54840   1
#> 10    Max(min follow-up, event cut) 70.54840   1
# Power by test
# Only use cuts for events, events + min follow-up
xx <- simfix(nsim=100,timingType=c(2,5),rg=tibble::tibble(rho=0,gamma=c(0,1)))
# Get power approximation for FH, data cutoff combination
xx %>% group_by(cut,rho,gamma) %>% summarise(mean(Z<=qnorm(.025)))
#> `summarise()` has grouped output by 'cut', 'rho'. You can override using the
#> `.groups` argument.
#> # A tibble: 4 × 4
#> # Groups:   cut, rho [2]
#>   cut                             rho gamma `mean(Z <= qnorm(0.025))`
#>   <chr>                         <dbl> <dbl>                     <dbl>
#> 1 Max(min follow-up, event cut)     0     0                      0.97
#> 2 Max(min follow-up, event cut)     0     1                      1   
#> 3 Targeted events                   0     0                      0.97
#> 4 Targeted events                   0     1                      0.99
# MaxCombo power estimate for cutoff at max of targeted events, minimum follow-up
p <- xx %>%  filter(cut != "Targeted events") %>% group_by(Sim) %>% group_map(pMaxCombo)
p <- unlist(p)
mean(p<.025)
#> [1] 0.98
# MaxCombo estimate for targeted events cutoff
p <- unlist(xx %>%  filter(cut == "Targeted events") %>% group_by(Sim) %>% group_map(pMaxCombo))
mean(p<.025)
#> [1] 0.99