Conversion of enrollment and failure rates from simfix() to simPWSurv() format

simfix2simPWSurv() converts failure rates and dropout rates entered in the simpler format for simfix() to that used for simtrial::simPWSurv(). The failRates argument for simfix() requires enrollment rates, failure rates hazard ratios and dropout rates by strata for a 2-arm trial, simtrial::simPWSurv() is in a more flexible but less obvious but more flexible format. Since simfix() automatically analyzes data and simtrial::simPWSurv() just produces a simulation dataset, the latter provides additional options to analyze or otherwise evaluate individual simulations in ways that simfix() does not.

Usage

simfix2simPWSurv(
  failRates = tibble::tibble(Stratum = "All", duration = c(3, 100), failRate =
    log(2)/c(9, 18), hr = c(0.9, 0.6), dropoutRate = rep(0.001, 2))
)

Arguments

failRates

Piecewise constant control group failure rates, hazard ratio for experimental vs control, and dropout rates by stratum and time period.

Value

A list of two tibble components formatted for simtrial::simPWSurv(): failRates and dropoutRates.

Examples

library(tidyr)
library(dplyr)
# Convert standard input
simfix2simPWSurv()
#> $failRates
#> # A tibble: 4 × 5
#>   Stratum period Treatment    duration   rate
#>   <chr>    <int> <chr>           <dbl>  <dbl>
#> 1 All          1 Control             3 0.0770
#> 2 All          2 Control           100 0.0385
#> 3 All          1 Experimental        3 0.0693
#> 4 All          2 Experimental      100 0.0231
#> 
#> $dropoutRates
#> # A tibble: 4 × 5
#>   Stratum period Treatment    duration  rate
#>   <chr>    <int> <chr>           <dbl> <dbl>
#> 1 All          1 Control             3 0.001
#> 2 All          2 Control           100 0.001
#> 3 All          1 Experimental        3 0.001
#> 4 All          2 Experimental      100 0.001
#> 
# Stratified example
failRates <- tibble::tibble(Stratum=c(rep("Low",3),rep("High",3)),
                            duration=rep(c(4,10,100),2),
                            failRate=c(.04,.1,.06,
                                       .08,.16,.12),
                            hr=c(1.5,.5,2/3,
                                 2, 10/16, 10/12),
                            dropoutRate=.01
)
x <- simfix2simPWSurv(failRates)
# Do a single simulation with the above rates
# Enroll 300 patients in ~12 months at constant rate
sim <-
    simPWSurv(n=300,
          strata=tibble::tibble(Stratum=c("Low","High"),p=c(.6,.4)),
          enrollRates=tibble::tibble(duration=12,rate=300/12),
          failRates=x$failRates,
          dropoutRates=x$dropoutRates)
# Cut after 200 events and do a stratified logrank test
dat <- sim %>%
       cutDataAtCount(200) %>%            # cut data
       tensurv(txval="Experimental") %>%  # convert format for tenFH
       tenFH(rg=tibble(rho=0,gamma=0))    # stratified logrank