Magirr and Burman Modestly Weighted Logrank Tests

Magirr and Burman (2019) have proposed a weighted logrank test to have better power than the logrank test when the treatment effect is delayed, but to still maintain good power under a proportional hazards assumption. The weights for some early interval specified by the user are the inverse of the combined treatment group empirical survival distribution; see details. After this initial period, weights are constant at the maximum of the previous weights. Another advantage of the test is that under strong null hypothesis that the underlying survival in the control group is greater than or equal to underlying survival in the experimental group, Type I error is controlled as the specified level.

Usage

wMB(x, delay = 4)

Arguments

x

a tensurv-class tibble with a counting process dataset

delay

The initial delay period where weights increase; after this, weights are constant at the final weigh in the delay period

Value

a vector with weights for the Magirr-Burman weighted logrank test for the data in x

Details

This function computes Magirr-Burman weights and adds them to a dataset created by the tensurv() function. These weights can then be used to compute a Z-statistic for the modestly weighted logrank test proposed.

We define \(t^*\) to be the input variable delay. This specifies an initial period during which weights increase. To define specific weights, we let \(S(t)\) denote the Kaplan-Meier survival estimate at time \(t\) for the combined data (control plus experimental treatment groups). The weight at time \(t\) is then defined as $$w(t)=S(\min(t,t^*))^{-1}.$$

References

Magirr, Dominic, and Carl‐Fredrik Burman. "Modestly weighted logrank tests." Statistics in Medicine 38.20 (2019): 3782-3790.

Examples

library(tidyr)
library(dplyr)
# Use default enrollment and event rates at cut at 100 events
x <- simPWSurv(n=200) %>% cutDataAtCount(125) %>% tensurv(txval="Experimental")
# compute Magirr-Burman weights with 
ZMB <- x %>% wMB(6) %>% 
             summarize(S=sum(OminusE*wMB),V=sum(Var*wMB^2),Z=S/sqrt(V))
# Compute p-value of modestly weighted logrank of Magirr-Burman
pnorm(ZMB$Z)
#> [1] 0.169131